### 2021-10-07 #3806 @emdupre
:wave: Hi @jameschapman19, and thank you for your submission to JOSS !

If you have any suggestions for potential reviewers, please let me know by listing their names or GitHub handles (without the @, so they don't receive a notification) here.
