### 2021-10-08 00defad
Moving datasets out of ccagame folder. Fixing __init__.py files
