+++ scripts/pls/xrmb.py
@@ -2,7 +2,7 @@
 
 # Imports
 
-from ccagame import datasets
+import datasets
 from ccagame.pls import Game, SGD, Incremental, Batch, Numpy, MSG
 
 # %%
