+++ ccagame/pca/__init__.py
@@ -1,2 +1,6 @@
 from ._pca import _PCA
 from .numpy import calc_numpy, Numpy
+from .game import Game
+from .oja import Oja
+from .krasulina import Krasulina
+from .gha import GHA
