+++ ccagame/pls/__init__.py
@@ -1,3 +1,7 @@
 from ._pls import _PLS
 from .msg import MSG
+from .game import Game
+from .incremental import Incremental
+from .batch import Batch
+from .sgd import SGD
 from .numpy import calc_numpy, Numpy
