+++ ccagame/cca/__init__.py
@@ -1,2 +1,7 @@
 from ._cca import _CCA
 from .numpy import calc_numpy, Numpy
+from .game import Game
+from .genoja import Genoja
+from .lagrange import Lagrange
+from .ccalin import CCALin
+from .alternatingleastsquares import AlternatingLeastSquares
