+++ scripts/pls/uk_biobank.py
@@ -5,7 +5,7 @@ import sys
 sys.executable
 #%%
 
-from ccagame import datasets
+import datasets
 from ccagame.pls import Game, SGD, Incremental, Batch, Numpy, MSG
 import numpy as np
 import os
