### 2021-10-08 #3806 @emdupre
You can suggest reviewers from any relevant project, though we often recommend starting with [this list of people](https://bit.ly/joss-reviewers) as they have already agreed to review for JOSS (please start at the bottom of the list).
