### 2021-10-08 69e3247
small change to data loading to output indices too... CCAGame is born thanks to BCD-style solution!
