+++ ccagame/cca/game.py
@@ -3,51 +3,86 @@ import time
 from functools import partial
 
 import jax.numpy as jnp
-from jax import grad
-from jax import jit
+from jax import grad, jit
 
+from ccagame.utils import data_stream, get_num_batches
 from . import _CCA
 from .utils import initialize, TCC
-from ..utils import data_stream, get_num_batches
 
 
-@partial(jit, static_argnums=5)
-def model(u, v, X, Y, V, k):
+@partial(jit, static_argnums=(4))
+def alpha_model(u, X, U, T, k: int):
     """
     Returns the utilities for the kth players
 
     Parameters
-        ----------
-        u :
-            current estimate for this level's left eigenvector
-        v :
-            current estimate for this level's right eigenvector
-        X :
-            batch of data for view X
-        Y :
-            batch of data for view Y
-        V :
-            all eigenvector estimates for each level
-        k :
-            level
+    ----------
+    u :
+        current estimate for this level's left eigenvector
+    X :
+        batch of data for view X
+    U :
+        all eigenvector estimates for each level
+    T :
+        the shared target
+    k :
+        level
 
     Returns
     -------
 
     """
-    C_xy = jnp.dot(jnp.transpose(X), Y)
-    C_xx = jnp.dot(jnp.transpose(X), X)
-    C_yy = jnp.dot(jnp.transpose(Y), Y)
-    rewards = (u.T @ C_xy @ v) ** 2 / (v.T @ C_yy @ v)
-    penalties = (u.T @ C_xy @ V[:, :k]) ** 2 / jnp.diag(V[:, :k].T @ C_yy @ V[:, :k])
-    return jnp.sum(rewards - penalties.sum()) / (u.T @ C_xx @ u)
+    C_xt = X.T @ T
+    rewards = -jnp.linalg.norm(T[:, k] - X @ u) ** 2
+    penalties = (u.T @ C_xt[:, :k]) ** 2 / jnp.diag(U[:, :k].T @ C_xt[:, :k])
+    return jnp.sum(rewards - penalties.sum())
+
+
+@partial(jit, static_argnums=(4))
+def mu_model(u, X, U, T, k: int):
+    """
+    Returns the gradients for the kth players
+
+    Parameters
+    ----------
+    u :
+        current estimate for this level's left eigenvector
+    X :
+        batch of data for view X
+    U :
+        all eigenvector estimates for each level
+    T :
+        the shared target
+    k :
+        level
+
+    Returns
+    -------
+
+    """
+    C_xt = X.T @ T
+    rewards = C_xt[:, k] - X.T@X@u
+    penalties = (u.T @ C_xt[:, :k]) * U[:, :k]
+    return rewards - penalties.sum(axis=1)
 
 
 # Update rule to be used for calculating eigenvectors
-@partial(jit, static_argnums=6, static_argnames=("lr", "riemannian_projection"))
-def update(u, v, X, Y, U, V, k, lr: float = 1.0, riemannian_projection=False):
+@partial(jit, static_argnums=7, static_argnames=("lr", "mu"))
+def update(
+        u,
+        v,
+        X,
+        Y,
+        U,
+        V,
+        T,
+        k: int,
+        lr: float = 1,
+        mu=True,
+):
     """
     Update the left and right singular vector estimates
+
     Parameters
     ----------
     u :
@@ -66,20 +101,26 @@ def update(u, v, X, Y, U, V, k, lr: float = 1.0, riemannian_projection=False):
         level
     lr :
         learning rate
-    riemannian_projection :
-        whether to use riemannian projection
+    mu :
[…]
