+++ ccagame/__init__.py
@@ -1,5 +1,4 @@
 from . import cca
-import datasets
 from . import pca
 from . import pls
 from . import solver
