+++ scripts/cca/example.py
@@ -4,7 +4,9 @@ import jax.numpy as jnp
 from jax import random
 
 # Imports
-from ccagame.cca import Numpy, Genoja, Game, Lagrange, AlternatingLeastSquares, CCALin
+from ccagame.cca import Numpy, Genoja, Game, Lagrange, AlternatingLeastSquares
+from sklearn.cross_decomposition import CCA
+import numpy as np
 
 # %%
 
@@ -14,9 +16,8 @@ n = 100
 p = 10
 q = 11
 latent_dims = 5
-max_iter = 300
-batch_size = 100
-epochs = 5
+batch_size = 5
+epochs = 100
 riemannian_projection = True
 initialization = 'random'
 lr = 1
@@ -33,15 +34,15 @@ X = X / jnp.linalg.norm(X, axis=0)
 Y = random.normal(subkey, (n, q))
 Y = Y / jnp.linalg.norm(Y, axis=0)
 
-# %%
-
+a = CCA(n_components=5).fit(X, Y)
+n = np.corrcoef(a.x_scores_, a.y_scores_, rowvar=False)
 # Model
 numpy = Numpy(scale=False, n_components=latent_dims).fit(X, Y)
 print("\n Eigenvalues calculated using numpy are :\n", numpy.score(X, Y))
 print("\n Time :\n", numpy.fit_time)
 
 game = Game(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True,
-            mu=True).fit(X, Y)
+            mu=True, simultaneous=True).fit(X, Y)
 print("\n Eigenvalues calculated using game are :\n", game.score(X, Y))
 print("\n Time :\n", game.fit_time)
 
