### 2021-10-11 cc1cc03
Tutorial Link updates for JOSS
