### 2021-10-13 8d6d780
adding some baselines for stochastic/scalable cca
