### 2021-10-14 0c6ec9c
adding some baselines for stochastic/scalable cca
