+++ ccagame/cca/_cca.py
@@ -3,8 +3,8 @@ from ccagame.pls import _PLS
 
 
 class _CCA(_PLS):
-    def __init__(self, n_components=2, *, scale=True, copy=True, wandb=True):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb)
+    def __init__(self, n_components=2, *, scale=True, copy=True, wandb=True, verbose=False, random_state=None):
+        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
 
     def score(self, X, y, sample_weight=None):
         return TCC(X, y, self.x_weights, self.y_weights)
