+++ ccagame/pls/msg.py
@@ -58,12 +58,10 @@ class MSG(_PLS):
             verbose=False,
             wandb=False
     ):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb)
+        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
         self.lr = lr
         self.epochs = epochs
-        self.random_state = random_state
         self.batch_size = batch_size
-        self.verbose = verbose
 
     def _fit(self, X, Y):
         X, X_val, Y, Y_val = train_test_split(
@@ -76,7 +74,7 @@ class MSG(_PLS):
         for epoch in range(self.epochs):
             start_time = time.time()
             for b in range(num_batches):
-                _,(X_i, Y_i) = next(batches)
+                _, (X_i, Y_i) = next(batches)
                 U, V = update(X_i, Y_i, U, V, self.n_components, lr=self.lr)
                 obj = TV(X, Y, U, V)
                 if self.wandb:
