+++ ccagame/pls/game.py
@@ -151,15 +151,13 @@ class Game(_PLS):
             verbose=False,
             wandb=False
     ):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb)
+        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
         self.lr = lr
         self.epochs = epochs
         self.riemannian_projection = riemannian_projection
-        self.random_state = random_state
         self.simultaneous = simultaneous
         self.batch_size = batch_size
         self.mu = mu
-        self.verbose = verbose
 
     def _fit(self, X, Y):
         """
@@ -185,7 +183,7 @@ class Game(_PLS):
             for epoch in range(self.epochs):
                 start_time = time.time()
                 for b in range(num_batches):
-                    _,(X_i, Y_i) = next(batches)
+                    _, (X_i, Y_i) = next(batches)
                     for k_ in range(self.n_components):
                         u, v = update(
                             U[:, k_],
@@ -218,7 +216,7 @@ class Game(_PLS):
                 for epoch in range(self.epochs):
                     start_time = time.time()
                     for b in range(num_batches):
-                        _,(X_i, Y_i) = next(batches)
+                        _, (X_i, Y_i) = next(batches)
                         u, v = update(
                             U[:, k_],
                             V[:, k_],
@@ -235,7 +233,7 @@ class Game(_PLS):
                         V = V.at[:, k_].set(v)
                         obj = TV(X, Y, U, V)
                         if self.wandb:
-                            wandb.log({f"Iteration/Objective/{k_}": obj},step=b)
+                            wandb.log({f"Iteration/Objective/{k_}": obj}, step=b)
                         else:
                             self.obj.append(obj)
                     obj = TV(X, Y, U, V)
