+++ ccagame/cca/game.py
@@ -109,7 +109,7 @@ class Game(_CCA):
             lr: float = 1.0,
             epochs: int = 100,
             riemannian_projection: bool = False,
-            random_state: int = 0,
+            random_state: int = None,
             simultaneous: bool = True,
             batch_size: int = 128,
             mu=True,
