+++ ccagame/pls/sgd.py
@@ -46,7 +46,7 @@ class SGD(_PLS):
             copy=True,
             lr: float = 1,
             epochs: int = 100,
-            random_state: int = 0,
+            random_state: int = None,
             batch_size: int = 128,
             verbose=False,
             wandb=False
