+++ ccagame/pls/game.py
@@ -144,7 +144,7 @@ class Game(_PLS):
             lr: float = 1,
             epochs: int = 100,
             riemannian_projection: bool = False,
-            random_state: int = 0,
+            random_state: int = None,
             simultaneous: bool = True,
             batch_size: int = 128,
             mu=True,
