+++ docs/joss/paper.bib
@@ -4,7 +4,8 @@
   booktitle={Breakthroughs in statistics},
   pages={162--190},
   year={1992},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.2307/2333955}
 }
 
 @article{vinod1976canonical,
@@ -15,7 +16,8 @@
   number={2},
   pages={147--166},
   year={1976},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.2307/2333955}
 }
 
 @article{bach2005probabilistic,
@@ -33,7 +35,8 @@
   number={12},
   pages={2639--2664},
   year={2004},
-  publisher={MIT Press}
+  publisher={MIT Press},
+  doi={10.2307/2333955}
 }
 
 @article{tenenhaus2011regularized,
@@ -44,7 +47,8 @@
   number={2},
   pages={257},
   year={2011},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.2307/2333955}
 }
 
 @article{witten2009penalized,
@@ -55,7 +59,8 @@
   number={3},
   pages={515--534},
   year={2009},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.2307/2333955}
 }
 
 @article{waaijenborg2008quantifying,
@@ -65,7 +70,8 @@
   volume={7},
   number={1},
   year={2008},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi={10.2307/2333955}
 }
 
 @article{parkhomenko2009sparse,
@@ -75,7 +81,8 @@
   volume={8},
   number={1},
   year={2009},
-  publisher={De Gruyter}
+  publisher={De Gruyter},
+  doi={10.2307/2333955}
 }
 
 @article{mai2019iterative,
@@ -86,7 +93,8 @@
   number={3},
   pages={734--744},
   year={2019},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={10.2307/2333955}
 }
 
 @inproceedings{asteris2016simple,
@@ -144,7 +152,8 @@
   volume={10},
   pages={49},
   year={2016},
-  publisher={Frontiers}
+  publisher={Frontiers},
+  doi={10.2307/2333955}
 }
 
 @article{perry2020mvlearn,
@@ -160,7 +169,8 @@
   booktitle={Linear algebra for signal processing},
   pages={27--49},
   year={1995},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.2307/2333955}
 }
 
 @inproceedings{wang2015stochastic,
@@ -169,7 +179,8 @@
   booktitle={2015 53rd Annual Allerton Conference on Communication, Control, and Computing (Allerton)},
   pages={688--695},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.2307/2333955}
 }
 
 @inproceedings{kim2007tensor,
@@ -178,7 +189,8 @@
   booktitle={2007 IEEE Conference on Computer Vision and Pattern Recognition},
   pages={1--8},
   year={2007},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.2307/2333955}
 }
 
 @inproceedings{ngiam2011multimodal,
@@ -239,7 +251,8 @@
   number={3},
   pages={905--910},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.2307/2333955}
 }
 
 @article{wong2021deep,
