+++ docs/joss/paper.bib
@@ -17,7 +17,7 @@
   pages={147--166},
   year={1976},
   publisher={Elsevier},
-  doi={10.2307/2333955}
+  doi={10.1016/0304-4076(76)90010-5}
 }
 
 @article{bach2005probabilistic,
@@ -36,7 +36,7 @@
   pages={2639--2664},
   year={2004},
   publisher={MIT Press},
-  doi={10.2307/2333955}
+  doi={10.1162/0899766042321814}
 }
 
 @article{tenenhaus2011regularized,
@@ -48,7 +48,7 @@
   pages={257},
   year={2011},
   publisher={Springer},
-  doi={10.2307/2333955}
+  doi={10.1007/s11336-011-9206-8}
 }
 
 @article{witten2009penalized,
@@ -60,7 +60,7 @@
   pages={515--534},
   year={2009},
   publisher={Oxford University Press},
-  doi={10.2307/2333955}
+  doi={10.1093/biostatistics/kxp008}
 }
 
 @article{waaijenborg2008quantifying,
@@ -71,7 +71,7 @@
   number={1},
   year={2008},
   publisher={De Gruyter},
-  doi={10.2307/2333955}
+  doi={10.2202/1544-6115.1329}
 }
 
 @article{parkhomenko2009sparse,
@@ -82,7 +82,7 @@
   number={1},
   year={2009},
   publisher={De Gruyter},
-  doi={10.2307/2333955}
+  doi={10.2202/1544-6115.1406}
 }
 
 @article{mai2019iterative,
@@ -94,7 +94,7 @@
   pages={734--744},
   year={2019},
   publisher={Wiley Online Library},
-  doi={10.2307/2333955}
+  doi={10.1111/biom.13043}
 }
 
 @inproceedings{asteris2016simple,
@@ -153,7 +153,7 @@
   pages={49},
   year={2016},
   publisher={Frontiers},
-  doi={10.2307/2333955}
+  doi={10.3389/fninf.2016.00049}
 }
 
 @article{perry2020mvlearn,
@@ -170,7 +170,7 @@
   pages={27--49},
   year={1995},
   publisher={Springer},
-  doi={10.2307/2333955}
+  doi={10.1007/978-1-4612-4228-4_3}
 }
 
 @inproceedings{wang2015stochastic,
@@ -180,7 +180,7 @@
   pages={688--695},
   year={2015},
   organization={IEEE},
-  doi={10.2307/2333955}
+  doi={10.1109/allerton.2015.7447071}
 }
 
 @inproceedings{kim2007tensor,
@@ -190,7 +190,7 @@
   pages={1--8},
   year={2007},
   organization={IEEE},
-  doi={10.2307/2333955}
+  doi={10.1109/cvpr.2007.383137}
 }
 
 @inproceedings{ngiam2011multimodal,
@@ -252,7 +252,7 @@
   pages={905--910},
   year={2007},
   publisher={IEEE},
-  doi={10.2307/2333955}
+  doi={10.1109/tnn.2007.891186}
 }
 
 @article{wong2021deep,
