### 2021-10-15 a091973
Fix random state by pushing utils into classes as static methods
