+++ ccagame/cca/msg.py
@@ -12,7 +12,6 @@ from sklearn.model_selection import train_test_split
 
 from ccagame.utils import data_stream, get_num_batches
 from . import _CCA
-from .utils import TCC, initialize
 
 
 def invsqrtm(M):
@@ -77,7 +76,7 @@ class MSG(_CCA):
         X, X_val, Y, Y_val = train_test_split(
             X, Y, random_state=self.random_state, train_size=0.9
         )
-        U, V = initialize(X, Y, self.n_components, "random", self.random_state)
+        U, V = self.initialize(X, Y, self.n_components, "random", self.random_state)
         Cx = jnp.zeros((X.shape[1], X.shape[1]))
         Cy = jnp.zeros((Y.shape[1], Y.shape[1]))
         batches = data_stream(X, Y, batch_size=self.batch_size)
@@ -88,12 +87,12 @@ class MSG(_CCA):
             for b in range(num_batches):
                 _, (X_i, Y_i) = next(batches)
                 U, V, Cx, Cy = update(X_i, Y_i, U, V, Cx, Cy, b, self.n_components, lr=self.lr)
-                obj = TCC(X, Y, U, V)
+                obj = self.TCC(X, Y, U, V)
                 if self.wandb:
                     wandb.log({"Iteration/Objective": obj}, step=b)
                 else:
                     self.obj.append(obj)
-            obj = TCC(X, Y, U, V)
+            obj = self.TCC(X, Y, U, V)
             if self.wandb:
                 wandb.log({"Epoch/Objective": obj}, step=epoch)
             if self.verbose:
