+++ ccagame/pls/_pls.py
@@ -1,6 +1,9 @@
 import time
 from abc import abstractmethod
 
+import jax.numpy as jnp
+import jax.scipy as jsp
+from jax import random
 from sklearn.base import (
     BaseEstimator,
     TransformerMixin,
@@ -8,7 +11,7 @@ from sklearn.base import (
     RegressorMixin,
 )
 
-from .utils import TV
+from ..utils import check_random_state
 
 
 class _PLS(BaseEstimator, TransformerMixin, MultiOutputMixin, RegressorMixin):
@@ -18,7 +21,7 @@ class _PLS(BaseEstimator, TransformerMixin, MultiOutputMixin, RegressorMixin):
         self.copy = copy
         self.wandb = wandb
         self.verbose = verbose
-        self.random_state = random_state
+        self.random_state = check_random_state(random_state)
 
     @abstractmethod
     def _fit(self, X, Y):
@@ -39,7 +42,7 @@ class _PLS(BaseEstimator, TransformerMixin, MultiOutputMixin, RegressorMixin):
         raise NotImplementedError
 
     def score(self, X, y, sample_weight=None):
-        return TV(X, y, self.x_weights, self.y_weights)
+        return self.TV(X, y, self.x_weights, self.y_weights)
 
     def center_scale(self, X, Y):
         x_mean = X.mean(axis=0)
@@ -52,3 +55,77 @@ class _PLS(BaseEstimator, TransformerMixin, MultiOutputMixin, RegressorMixin):
             X /= x_std
             Y /= y_std
         return X, Y, x_mean, y_mean, x_std, y_std
+
+    @staticmethod
+    def initialize(X, Y, k, type="uniform", random_state=None):
+        if type == "svd":
+            U1, _, V1 = jnp.linalg.svd(X.T @ Y)
+            U1 = U1[:, :k]
+            V1 = V1[:, :k]
+        elif type == "uniform":
+            U1 = jnp.ones((X.shape[1], k))
+            V1 = jnp.ones((Y.shape[1], k))
+            U1 = U1 / jnp.linalg.norm(U1, axis=0)
+            V1 = V1 / jnp.linalg.norm(V1, axis=0)
+        elif type == "random":
+            key, subkey = random.split(random_state)
+            U1 = random.normal(key, (X.shape[1], k))
+            V1 = random.normal(subkey, (Y.shape[1], k))
+            U1 = U1 / jnp.linalg.norm(U1, axis=0)
+            V1 = V1 / jnp.linalg.norm(V1, axis=0)
+        else:
+            print(f'Initialization "{type}" not implemented')
+            return
+        return U1, V1
+
+    @staticmethod
+    def TV(X, Y, Wx, Wy):
+        X_hat = X @ Wx
+        Y_hat = Y @ Wy
+        C = X_hat.T @ Y_hat
+        _, S, _ = jnp.linalg.svd(C)
+        return S.sum()
+
+    @staticmethod
+    def TCC(X, Y, Wx, Wy):
+        dof = X.shape[0] - 1
+        X_hat = X @ Wx
+        Y_hat = Y @ Wy
+        C = jnp.hstack((X_hat, Y_hat))
+        C = C.T @ C / dof
+        # Get the block covariance matrix placing Xi^TX_i on the diagonal
+        D = jsp.linalg.block_diag(*[m.T @ m for i, m in enumerate([X_hat, Y_hat])]) / dof
+        C = C - jsp.linalg.block_diag(*[view.T @ view / dof for view in [X_hat, Y_hat]]) + D
+        R = jnp.linalg.inv(jnp.linalg.cholesky(D))
+        # In MCCA our eigenvalue problem Cv = lambda Dv
[…]
