+++ ccagame/pls/game.py
@@ -9,7 +9,6 @@ from sklearn.model_selection import train_test_split
 
 from ccagame.utils import data_stream, get_num_batches
 from . import _PLS
-from .utils import initialize, TV
 
 
 @partial(jit, static_argnums=(6))
@@ -174,7 +173,7 @@ class Game(_PLS):
         X, X_val, Y, Y_val = train_test_split(
             X, Y, random_state=self.random_state, train_size=0.9
         )
-        U, V = initialize(X, Y, self.n_components, "random", self.random_state)
+        U, V = self.initialize(X, Y, self.n_components, "random", self.random_state)
         batches = data_stream(X, Y, batch_size=self.batch_size)
         num_batches = get_num_batches(X, Y, batch_size=self.batch_size)
         self.obj = []
@@ -199,12 +198,12 @@ class Game(_PLS):
                         )
                         U = U.at[:, k_].set(u)
                         V = V.at[:, k_].set(v)
-                    obj = TV(X, Y, U, V)
+                    obj = self.TV(X, Y, U, V)
                     if self.wandb:
                         wandb.log({"Iteration/Objective": obj}, step=b)
                     else:
                         self.obj.append(obj)
-                obj = TV(X, Y, U, V)
+                obj = self.TV(X, Y, U, V)
                 if self.wandb:
                     wandb.log({"Epoch/Objective": obj}, step=epoch)
                 if self.verbose:
@@ -231,12 +230,12 @@ class Game(_PLS):
                         )
                         U = U.at[:, k_].set(u)
                         V = V.at[:, k_].set(v)
-                        obj = TV(X, Y, U, V)
+                        obj = self.TV(X, Y, U, V)
                         if self.wandb:
                             wandb.log({f"Iteration/Objective/{k_}": obj}, step=b)
                         else:
                             self.obj.append(obj)
-                    obj = TV(X, Y, U, V)
+                    obj = self.TV(X, Y, U, V)
                     if self.wandb:
                         wandb.log({f"Epoch/Objective/{k_}": obj})
                     if self.verbose:
