### 2021-10-15 82a8b66
Fix random state by pushing utils into classes as static methods
