+++ ccagame/cca/_cca.py
@@ -1,4 +1,4 @@
-from ccagame.cca.utils import TCC
+
 from ccagame.pls import _PLS
 
 
