### 2021-10-15 377f453
Fix random state by pushing utils into classes as static methods
