### 2021-10-15 67c2b86
Fix random state by pushing utils into classes as static methods
