### 2021-10-15 89d45e6
Fix random state by pushing utils into classes as static methods
