+++ ccagame/pls/incremental.py
@@ -73,7 +73,7 @@ class Incremental(_PLS):
         self.lr = lr
         self.epochs = epochs
 
-    def _fit(self, X, Y):
+    def _fit(self, X, Y, X_val=None, Y_val=None):
         U, V = self.initialize(X, Y, self.n_components, "random", self.random_state)
         batches = data_stream(X, Y, batch_size=1)
         num_batches = get_num_batches(X, Y, batch_size=1)
@@ -84,12 +84,12 @@ class Incremental(_PLS):
             for b in range(num_batches):
                 _, (X_i, Y_i) = next(batches)
                 U, S, V = update(X_i, Y_i, U, S, V, self.n_components)
-                obj = self.TV(X, Y, U, V)
+                obj = self.TV(X@U, Y@V)
                 if self.wandb:
                     wandb.log({"Iteration/Objective": obj}, step=b)
                 else:
                     self.obj.append(obj)
-            obj = self.TV(X, Y, U, V)
+            obj = self.TV(X@U, Y@V)
             if self.wandb:
                 wandb.log({"Epoch/Objective": obj}, step=epoch)
             if self.verbose:
