+++ ccagame/cca/napi.py
@@ -43,7 +43,7 @@ class NAPI(_CCA):
         self.lr = lr
         self.solver=solver
 
-    def _fit(self, X, Y):
+    def _fit(self, X, Y, X_val=None, Y_val=None):
         p = X.shape[1]
         A, B = self.initialize_gep(X, Y)
         W, V = self.initialize(

[…]
