### 2021-10-15 ebfcb3b
Fix random state by pushing utils into classes as static methods
