### 2021-10-15 63326e2
Fix random state by pushing utils into classes as static methods
