+++ ccagame/pls/incremental.py
@@ -84,16 +84,16 @@ class Incremental(_PLS):
             for b in range(num_batches):
                 _, (X_i, Y_i) = next(batches)
                 U, S, V = update(X_i, Y_i, U, S, V, self.n_components)
-                obj = self.TV(X@U, Y@V)
+                obj_tr = self.TV(X @ U, Y @ V)
+                obj_val = self.TV(X_val @ U, Y_val @ V)
                 if self.wandb:
-                    wandb.log({"Iteration/Objective": obj}, step=b)
+                    wandb.log({"Iteration/Objective (Train)": obj_tr,
+                               "Iteration/Objective (Val)": obj_val}, step=b)
                 else:
-                    self.obj.append(obj)
-            obj = self.TV(X@U, Y@V)
-            if self.wandb:
-                wandb.log({"Epoch/Objective": obj}, step=epoch)
+                    self.obj.append([obj_tr, obj_val])
             if self.verbose:
                 epoch_time = time.time() - start_time
                 print(f"Epoch {epoch} in {epoch_time} sec")
-                print(f"epoch {epoch} objective: {obj}")
+                print(f"Epoch {epoch} objective (Train): {obj_tr}")
+                print(f"Epoch {epoch} objective (Train): {obj_val}")
         return U, V
