### 2021-10-15 3f0f9d1
Fix random state by pushing utils into classes as static methods
