+++ ccagame/cca/ccalin.py
@@ -51,8 +51,8 @@ class CCALin(_CCA):
             start_time = time.time()
             W = update(A, B, W)
             W = self.gram_schmidt_matrix(W, B)
-            obj_tr = self.TV(X @ W[:p], Y @ W[p:])
-            obj_val = self.TV(X_val @ W[:p], Y_val @ W[p:])
+            obj_tr = self.TCC(X @ W[:p], Y @ W[p:])
+            obj_val = self.TCC(X_val @ W[:p], Y_val @ W[p:])
             if self.wandb:
                 wandb.log({"Iteration/Objective (Train)": obj_tr,
                            "Iteration/Objective (Val)": obj_val}, step=epoch)
