### 2021-10-15 4683208
Fix random state by pushing utils into classes as static methods
