### 2021-10-15 3baf937
Callbacks as standard method to reduce lines of code
