+++ ccagame/cca/lagrange.py
@@ -52,16 +52,7 @@ class Lagrange(_CCA):
             W = update(A, B, W, self.lr)
             obj_tr = self.TCC(X @ W[:p], Y @ W[p:])
             obj_val = self.TCC(X_val @ W[:p], Y_val @ W[p:])
-            if self.wandb:
-                wandb.log({"Iteration/Objective (Train)": obj_tr,
-                           "Iteration/Objective (Val)": obj_val}, step=epoch)
-            else:
-                self.obj.append([obj_tr, obj_val])
-            if self.verbose:
-                epoch_time = time.time() - start_time
-                print(f"Epoch {epoch} in {epoch_time} sec")
-                print(f"Epoch {epoch} objective (Train): {obj_tr}")
-                print(f"Epoch {epoch} objective (Train): {obj_val}")
+            self.callback(obj_tr, obj_val, epoch, start_time)
         U = self.gram_schmidt_matrix(W[:p], B[:p, :p])
         V = self.gram_schmidt_matrix(W[p:], B[p:, p:])
         return U, V
