+++ ccagame/cca/appgrad.py
@@ -49,14 +49,5 @@ class AppGrad(_CCA):
             U, V, phi_, psi_ = update(U, V, phi_, psi_)
             obj_tr = self.TCC(X @ U, Y @ V)
             obj_val = self.TCC(X_val @ U, Y_val @ V)
-            if self.wandb:
-                wandb.log({"Iteration/Objective (Train)": obj_tr,
-                           "Iteration/Objective (Val)": obj_val}, step=epoch)
-            else:
-                self.obj.append([obj_tr, obj_val])
-            if self.verbose:
-                epoch_time = time.time() - start_time
-                print(f"Epoch {epoch} in {epoch_time} sec")
-                print(f"Epoch {epoch} objective (Train): {obj_tr}")
-                print(f"Epoch {epoch} objective (Train): {obj_val}")
+            self.callback(obj_tr, obj_val, epoch, start_time)
         return U, V

[…]
