### 2021-10-15 793f2ab
Callbacks as standard method to reduce lines of code
