+++ ccagame/pca/gha.py
@@ -51,7 +51,7 @@ class GHA(_PCA):
         self.epochs = epochs
         self.batch_size = batch_size
 
-    def _fit(self, X):
+    def _fit(self, X, X_val):
         U = self.initialize(
             X, self.n_components, type="random", random_state=self.random_state
         )
@@ -63,16 +63,8 @@ class GHA(_PCA):
             for _ in range(num_batches):
                 _, X = next(batches)
                 U = update(U, X, lr=self.lr)
-                obj = self.TV(X)
-                if self.wandb:
-                    wandb.log({"Iteration/Objective": obj})
-                else:
-                    self.obj.append(obj)
-            obj = self.TV(X)
-            if self.wandb:
-                wandb.log({"Epoch/Objective": obj})
-            if self.verbose:
-                epoch_time = time.time() - start_time
-                print(f"Epoch {epoch} in {epoch_time} sec")
-                print(f"epoch {epoch}: {obj}")
+                obj_tr = self.TV(X @ U)
+                obj_val = self.TV(X_val @ U)
+            self.callback(obj_tr, obj_val, b)
+            self.callback(obj_tr, obj_val, epoch, start_time)
         return U
