+++ ccagame/pca/krasulina.py
@@ -54,7 +54,7 @@ class Krasulina(_PCA):
         self.epochs = epochs
         self.batch_size = batch_size
 
-    def _fit(self, X):
+    def _fit(self, X, X_val):
         U = self.initialize(
             X, self.n_components, type="random", random_state=self.random_state
         )
@@ -63,19 +63,11 @@ class Krasulina(_PCA):
         self.obj = []
         for epoch in range(self.epochs):
             start_time = time.time()
-            for _ in range(num_batches):
+            for b in range(num_batches):
                 _, X_i = next(batches)
                 U = update(U, X_i, lr=self.lr)
-                obj = self.TV(X)
-                if self.wandb:
-                    wandb.log({"Iteration/Objective": obj})
-                else:
-                    self.obj.append(obj)
-            obj = self.TV(X)
-            if self.wandb:
-                wandb.log({"Epoch/Objective": obj})
-            if self.verbose:
-                epoch_time = time.time() - start_time
-                print(f"Epoch {epoch} in {epoch_time} sec")
-                print(f"epoch {epoch}: {obj}")
+                obj_tr = self.TV(X @ U)
+                obj_val = self.TV(X_val @ U)
+                self.callback(obj_tr, obj_val, b)
+            self.callback(obj_tr, obj_val, epoch, start_time)
         return U
