+++ ccagame/pls/msg.py
@@ -6,9 +6,7 @@ import time
 from functools import partial
 
 import jax.numpy as jnp
-import wandb
 from jax import jit
-from sklearn.model_selection import train_test_split
 
 from ccagame.utils import data_stream, get_num_batches
 from . import _PLS
