+++ ccagame/cca/_cca.py
@@ -1,4 +1,3 @@
-
 from ccagame.pls import _PLS
 
 
