+++ ccagame/pls/game.py
@@ -3,9 +3,7 @@ import time
 from functools import partial
 
 import jax.numpy as jnp
-import wandb
 from jax import grad, jit
-from sklearn.model_selection import train_test_split
 
 from ccagame.utils import data_stream, get_num_batches
 from . import _PLS
