+++ ccagame/cca/game.py
@@ -150,8 +150,8 @@ class Game(_CCA):
                         V = V.at[:, k_].set(v)
                         obj_tr = self.TCC(X @ U, Y @ V)
                         obj_val = self.TCC(X_val @ U, Y_val @ V)
-                        self.callback(obj_tr, obj_val, b)
-                    self.callback(obj_tr, obj_val, b, start_time)
+                        self.callback(obj_tr, obj_val)
+                    self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         else:
             for k_ in range(self.n_components):
                 for epoch in range(self.epochs):
@@ -174,6 +174,6 @@ class Game(_CCA):
                         V = V.at[:, k_].set(v)
                         obj_tr = self.TCC(X @ U, Y @ V)
                         obj_val = self.TCC(X_val @ U, Y_val @ V)
-                        self.callback(obj_tr, obj_val, b)
-                    self.callback(obj_tr, obj_val, b, start_time)
+                        self.callback(obj_tr, obj_val)
+                    self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         return U, V
