+++ ccagame/pls/sgd.py
@@ -66,6 +66,6 @@ class SGD(_PLS):
                 V = update(Y_i, X_i, V, U, lr=self.lr)
                 obj_tr = self.TV(X @ U, Y @ V)
                 obj_val = self.TV(X_val @ U, Y_val @ V)
-                self.callback(obj_tr, obj_val, b)
-            self.callback(obj_tr, obj_val, b, start_time)
+                self.callback(obj_tr, obj_val)
+            self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         return U, V
