+++ ccagame/cca/ccalin.py
@@ -53,7 +53,7 @@ class CCALin(_CCA):
             W = self.gram_schmidt_matrix(W, B)
             obj_tr = self.TCC(X @ W[:p], Y @ W[p:])
             obj_val = self.TCC(X_val @ W[:p], Y_val @ W[p:])
-            self.callback(obj_tr, obj_val, epoch, start_time)
+            self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         U = self.gram_schmidt_matrix(W[:p], B[:p, :p])
         V = self.gram_schmidt_matrix(W[p:], B[p:, p:])
         return U, V
