+++ ccagame/pca/krasulina.py
@@ -67,6 +67,6 @@ class Krasulina(_PCA):
                 U = update(U, X_i, lr=self.lr)
                 obj_tr = self.TV(X @ U)
                 obj_val = self.TV(X_val @ U)
-                self.callback(obj_tr, obj_val, b)
-            self.callback(obj_tr, obj_val, epoch, start_time)
+                self.callback(obj_tr, obj_val)
+            self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         return U

[…]
