+++ scripts/pls/example.py
@@ -14,7 +14,7 @@ n = 100
 p = 10
 q = 11
 latent_dims = 3
-epochs = 50
+epochs = 20
 riemannian_projection = False
 lr = 1e-2
 batch_size = 100
