+++ scripts/cca/example.py
@@ -17,7 +17,7 @@ p = 10
 q = 11
 latent_dims = 5
 batch_size = 5
-epochs = 10
+epochs = 20
 riemannian_projection = True
 initialization = 'random'
 lr = 1e-1
