### 2021-10-15 4fad99b
Small fixes to CCA methods
