+++ ccagame/cca/genoja.py
@@ -53,7 +53,9 @@ class Genoja(_CCA):
         for epoch in range(self.epochs):
             start_time = time.time()
             W, P = update(A, B, W, P, self.beta_0 / (1 + 1e-4 * epoch), self.alpha)
-            obj_tr = self.TCC(X @ W[:p], Y @ W[p:])
+            U = self.gram_schmidt_matrix(W[:p], B[:p, :p])
+            V = self.gram_schmidt_matrix(W[p:], B[p:, p:])
+            obj_tr = self.TCC(X @ U, Y @ V)
             obj_val = self.TCC(X_val @ W[:p], Y_val @ W[p:])
             self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
         U = self.gram_schmidt_matrix(W[:p], B[:p, :p])
