### 2021-10-26 cc1c033
update ukbb script
