+++ scripts/pls/uk_biobank.py
@@ -10,9 +10,9 @@ from ccagame.pls import Game, SGD, Incremental, Batch, Numpy, MSG
 import numpy as np
 import os
 import matplotlib.pyplot as plt
-import numpy as np
 
-os.chdir('ukbb_results')
+os.chdir('./results/ukbb_results/211025_v2')
+print(os.getcwd())
 # %%
 
 # Parameters
@@ -25,53 +25,65 @@ epochs = 1
 riemannian_projection = False
 lr = 1e-3
 batch_size = 1
-path = 'C:/Users/anala/Documents/PhD/year_1/project_work/Epilepsy/data/PLS/210326'
 # This implements the version with unbiased updates analagous to Eigengame:Unloaded
-train_1, train_2, test_1, test_2 = datasets.ukbiobank(path)
-
+train_1, train_2 = datasets.ukbiobank(path=os.getcwd(), save=True)
+print(train_1.shape)
 # Model
 numpy = Numpy(scale=False, n_components=latent_dims).fit(train_1, train_2)
 print("\n Eigenvalues calculated using numpy are :\n", numpy.score(train_1, train_2))
 print("\n Time :\n", numpy.fit_time)
 
-game = Game(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True,
+game = Game(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=False,
             mu=True).fit(train_1, train_2)
 print("\n Eigenvalues calculated using game are :\n", game.score(train_1, train_2))
 print("\n Time :\n", game.fit_time)
-np.save('game', game.obj)
+np.save('game_train', list(zip(*game.obj))[0])
+np.save('game_val', list(zip(*game.obj))[1])
 
-sgd = SGD(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True).fit(train_1,
+sgd = SGD(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=False).fit(train_1,
                                                                                                                 train_2)
 print("\n Eigenvalues calculated using sgd are :\n", sgd.score(train_1, train_2))
 print("\n Time :\n", sgd.fit_time)
-np.save('sgd', sgd.obj)
 
-incremental = Incremental(scale=False, lr=lr, epochs=epochs, n_components=latent_dims, verbose=True).fit(train_1,
+np.save('sgd_train', list(zip(*sgd.obj))[0])
+np.save('sgd_val',list(zip(*sgd.obj))[1])
+
+incremental = Incremental(scale=False, lr=lr, epochs=epochs, n_components=latent_dims, verbose=False).fit(train_1,
                                                                                                          train_2)
 print("\n Eigenvalues calculated using incremental are :\n", incremental.score(train_1, train_2))
 print("\n Time :\n", incremental.fit_time)
-np.save('inc', incremental.obj)
+np.save('inc_train', list(zip(*incremental.obj))[0])
+np.save('inc_val', list(zip(*incremental.obj))[1])
 
-msg = MSG(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True).fit(train_1,
+msg = MSG(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=False).fit(train_1,
                                                                                                                 train_2)
[…]
