+++ ccagame/cca/game.py
@@ -93,6 +93,8 @@ def update(
     else:
         du = grad(alpha_model)(u, X, U, T, k)
         dv = grad(alpha_model)(v, Y, V, T, k)
+    du = du * X.shape[0]
+    dv = dv * X.shape[0]
     uhat = u + lr * du
     vhat = v + lr * dv
     return uhat, vhat
@@ -130,7 +132,6 @@ class Game(_CCA):
         self.obj = []
         if self.simultaneous:
             for epoch in range(self.epochs):
-                # T = update_T(X, Y, U, V)
                 start_time = time.time()
                 for b in range(num_batches):
                     idx, (X_i, Y_i) = next(batches)
@@ -143,6 +144,7 @@ class Game(_CCA):
                             U,
                             V,
                             k_,
+                            X.shape[0],
                             lr=self.lr,
                             mu=self.mu
                         )
@@ -166,6 +168,7 @@ class Game(_CCA):
                             U,
                             V,
                             k_,
+                            X.shape[0],
                             lr=self.lr,
                             mu=self.mu
                         )
