### 2021-10-26 71256e0
Added in a test to check things run on commits
