+++ ccagame/cca/game.py
@@ -93,8 +93,6 @@ def update(
     else:
         du = grad(alpha_model)(u, X, U, T, k)
         dv = grad(alpha_model)(v, Y, V, T, k)
-    du = du * X.shape[0]
-    dv = dv * X.shape[0]
     uhat = u + lr * du
     vhat = v + lr * dv
     return uhat, vhat
@@ -144,7 +142,6 @@ class Game(_CCA):
                             U,
                             V,
                             k_,
-                            X.shape[0],
                             lr=self.lr,
                             mu=self.mu
                         )
