### 2021-10-26 f67fe24
Added in a test to check things run on commits
