+++ ccagame/cca/appgrad.py
@@ -28,17 +28,24 @@ def update(X, Y, phi, psi, phi_, psi_, lr):
 
 class AppGrad(_CCA):
     def __init__(
-            self,
-            n_components=2,
-            *,
-            scale=True,
-            copy=True,
-            epochs: int = 100,
-            random_state: int = None,
-            verbose=False,
-            wandb=False
+        self,
+        n_components=2,
+        *,
+        scale=True,
+        copy=True,
+        epochs: int = 100,
+        random_state: int = None,
+        verbose=False,
+        wandb=False
     ):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
+        super().__init__(
+            n_components,
+            scale=scale,
+            copy=copy,
+            wandb=wandb,
+            verbose=verbose,
+            random_state=random_state,
+        )
         self.epochs = epochs
 
     def _fit(self, X, Y, X_val=None, Y_val=None):
