+++ ccagame/cca/game.py
@@ -68,15 +68,15 @@ def mu_model(u, X, U, T, k: int):
 # Update rule to be used for calculating eigenvectors
 @partial(jit, static_argnums=6, static_argnames=("lr", "mu"))
 def update(
-        u,
-        v,
-        X,
-        Y,
-        U,
-        V,
-        k: int,
-        lr: float = 1,
-        mu=True,
+    u,
+    v,
+    X,
+    Y,
+    U,
+    V,
+    k: int,
+    lr: float = 1,
+    mu=True,
 ):
     """
     Update the left and right singular vector estimates
@@ -100,22 +100,29 @@ def update(
 
 class Game(_CCA):
     def __init__(
-            self,
-            n_components=4,
-            *,
-            scale=True,
-            copy=True,
-            lr: float = 1.0,
-            epochs: int = 100,
-            riemannian_projection: bool = False,
-            random_state: int = None,
-            simultaneous: bool = True,
-            batch_size: int = 128,
-            mu=True,
-            verbose=False,
-            wandb=False
+        self,
+        n_components=4,
+        *,
+        scale=True,
+        copy=True,
+        lr: float = 1.0,
+        epochs: int = 100,
+        riemannian_projection: bool = False,
+        random_state: int = None,
+        simultaneous: bool = True,
+        batch_size: int = 128,
+        mu=True,
+        verbose=False,
+        wandb=False
     ):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
+        super().__init__(
+            n_components,
+            scale=scale,
+            copy=copy,
+            wandb=wandb,
+            verbose=verbose,
+            random_state=random_state,
+        )
         self.lr = lr
         self.epochs = epochs
         self.riemannian_projection = riemannian_projection
@@ -143,7 +150,7 @@ class Game(_CCA):
                             V,
                             k_,
                             lr=self.lr,
-                            mu=self.mu
+                            mu=self.mu,
                         )
                         U = U.at[:, k_].set(u)
                         V = V.at[:, k_].set(v)
@@ -167,7 +174,7 @@ class Game(_CCA):
                             k_,
                             X.shape[0],
                             lr=self.lr,
-                            mu=self.mu
+                            mu=self.mu,
                         )
                         # T,
                         U = U.at[:, k_].set(u)
