+++ ccagame/cca/_cca.py
@@ -2,9 +2,26 @@ from ccagame.pls import _PLS
 import jax.numpy as jnp
 import jax.scipy as jsp
 
+
 class _CCA(_PLS):
-    def __init__(self, n_components=2, *, scale=True, copy=True, wandb=True, verbose=False, random_state=None):
-        super().__init__(n_components, scale=scale, copy=copy, wandb=wandb, verbose=verbose, random_state=random_state)
+    def __init__(
+        self,
+        n_components=2,
+        *,
+        scale=True,
+        copy=True,
+        wandb=True,
+        verbose=False,
+        random_state=None
+    ):
+        super().__init__(
+            n_components,
+            scale=scale,
+            copy=copy,
+            wandb=wandb,
+            verbose=verbose,
+            random_state=random_state,
+        )
 
     def score(self, X, y=None, sample_weight=None):
         X_hat, Y_hat = self.transform(X, y)
