### 2021-10-26 ccac094
Added in a test to check things run on commits
