+++ .github/workflows/python-publish.yml
@@ -27,11 +27,5 @@ jobs:
         python -m pip install codecov
         python -m pip install pytest-cov
         python -m pip install -r requirements.txt
-    - name: Lint with flake8
-      run: |
-        # stop the build if there are Python syntax errors or undefined names
-        flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Run Test
       run: pytest --cov-report=xml --cov=ccagame ./cca_game/test/
