### 2021-10-26 8df077e
Added in a test to check things run on commits
