### 2021-10-26 4dc6468
Added in a test to check things run on commits
