### 2021-10-26 b8780bf
sparse game removed the gradient scaling thing
