+++ ccagame/pls/sparsegame.py
@@ -127,8 +127,6 @@ def update(
     else:
         du = grad(alpha_model)(u, v, X, Y, U, V, k)
         dv = grad(alpha_model)(v, u, Y, X, V, U, k)
-    du = du * X.shape[0] / n
-    dv = dv * X.shape[0] / n
     if riemannian_projection:
         dur = du - (u.T @ u) * u
         uhat = u + lr * dur
