### 2021-10-26 02dc82e
remove scaling and collecting cov after each batch iteration
