+++ ccagame/pls/game.py
@@ -175,6 +175,8 @@ class Game(_PLS):
         # We can either solve the eigenvectors simulataneously or complete each one
         if self.simultaneous:
             for epoch in range(self.epochs):
+                obj_tr = 0
+                obj_val = 0
                 start_time = time.time()
                 for b in range(num_batches):
                     _, (X_i, Y_i) = next(batches)
@@ -193,13 +195,14 @@ class Game(_PLS):
                         )
                         U = U.at[:, k_].set(u)
                         V = V.at[:, k_].set(v)
-                    obj_tr = self.TV(X @ U, Y @ V)
-                    obj_val = self.TV(X_val @ U, Y_val @ V)
-                    self.callback(obj_tr, obj_val)
-                self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
+                    obj_tr += self.TV(X @ U, Y @ V)
+                    obj_val += self.TV(X_val @ U, Y_val @ V)
+                self.callback(obj_tr/num_batches, obj_val/num_batches, epoch=epoch, start_time=start_time)
         else:
             for k_ in range(self.n_components):
                 for epoch in range(self.epochs):
+                    obj_tr = 0
+                    obj_val = 0
                     start_time = time.time()
                     for b in range(num_batches):
                         _, (X_i, Y_i) = next(batches)
@@ -217,8 +220,8 @@ class Game(_PLS):
                         )
                         U = U.at[:, k_].set(u)
                         V = V.at[:, k_].set(v)
-                        obj_tr = self.TV(X @ U, Y @ V)
-                        obj_val = self.TV(X_val @ U, Y_val @ V)
-                        self.callback(obj_tr, obj_val)
-                    self.callback(obj_tr, obj_val, epoch=epoch, start_time=start_time)
+                        obj_tr += self.TV(X @ U, Y @ V)
+                        obj_val += self.TV(X_val @ U, Y_val @ V)
+                        #self.callback(obj_tr, obj_val)
+                    self.callback(obj_tr/num_batches, obj_val/num_batches, epoch=epoch, start_time=start_time)
         return U, V
