+++ scripts/pls/mnist.py
@@ -6,8 +6,9 @@ import datasets
 from ccagame.pls import Game, SGD, Incremental, Batch, Numpy, MSG
 import numpy as np
 import os
-
-os.chdir('mnist_results')
+import matplotlib.pyplot as plt
+import random
+os.chdir('./results/mnist_results/211026')
 # %%
 
 # Parameters
@@ -16,45 +17,71 @@ n = 100
 p = 10
 q = 11
 latent_dims = 3
-epochs = 1
+epochs = 20
 riemannian_projection = False
 lr = 1e-3
-batch_size = 1
+batch_size = 20
 # This implements the version with unbiased updates analagous to Eigengame:Unloaded
 mu = True
 train, train_labels, test, test_labels = datasets.mnist()
-train_1 = train[:, :392]
-train_2 = train[:, 392:]
+
+random.seed(42)
+idx = random.sample(list(range(0, train.shape[0])), 500)
+train_1 = train[idx, :392]
+train_2 = train[idx, 392:]
 
 # Model
 numpy = Numpy(scale=False, n_components=latent_dims).fit(train_1, train_2)
 print("\n Eigenvalues calculated using numpy are :\n", numpy.score(train_1, train_2))
 print("\n Time :\n", numpy.fit_time)
 
-sgd = SGD(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True).fit(train_1,
+sgd = SGD(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=False).fit(train_1,
                                                                                                                 train_2)
-print("\n Eigenvalues calculated using sgd are :\n", sgd.score(train_1, train_2))
+print("\n Eigenvalues calculated using sgd are :\n", list(zip(*sgd.obj))[0][-1])
 print("\n Time :\n", sgd.fit_time)
-np.save(f'sgd_{batch_size}', sgd.obj)
+np.save('sgd_train', list(zip(*sgd.obj))[0])
+np.save('sgd_val',list(zip(*sgd.obj))[1])
 
-game = Game(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=True,
+game = Game(scale=False, lr=lr, batch_size=batch_size, epochs=epochs, n_components=latent_dims, verbose=False,
             mu=True).fit(train_1, train_2)
-print("\n Eigenvalues calculated using game are :\n", game.score(train_1, train_2))
+print("\n Eigenvalues calculated using game are :\n", list(zip(*game.obj))[0][-1])
 print("\n Time :\n", game.fit_time)
-np.save(f'game_{batch_size}', game.obj)
+np.save('game_train', list(zip(*game.obj))[0])
+np.save('game_val', list(zip(*game.obj))[1])
 
-incremental = Incremental(scale=False, lr=lr, epochs=epochs, n_components=latent_dims, verbose=True).fit(train_1,
+incremental = Incremental(scale=False, lr=lr, epochs=epochs, n_components=latent_dims, verbose=False).fit(train_1,
                                                                                                          train_2)
-print("\n Eigenvalues calculated using incremental are :\n", incremental.score(train_1, train_2))
+print("\n Eigenvalues calculated using incremental are :\n", list(zip(*incremental.obj))[0][-1])
 print("\n Time :\n", incremental.fit_time)
-np.save('inc', incremental.obj)
+np.save('inc_train', list(zip(*incremental.obj))[0])
+np.save('inc_val', list(zip(*incremental.obj))[1])
 
[…]
