### 2021-10-26 13d5e2f
Create plslightningjames.py
Think this works!
