### 2021-10-26 e81ffa3
Delete plslightningjames.py
