### 2021-10-27 daa77a1
Switching wrapper to pytorch-lightning
