+++ cca_zoo/test/test_deepmodels.py
@@ -1,11 +1,14 @@
 import numpy as np
+import pytorch_lightning as pl
+import torch.utils.data
 from sklearn.utils.validation import check_random_state
 from torch import optim, manual_seed
 from torch.utils.data import Subset
 
 from cca_zoo import data
 from cca_zoo.data import Noisy_MNIST_Dataset
-from cca_zoo.deepmodels import DCCA, DCCAE, DVCCA, DCCA_NOI, DTCCA, SplitAE, DeepWrapper
+from cca_zoo.deepmodels import DCCA, DCCAE, DVCCA, DCCA_NOI, DTCCA, SplitAE, CCALightning, get_dataloaders, \
+    process_data
 from cca_zoo.deepmodels import objectives, architectures
 from cca_zoo.models import CCA
 
@@ -16,12 +19,19 @@ Y = rng.rand(200, 10)
 Z = rng.rand(200, 10)
 X_conv = rng.rand(100, 1, 16, 16)
 Y_conv = rng.rand(100, 1, 16, 16)
-train_dataset = data.CCA_Dataset([X, Y])
+dataset = data.CCA_Dataset([X, Y, Z])
+train_dataset, val_dataset = process_data(
+    dataset,
+    val_split=0.2)
+train_dataset_numpy, val_dataset_numpy = process_data(
+    (X, Y, Z),
+    val_split=0.2)
+train_loader, val_loader = get_dataloaders(train_dataset, val_dataset)
+train_loader_numpy, val_loader_numpy = get_dataloaders(train_dataset, val_dataset)
 
 
 def test_input_types():
     latent_dims = 2
-    device = "cpu"
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
     encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
     # DCCA
@@ -30,16 +40,14 @@ def test_input_types():
         encoders=[encoder_1, encoder_2],
         objective=objectives.CCA,
     )
+    dcca_model = CCALightning(dcca_model)
+    trainer = pl.Trainer(gpus=0, max_epochs=5, progress_bar_refresh_rate=1, log_every_n_steps=1, logger=False)
+    trainer.fit(dcca_model, train_loader, val_loader)
+    trainer_numpy = pl.Trainer(gpus=0, max_epochs=5, progress_bar_refresh_rate=1, log_every_n_steps=1, logger=False)
+    trainer_numpy.fit(dcca_model, train_loader_numpy, val_loader_numpy)
 
-    dcca_model = DeepWrapper(dcca_model, device=device)
-    dcca_model.fit(train_dataset, epochs=3)
-    dcca_model.fit(train_dataset, val_dataset=train_dataset, epochs=3)
-    dcca_model.fit((X, Y), val_dataset=(X, Y), epochs=3)
-    dcca_model.fit((X, Y), val_split=0.2, epochs=3)
 
-
-def tutorial_test():
-    # Load MNIST Data
+def test_tutorial():
     N = 500
     latent_dims = 2
     dataset = Noisy_MNIST_Dataset(mnist_type="FashionMNIST", train=True)
@@ -52,14 +60,16 @@ def tutorial_test():
     test_ids = np.arange(min(N, len(test_dataset)))
     np.random.shuffle(test_ids)
     test_dataset = Subset(test_dataset, test_ids)
-    print("DCCA")
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
     encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+    train_loader, val_loader = get_dataloaders(train_dataset, val_dataset)
[…]
