+++ cca_zoo/deepmodels/trainers.py
@@ -1,0 +1,195 @@
+import itertools
+import sys
+from typing import Optional
+
+import numpy as np
+import torch
+from pytorch_lightning import LightningModule
+from torch.utils.data import DataLoader
+
+
+class CCALightning(LightningModule):
+    def __init__(
+            self,
+            model,
+            optimizer='Adam',
+            learning_rate=1e-3,
+            weight_decay=0.1,
+            lr_scheduler=None,
+    ):
+        super().__init__()
+        self.save_hyperparameters()
+        self.model = model
+
+    def forward(self, *args):
+        z = self.encode(*args)
+        return z
+
+    def loss(self, *args, **kwargs):
+        return self.model.loss(*args, **kwargs)
+
+    # Configuration.  Add more for learning schedulers, etc.?
+    def configure_optimizers(self):
+        if isinstance(self.hparams.optimizer, torch.optim.Optimizer):
+            optimizer = self.hparams.optimizer
+        elif self.hparams.optimizer == "Adam":
+            optimizer = torch.optim.Adam(
+                self.parameters(),
+                lr=self.hparams.learning_rate,
+                weight_decay=self.hparams.weight_decay,
+            )
+        elif self.hparams.optimizer == "SGD":
+            # Left out the momentum options for now
+            optimizer = torch.optim.SGD(
+                self.parameters(),
+                lr=self.hparams.learning_rate,
+                weight_decay=self.hparams.weight_decay,
+            )
+        elif self.hparams.optimizer == "LBFGS":
+            optimizer = torch.optim.LBFGS(
+                self.parameters(),
+                # or can have self.hparams.learning_rate with warning if too low.
+                lr=1,
+                tolerance_grad=1e-5,  # can add to parameters if useful.
+                tolerance_change=1e-9,  # can add to parameters if useful.
+            )
+        else:
+            print("Invalid optimizer.  See --help")
+            sys.exit()
+
+        if self.hparams.lr_scheduler == None:
+            return optimizer
+        elif isinstance(self.hparams.lr_scheduler, torch.optim.lr_scheduler):
+            scheduler = self.hparams.lr_scheduler
+        elif self.hparams.lr_scheduler == "StepLR":
+            step_size = self.hparams.StepLR_step_size
+            gamma = self.hparams.StepLR_gamma
+            scheduler = torch.optim.lr_scheduler.StepLR(
+                optimizer, step_size, gamma)
+        elif self.hparams.lr_scheduler == "ReduceLROnPlateau":
+            factor = self.hparams.lr_factor
+            patience = self.hparams.lr_patience
+            scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(
+                optimizer, mode="min", factor=factor, patience=patience
+            )
+            return {
+                "optimizer": optimizer,
+                "lr_scheduler": scheduler,
+                "monitor": self.hparams.LRScheduler_metric,
+            }
+        elif self.hparams.lr_scheduler == "OneCycleLR":
+            max_lr = self.hparams.OneCycleLR_max_lr
+            epochs = self.hparams.OneCycleLR_epochs
+            steps_per_epoch = self.hparams.train_trajectories * (
+                    self.hparams.T + 1
+            )
+            scheduler = torch.optim.lr_scheduler.OneCycleLR(
+                optimizer,
+                max_lr=max_lr,
+                epochs=epochs,
+                steps_per_epoch=steps_per_epoch,
+            )
+        else:
+            print("Invalid scheduler configuration.  See --help")
+            raise
+        return [optimizer], [scheduler]
+
+    def training_step(self, batch, batch_idx):
+        data, label = batch
+        loss = self.model.loss(*data)
+        return loss
+
+    def validation_step(self, batch, batch_idx):
+        data, label = batch
[…]
