+++ cca_zoo/deepmodels/__init__.py
@@ -4,7 +4,8 @@ from ._dcca_base import _DCCA_base
 from .dcca import DCCA
 from .dcca_noi import DCCA_NOI
 from .dccae import DCCAE
-from .deepwrapper import DeepWrapper
 from .dtcca import DTCCA
 from .dvcca import DVCCA
 from .splitae import SplitAE
+from .trainers import CCALightning
+from .utils import get_dataloaders, process_data
