+++ cca_zoo/models/__init__.py
@@ -1,5 +1,4 @@
 from .cca_base import _CCA_Base
-from .rcca import rCCA, CCA, PLS
 from .gcca import GCCA, KGCCA
 from .iterative import (
     _Iterative,
@@ -14,4 +13,6 @@ from .iterative import (
     SWCCA,
 )
 from .mcca import MCCA, KCCA
+from .ncca import NCCA
+from .rcca import rCCA, CCA, PLS
 from .tcca import TCCA, KTCCA
