### 2021-11-03 #3823 @hugorichard
@emdupre, here is my review:

This is a really nice piece of software. I believe it is really useful to have all these CCA methods gathered at the same place.

I recommend acceptance.

As a side note, I believe the vision of mvlearn and cca-zoo coincide. It would be interesting to study how to fuse the two in my opinion.

A few comments:
- The statement of need is missing in the documentation
- I don't like to have examples as notebooks. I believe it is much better to do it like in sklearn (with .py files that generate figures along with some text that describes why the experiment is performed and why the model is useful).
- I believe the tests are not rigorous enough. In many tests, nothing is actually checked (there is no assert statement). In other tests, it is only checked that the performance of some methods is lower than others. It would be more convincing to test that the method actually works: take some synthetic data that match the CCA model and show that the method can learn projection so that the correlation of the projected data is above 0.9.
- I get an ImportError: cannot import name 'delayed' from 'sklearn.utils.fixes'
after running pip install cca-zoo. It is fixed if I force the version of scikit-learn to be 0.24 so something needs to be fixed there.
- The tests are extremely long to run... much longer than sklearn although the library is not as huge. This should definitely be fixed.
- Some tests are failing on my end when I run `py.test test_models.py` with the error ValueError: Unknown projection '3d'
