### 2021-11-05 a055b04
Switching wrapper to pytorch-lightning
