+++ cca_zoo/test/test_deepmodels.py
@@ -15,8 +15,8 @@ from cca_zoo.models import CCA
 manual_seed(0)
 rng = check_random_state(0)
 X = rng.rand(200, 10)
-Y = rng.rand(200, 10)
-Z = rng.rand(200, 10)
+Y = rng.rand(200, 12)
+Z = rng.rand(200, 14)
 X_conv = rng.rand(100, 1, 16, 16)
 Y_conv = rng.rand(100, 1, 16, 16)
 dataset = data.CCA_Dataset([X, Y, Z])
@@ -36,7 +36,7 @@ conv_loader = get_dataloaders(conv_dataset)
 def test_input_types():
     latent_dims = 2
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     # DCCA
     dcca_model = DCCA(
         latent_dims=latent_dims,
@@ -103,7 +103,7 @@ def test_DCCA_methods():
     cca_model = CCA(latent_dims=latent_dims).fit((X, Y))
     # DCCA_NOI
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dcca_noi_model = DCCA_NOI(
         latent_dims, X.shape[0], encoders=[encoder_1, encoder_2], rho=0
     )
@@ -119,7 +119,7 @@ def test_DCCA_methods():
     )
     # DCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -137,7 +137,7 @@ def test_DCCA_methods():
     )
     # DGCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dgcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -155,7 +155,7 @@ def test_DCCA_methods():
     )
     # DMCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dmcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -177,7 +177,7 @@ def test_DTCCA_methods():
     latent_dims = 2
     epochs = 5
     encoder_1 = architectures.Encoder(latent_dims=10, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=10, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=10, feature_size=12)
     dtcca_model = DTCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
[…]
