### 2021-11-05 52ff451
Switching wrapper to pytorch-lightning
