### 2021-11-07 6a9bd05
Update tests to test different dimensions in deep training
