+++ cca_zoo/test/test_deepmodels.py
@@ -12,8 +12,8 @@ from cca_zoo.models import CCA
 manual_seed(0)
 rng = check_random_state(0)
 X = rng.rand(200, 10)
-Y = rng.rand(200, 10)
-Z = rng.rand(200, 10)
+Y = rng.rand(200, 12)
+Z = rng.rand(200, 14)
 X_conv = rng.rand(100, 1, 16, 16)
 Y_conv = rng.rand(100, 1, 16, 16)
 train_dataset = data.CCA_Dataset([X, Y])
@@ -23,7 +23,7 @@ def test_input_types():
     latent_dims = 2
     device = "cpu"
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     # DCCA
     dcca_model = DCCA(
         latent_dims=latent_dims,
@@ -89,7 +89,7 @@ def test_DCCA_methods_cpu():
     epochs = 100
     # DCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -106,7 +106,7 @@ def test_DCCA_methods_cpu():
     )
     # DGCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dgcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -123,7 +123,7 @@ def test_DCCA_methods_cpu():
     )
     # DMCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dmcca_model = DCCA(
         latent_dims=latent_dims,
         encoders=[encoder_1, encoder_2],
@@ -140,7 +140,7 @@ def test_DCCA_methods_cpu():
     )
     # DCCA_NOI
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dcca_noi_model = DCCA_NOI(
         latent_dims, X.shape[0], encoders=[encoder_1, encoder_2], rho=0
     )
@@ -159,8 +159,8 @@ def test_DTCCA_methods_cpu():
     latent_dims = 2
     device = "cpu"
     encoder_1 = architectures.Encoder(latent_dims=10, feature_size=10)
-    encoder_2 = architectures.Encoder(latent_dims=10, feature_size=10)
-    encoder_3 = architectures.Encoder(latent_dims=10, feature_size=10)
+    encoder_2 = architectures.Encoder(latent_dims=10, feature_size=12)
+    encoder_3 = architectures.Encoder(latent_dims=10, feature_size=14)
[…]
