### 2021-11-09 #3823 @robbisg
Hi everyone,

@emdupre I will not open an issue in the `cca-zoo` repo since I think that on my side there are few requests that can be addressed in this thread. But if you prefer, I can open the issue! ;)

### Review
I think `cca-zoo` is a very useful package that can be used to relate multivariate datasets. Personally, I used very basic implementations of these technique and I am very happy for the possibility of using more advance technique included in this tool!

#### Paper
The paper is well written and presents an extensive comparison of other similar packages and the implementation philosophy.
These edits should be included, in my point of view:
- Although you are presenting a software tool, I believe that a brief mathematical formulation of the problem, as presented in the tutorial notebook, can help a partially-expert reader to understand the application of the tool
- As before, a very simple example of the application of these technique can help a non-expert reader to understand the package.

### Software
The software uses the `scikit-learn` standard and this is very useful for the users, moreover CI and tests are valuable efforts to make the package more robust.
I hadn't any test or import errors, and by inspecting github CI, it seems that everything is ok, also using different python versions.

- I agree with @hugorichard, that examples can be "prettified" as in `scikit-learn` or `mvlearn` and is quite easy using `sphinx-gallery`.
- I think that you can also add in the documentation the mathematical formulation of these algorithms.

I think that the tool is in a very good shape, I recommend acceptance.

Thanks @jameschapman19 and @emdupre, and sorry for the delay.
