### 2021-11-09 #3823 @jameschapman19
I was wary of replying as I don't know the process but it feels appropriate for me to say thanks both. All of these comments I agree with and will action them.

- examples in .py files (another benefit I have realized of this is that they can be tested cleanly on commit)
- smaller dimensionality tests should speed things up and improve the nature of the tests (and also make hard tests on results less unwieldy i.e. I can put correct array results in the script if they aren't high dimensional). I notice this is essentially how the scikit-learn ones work. I also agree that a quick-and-easy approach is just to check correlations are reasonably high (and perhaps that regularization makes them lower). I took this approach a little with the deep tests (i.e. DCCA higher correlation than linear CCA).
- maths in the docs is a great idea.
