### 2021-11-09 jameschapman19/cca_zoo#75 @jameschapman19
Primary Examples/Tutorials in .py files
I think the best route to go is a minimal colab notebook and some sphinx-gallery examples in .py files.
