### 2021-11-09 jameschapman19/cca_zoo#76 @jameschapman19
Update tests
Reduce the dimensionality of the tests substantially. Add in some hard checks such as correlation thresholds where they are missing.
