+++ ccagame/__init__.py
@@ -1,4 +1,3 @@
 from . import cca
-from . import pca
 from . import pls
 from . import solver
