+++ ccagame/pls/__init__.py
@@ -3,5 +3,5 @@ from .batch import Batch
 from .game import Game
 from .incremental import Incremental
 from .msg import MSG
-from .numpy import calc_numpy, Numpy
+from .numpy import Numpy
 from .sgd import SGD
