+++ ccagame/cca/__init__.py
@@ -4,4 +4,4 @@ from .game import Game
 from .genoja import Genoja
 from .lagrange import Lagrange
 from .msg import MSG
-from .numpy import calc_numpy, Numpy
+from .numpy import Numpy
