+++ ccagame/jaxline/experiment.py
@@ -1,0 +1,188 @@
+import functools
+from typing import Optional
+
+import jax
+import jax.numpy as jnp
+import numpy as np
+import optax
+from absl import app
+from absl import flags
+from jax._src.random import PRNGKey
+from jaxline import platform
+from jaxline import utils
+from jaxline.base_config import get_base_config
+from jaxline.experiment import AbstractExperiment
+from os import environ
+from ccagame.utils import data_stream
+from datasets.mnist import mnist
+
+CORES = 2
+environ['XLA_FLAGS'] = f'--xla_force_host_platform_device_count={CORES}'
+input_data_iterator = None
+FLAGS = flags.FLAGS
+X, _, X_te, _ = mnist()
+input_data_iterator = data_stream(X, Y=None, batch_size=None)
+
+
+def get_config():
+    """Return config object for training."""
+    config = get_base_config()
+    config.experiment_kwargs = {
+        'k_per_device': 3,
+        'num_devices': CORES,
+        'dims': X.shape[1],
+        'axis_index_groups': None,
+        'data_stream': input_data_iterator,
+        'algorithm': 'game',
+    }
+    config.checkpoint_dir = 'C:/Users/chapm/PycharmProjects/ccagame/jaxlog'
+    config.train_checkpoint_all_hosts = True
+    config.lock()
+    return config
+
+
+def eg_grads(vi: jnp.ndarray,
+             weights: jnp.ndarray,
+             eigs: jnp.ndarray,
+             data: jnp.ndarray) -> jnp.ndarray:
+    """
+    Args:
+    vi: shape (d,), eigenvector to be updated
+    weights: shape (k,), mask for penalty coefficients,
+    eigs: shape (k, d), i.e., vectors on rows
+    data: shape (N, d), minibatch X_t
+    Returns:
+    grads: shape (d,), gradient for vi
+    """
+    weights_ij = (jnp.sign(weights + 0.5) - 1.) / 2.  # maps -1 to -1 else to 0
+    data_vi = data @ vi
+    data_eigs = (data @ eigs.T).T  # Xvj on row j
+    vi_m_vj = data_eigs @ data_vi
+    penalty_grads = vi_m_vj * eigs.T
+    penalty_grads = penalty_grads @ weights_ij
+    grads = data.T @ data_vi + penalty_grads
+    return grads
+
+
+def utility(vi, weights, eigs, data):
+    """Compute Eigengame utilities.
+    util: shape (1,), utility for vi
+    """
+    data_vi = data @ vi
+    data_eigs = (data @ eigs.T).T  # Xvj on row j
+    vi_m_vj2 = data_eigs @ data_vi ** 2.
+    vj_m_vj = jnp.sum(data_eigs * data_eigs, axis=1)
+    r_ij = vi_m_vj2 / vj_m_vj
+    util = r_ij @ weights
+    return util
+
+
+class Experiment(AbstractExperiment):
+    def __init__(self, mode, init_rng=None, num_devices=1, k_per_device=1, dims=1, axis_index_groups=None,
+                 algorithm='game', data_stream=None):
+        super(Experiment, self).__init__(mode=mode, init_rng=init_rng)
+        """Constructs the experiment.
+        Args:
+          mode: A string, equivalent to FLAGS.jaxline_mode when running normally.
+          init_rng: A `PRNGKey` to use for experiment initialization.
+        """
+        """Initialization function for a Jaxline experiment."""
+        self._total_k = k_per_device * num_devices
+        self._dims = dims
+        self.algorithm = algorithm
+        self.data_stream = data_stream
+        weights = np.eye(self._total_k) * 2 - np.ones((self._total_k, self._total_k))
+        weights[np.triu_indices(self._total_k, 1)] = 0
[…]
