### 2021-11-09 #3823 @emdupre
Thanks both @hugorichard and @robbisg for your reviews, and @jameschapman19 for your thoughtful response !

Please do feel free to continue discussions in thread, though as specific issues on `cca-zoo` arise from this review process it will be helpful to cross-link them here !
