+++ ccagame/pls/meg.py
@@ -1,2 +1,0 @@
-# TODO
-# http://proceedings.mlr.press/v48/aroraa16.pdf
