+++ ccagame/cca/streamingshiftinvert.py
@@ -1,2 +1,0 @@
-# TODO
-# https://jmlr.csail.mit.edu/papers/volume20/18-095/18-095.pdf
