+++ ccagame/pls/game.py
@@ -1,231 +1,0 @@
-# Importing necessary libraries
-import time
-from functools import partial
-
-import jax.numpy as jnp
-from jax import grad, jit
-
-from ccagame.utils import data_stream, get_num_batches
-from . import _PLS
-
-
-@partial(jit, static_argnums=(6))
-def alpha_model(u, v, X, Y, U, V, k: int):
-    """
-    Returns the utilities for the kth players
-
-    Parameters
-    ----------
-    u :
-        current estimate for this level's left eigenvector
-    v :
-        current estimate for this level's right eigenvector
-    X :
-        batch of data for view X
-    Y :
-        batch of data for view Y
-    U :
-        all eigenvector estimates for each level
-    V :
-        all eigenvector estimates for each level
-    k :
-        level
-
-    Returns
-    -------
-
-    """
-    C_xy = X.T @ Y
-    rewards = u.T @ C_xy @ v
-    penalties = (u.T @ C_xy @ V[:, :k]) ** 2 / jnp.diag(U[:, :k].T @ C_xy @ V[:, :k])
-    return jnp.sum(rewards - penalties.sum())
-
-
-@partial(jit, static_argnums=(6))
-def mu_model(u, v, X, Y, U, V, k: int):
-    """
-    Returns the gradients for the kth players
-
-    Parameters
-    ----------
-    u :
-        current estimate for this level's left eigenvector
-    v :
-        current estimate for this level's right eigenvector
-    X :
-        batch of data for view X
-    Y :
-        batch of data for view Y
-    U :
-        all eigenvector estimates for each level
-    V :
-        all eigenvector estimates for each level
-    k :
-        level
-
-    Returns
-    -------
-
-    """
-    C_xy = X.T @ Y
-    rewards = C_xy @ v
-    penalties = (u.T @ C_xy @ V[:, :k]) * U[:, :k]
-    return rewards - penalties.sum(axis=1)
-
-
-# Update rule to be used for calculating eigenvectors
-@partial(jit, static_argnums=6, static_argnames=("lr", "riemannian_projection", "mu"))
-def update(
-    u,
-    v,
-    X,
-    Y,
-    U,
-    V,
-    k: int,
-    lr: float = 1,
-    riemannian_projection: bool = False,
-    mu=False,
-):
-    """
-    Update the left and right singular vector estimates
-
-    Parameters
-    ----------
-    u :
-        current estimate for this level's left eigenvector
-    v :
-        current estimate for this level's right eigenvector
-    X :
-        batch of data for view X
-    Y :
-        batch of data for view Y
-    U :
-        all eigenvector estimates for each level
-    V :
-        all eigenvector estimates for each level
-    k :
-        level
-    lr :
-        learning rate
-    riemannian_projection :
-        whether to use riemannian projection
-    mu :
-        which game model to use. If True uses unbiased estimate as in eigengame:unloaded if False uses biased estimate as in original eigengame
-    """
-    if mu:
-        du = mu_model(u, v, X, Y, U, V, k)
-        dv = mu_model(v, u, Y, X, V, U, k)
-    else:
-        du = grad(alpha_model)(u, v, X, Y, U, V, k)
-        dv = grad(alpha_model)(v, u, Y, X, V, U, k)
-    if riemannian_projection:
-        dur = du - (u.T @ u) * u
-        uhat = u + lr * dur
-        dvr = dv - (v.T @ v) * v
-        vhat = v + lr * dvr
-    else:
-        uhat = u + lr * du
-        vhat = v + lr * dv
-    return uhat / jnp.linalg.norm(uhat), vhat / jnp.linalg.norm(vhat)
-
-
-# Object form
-class Game(_PLS):
-    def __init__(
-        self,
[…]
