+++ ccagame/cca/__init__.py
@@ -1,7 +1,0 @@
-from ._cca import _CCA
-from .ccalin import CCALin
-from .game import Game
-from .genoja import Genoja
-from .lagrange import Lagrange
-from .msg import MSG
-from .numpy import Numpy
