+++ ccagame/pls/__init__.py
@@ -1,7 +1,0 @@
-from ._pls import _PLS
-from .batch import Batch
-from .game import Game
-from .incremental import Incremental
-from .msg import MSG
-from .numpy import Numpy
-from .sgd import SGD

[…]
