+++ ccagame/pls/gameexperiment.py
@@ -1,0 +1,195 @@
+import functools
+from os import environ
+from typing import Optional
+
+import jax
+import jax.numpy as jnp
+import optax
+from absl import app, flags
+from ccagame.pls.plsexperiment import PLSExperiment
+from ccagame.utils import data_stream
+from datasets.mnist import mnist
+from jax._src.random import PRNGKey
+from jaxline import platform, utils
+from jaxline.base_config import get_base_config
+from jaxline.experiment import AbstractExperiment
+
+import numpy as np
+
+CORES = 2
+environ["XLA_FLAGS"] = f"--xla_force_host_platform_device_count={CORES}"
+input_data_iterator = None
+FLAGS = flags.FLAGS
+X, _, X_te, _ = mnist()
+input_data_iterator = data_stream(X, Y=None, batch_size=None)
+
+
+def get_config():
+    """Return config object for training."""
+    config = get_base_config()
+    config.experiment_kwargs = {
+        "k_per_device": 3,
+        "num_devices": CORES,
+        "dims": X.shape[1],
+        "axis_index_groups": None,
+        "data_stream": input_data_iterator,
+        "algorithm": "game",
+    }
+    config.checkpoint_dir = "C:/Users/chapm/PycharmProjects/ccagame/jaxlog"
+    config.train_checkpoint_all_hosts = True
+    config.lock()
+    return config
+
+
+def eg_grads(
+    vi: jnp.ndarray, weights: jnp.ndarray, eigs: jnp.ndarray, data: jnp.ndarray
+) -> jnp.ndarray:
+    """
+    Args:
+    vi: shape (d,), eigenvector to be updated
+    weights: shape (k,), mask for penalty coefficients,
+    eigs: shape (k, d), i.e., vectors on rows
+    data: shape (N, d), minibatch X_t
+    Returns:
+    grads: shape (d,), gradient for vi
+    """
+    weights_ij = (jnp.sign(weights + 0.5) - 1.0) / 2.0  # maps -1 to -1 else to 0
+    data_vi = data @ vi
+    data_eigs = (data @ eigs.T).T  # Xvj on row j
+    vi_m_vj = data_eigs @ data_vi
+    penalty_grads = vi_m_vj * eigs.T
+    penalty_grads = penalty_grads @ weights_ij
+    grads = data.T @ data_vi + penalty_grads
+    return grads
+
+
+def utility(vi, weights, eigs, data):
+    """Compute Eigengame utilities.
+    util: shape (1,), utility for vi
+    """
+    data_vi = data @ vi
+    data_eigs = (data @ eigs.T).T  # Xvj on row j
+    vi_m_vj2 = data_eigs @ data_vi ** 2.0
+    vj_m_vj = jnp.sum(data_eigs * data_eigs, axis=1)
+    r_ij = vi_m_vj2 / vj_m_vj
+    util = r_ij @ weights
+    return util
+
+
+def _grads_and_utils(vi, weights, V, inputs):
+    """Compute utiltiies and update directions ("grads").
+    23 Wrap in jax.vmap for k_per_device dimension."""
+    utilities = utility(vi, weights, V, inputs)
+    grads = eg_grads(vi, weights, V, inputs)
+    return grads, utilities
+
+
+class GameExperiment(PLSExperiment):
+    def __init__(
+        self,
+        mode,
+        init_rng=None,
+        num_devices=1,
+        k_per_device=1,
+        dims=1,
+        axis_index_groups=None,
+        data_stream=None,
+    ):
+        super(PLSExperiment, self).__init__(
+            mode,
+            init_rng=init_rng,
+            num_devices=num_devices,
+            k_per_device=k_per_device,
+            dims=dims,
+            axis_index_groups=axis_index_groups,
[…]
