+++ ccagame/__init__.py
@@ -1,0 +1,2 @@
+from . import pca
+from . import pls

[…]
