+++ ccagame/pca/experiments/__init__.py
@@ -1,0 +1,3 @@
+from .pcaexperiment import PCAExperiment
+from .game import Game
+from .oja import Oja
