+++ ccagame/pls/experiments/__init__.py
@@ -1,0 +1,2 @@
+from .plsexperiment import PLSExperiment
+from .game import Game

[…]
