+++ README.md
@@ -1,4 +1,18 @@
 # ccagame
 
+## How to run an experiment
+At the moment to run PLS methods:
+
+ - Eigengame: ccagame/pls/experiments/game.py
+ - Oja: ccagame/pls/experiments/oja.py
+
+To run PCA methods:
+
+ - Eigengame: ccagame/pca/experiments/game.py
+ - Oja: ccagame/pls/experiments/oja.py
+ - GHA: ccagame/pls/experiments/gha.py
+
+## Description
+
 cca as a game a bit like https://deepmind.com/research/publications/2020/EigenGame-PCA-as-a-Nash-Equilibrium
 / https://arxiv.org/pdf/2010.00554.pdf
