+++ jaxline/utils_test.py
@@ -1,0 +1,326 @@
+# Copyright 2020 DeepMind Technologies Limited. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""Tests for jaxline's utils."""
+
+import functools
+import itertools as it
+import time
+from unittest import mock
+
+from absl.testing import absltest
+from absl.testing import flagsaver
+import jax
+import jax.numpy as jnp
+from jaxline import utils
+import numpy as np
+
+
+class PyPrefetchTest(absltest.TestCase):
+
+  def testEmpty(self):
+    self.assertEqual(list(utils.py_prefetch(lambda: ())), [])
+
+  def testBaseCase(self):
+    self.assertEqual(list(utils.py_prefetch(lambda: range(100))),
+                     list(range(100)))
+
+  def testBadFunction(self):
+
+    def _bad_function():
+      raise ValueError
+
+    iterable = utils.py_prefetch(_bad_function)
+    with self.assertRaises(ValueError):
+      next(iterable)
+
+  def testBadFunctionIteration(self):
+
+    def _bad_iterable():
+      yield 1
+      raise ValueError
+
+    iterable = utils.py_prefetch(_bad_iterable)
+    self.assertEqual(next(iterable), 1)
+    with self.assertRaises(ValueError):
+      next(iterable)
+
+
+class TreePsumTest(absltest.TestCase):
+
+  def testBaseCase(self):
+    # pick leaf objects with leading dimension one as these tests will
+    # be run on a single device.
+    data = {"a": jnp.array([1]), "b": jnp.array([2])}
+    data_summed = jax.pmap(
+        lambda x: utils.tree_psum(x, axis_name="i"), axis_name="i")(data)
+    self.assertEqual(data_summed, data)
+
+  def testEmpty(self):
+    data = {"a": jnp.array([]), "b": jnp.array([])}
+    with self.assertRaises(ZeroDivisionError):
+      jax.pmap(lambda x: utils.tree_psum(x, axis_name="i"), axis_name="i")(data)
+
+  def testSingleLeafTree(self):
+    data = jnp.array([1])
+    data_summed = jax.pmap(
+        lambda x: utils.tree_psum(x, axis_name="i"), axis_name="i")(data)
+    self.assertEqual(data_summed, data)
+
+  def testNotNumpy(self):
+    data = [1]
+    with self.assertRaises(ValueError):
+      jax.pmap(lambda x: utils.tree_psum(x, axis_name="i"), axis_name="i")(data)
+
+  def testNumDevicesMismatch(self):
+    data = jnp.array([1, 2])  # assumes 2 devices but we only have 1
+    with self.assertRaises(ValueError):
+      jax.pmap(lambda x: utils.tree_psum(x, axis_name="i"), axis_name="i")(data)
+
+  def testNoPmapWrapper(self):
+    with self.assertRaises(NameError):  # axis_name will be undefined
+      utils.tree_psum(jnp.array([1]), axis_name="i")
+
+  def testAxisNameMismatch(self):
+    data = jnp.array([1])
+    with self.assertRaises(NameError):
[…]
