+++ jaxline/platform.py
@@ -1,0 +1,152 @@
+# Copyright 2020 DeepMind Technologies Limited. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+# ==============================================================================
+"""A Deepmind-specific platform for running Experiments with Jaxline."""
+
+from concurrent import futures
+import os
+from typing import Any, Mapping
+
+from absl import flags
+from absl import logging
+
+import chex
+import jax
+
+from jaxline import base_config
+from jaxline import train
+from jaxline import utils
+from ml_collections import config_dict
+from ml_collections import config_flags
+import numpy as np
+
+import tensorflow as tf
+
+FLAGS = flags.FLAGS
+
+
+# TODO(tomhennigan) Add support for ipdb and pudb.
+config_flags.DEFINE_config_file(
+    "config", help_string="Training configuration file.")
+# This flag is expected to be used only internally by jaxline.
+# It is prefixed by "jaxline" to prevent a conflict with a "mode" flag defined
+# by Monarch.
+_JAXLINE_MODE = flags.DEFINE_string(
+    "jaxline_mode", "train",
+    ("Execution mode: `train` will run training, `eval` will run evaluation."))
+_JAXLINE_TPU_DRIVER = flags.DEFINE_string("jaxline_tpu_driver", "",
+                                          "Whether to use tpu_driver.")
+_JAXLINE_ENSURE_TPU = flags.DEFINE_bool(
+    "jaxline_ensure_tpu", False, "Whether to ensure we have a TPU connected.")
+
+
+def create_checkpointer(
+    config: config_dict.ConfigDict,
+    mode: str,
+) -> utils.Checkpointer:
+  """Creates an object to be used as a checkpointer."""
+  return utils.InMemoryCheckpointer(config, mode)
+
+
+class TensorBoardLogger:
+  """Writer to write experiment data to stdout."""
+
+  def __init__(self, config, mode: str):
+    """Initializes the writer."""
+    log_dir = os.path.join(config.checkpoint_dir, mode)
+    self._writer = tf.summary.create_file_writer(log_dir)
+
+  def write_scalars(self, global_step: int, scalars: Mapping[str, Any]):
+    """Writes scalars to stdout."""
+    global_step = int(global_step)
+    with self._writer.as_default():
+      for k, v in scalars.items():
+        tf.summary.scalar(k, v, step=global_step)
+    self._writer.flush()
+
+  def write_images(self, global_step: int, images: Mapping[str, np.ndarray]):
+    """Writes images to writers that support it."""
+    global_step = int(global_step)
+    with self._writer.as_default():
+      for k, v in images.items():
+        # Tensorboard only accepts [B, H, W, C] but we support [H, W] also.
+        if v.ndim == 2:
+          v = v[None, ..., None]
+        tf.summary.image(k, v, step=global_step)
+    self._writer.flush()
+
+
+def create_writer(config: config_dict.ConfigDict, mode: str) -> Any:
+  """Creates an object to be used as a writer."""
+  return TensorBoardLogger(config, mode)
+
+
+@utils.debugger_fallback
+def main(experiment_class, argv, checkpointer_factory=create_checkpointer):
[…]
