### 2021-11-11 jameschapman19/cca_zoo#77 @ejolly
[JOSS Review] - CCA_Tutorial notebook produces errors on fresh install
Including this here as part of my [JOSS review](https://github.com/openjournals/joss-reviews/issues/3823):

Looks like the first notebook produces errors in the following sections. These can also been seen in the notebook cell outputs if opening in Google Collab. Building tutorial notebook as part of your CI might be a helpful way to catch these and other issues in the future:

  -  The "Overfitting and Sample Size" section produces an error:  `rCCA object has no attribute scores`
  - The "Ridge Regularised CCA: from CCA to PLS produces an error: `Expected 2D array, got 1D array instead`
  - The "Regularised CCA" under the neuroimaging example also produces a `rCCA object has no attribute scores`
