### 2021-11-11 #3823 @ejolly
Hey @emdupre @jameschapman19 here's my review:

## Review

- Super cool package and it's nice to see to all these algorithms (including their deep learning variants) in one place!
- I found the manuscript clear and adhering to JOSS requirements
- The API reference is also great and nicely documented
- Recommend acceptance

## Suggested Changes
- I do agree with the other reviewers that it would be nice to move some of the math + high-level explanations from the tutorial notebooks into a dedicated page in the documentation site. This would help non-experts get acquainted with the methods and more clearly see how they could use `cca-zoo` for their own problems
- I'd also recommend using `sphinx-gallery` to combine your tutorials with your documentation. You might consider adding another step in your CI pipeline to build and deploy your documentation after your test suite finishes. This has a few advantages:
   - It ensures that your live docs automatically stay updated with any new commits
   - It executes the code in the tutorial notebooks which serve as a secondary layer of testing that your test-suite might not cover

### Minor suggested changes

 - Not strictly necessary but it might be nice to have a `pip install cca-zoo[all]` option
 - Since the intended audience seem to be folks who are familiar with scikit-learn, you might consider adopting the American-English spelling of "center" as oppose to "centre" for any method and function argument as they do in their API and documentation for consistency and user-experience.

## Test and Issues
- I had no issues running the full testing suite in a fresh environment on macOS with Python 3.8.12
- There were a few issues with one of the tutorial notebooks which I noted [here](https://github.com/jameschapman19/cca_zoo/issues/77)
