+++ ccagame/pls/experiments/game.py
@@ -25,6 +25,7 @@ class Game(PLSExperiment):
         k_per_device=1,
         dims=1,
         data_stream=None,
+        whole_batch=False
     ):
         super(Game, self).__init__(
             mode,
@@ -33,6 +34,7 @@ class Game(PLSExperiment):
             k_per_device=k_per_device,
             dims=dims,
             data_stream=data_stream,
+            whole_batch=whole_batch
         )
         """Constructs the experiment.
         Args:
