+++ ccagame/baseexperiment.py
@@ -17,6 +17,7 @@ def get_config(
     dims,
     log_tensors_interval=1,
     log_train_data_interval=1,
+    whole_batch=False
 ):
     """Return config object for training."""
     config = get_base_config()
@@ -25,6 +26,7 @@ def get_config(
         "num_devices": devices,
         "dims": dims,
         "data_stream": data_stream,
+        "whole_batch": whole_batch
     }
     config.checkpoint_dir = "jaxlog"
     config.train_checkpoint_all_hosts = True
@@ -42,6 +44,7 @@ class BaseExperiment(AbstractExperiment):
         num_devices=1,
         k_per_device=1,
         data_stream=None,
+        whole_batch=False
     ):
         super(BaseExperiment, self).__init__(mode=mode, init_rng=init_rng)
         """Constructs the experiment.
@@ -54,6 +57,9 @@ class BaseExperiment(AbstractExperiment):
         self.data_stream = data_stream
         self.local_rng = jax.random.fold_in(PRNGKey(123), jax.host_id())
         self._num_devices = num_devices
+        self.whole_batch = whole_batch
+        if self.whole_batch:
+            self.inputs=None
 
     def step(
         self,
@@ -63,8 +69,11 @@ class BaseExperiment(AbstractExperiment):
         writer: Optional[utils.Writer],
     ):
         """Step function for a Jaxline experiment"""
-        inputs = next(self.data_stream)
-        outputs = self._update(inputs, global_step)
+        if self.whole_batch:
+            outputs = self._update(self.inputs, global_step)
+        else:
+            inputs = next(self.data_stream)
+            outputs = self._update(inputs, global_step)
         return self._get_scalars(outputs)
 
     def _get_scalars(self, outputs):
