+++ ccagame/pca/experiments/game.py
@@ -154,8 +154,7 @@ class Game(PCAExperiment):
 if __name__ == "__main__":
     X, _, X_te, _ = mnist()
     input_data_iterator = data_stream(X, Y=None, batch_size=None)
-    k_per_device = 15
-    whole_batch = True
-    FLAGS.config = get_config(input_data_iterator, CORES, k_per_device, X.shape[1], whole_batch=whole_batch)
+    k_per_device = 5
+    FLAGS.config = get_config(input_data_iterator, CORES, X.shape[1],k_per_device=k_per_device)
     flags.mark_flag_as_required("config")
     app.run(functools.partial(platform.main, Game))
