+++ cca_zoo/models/__init__.py
@@ -1,4 +1,5 @@
 from .cca_base import _CCA_Base
+from .rcca import rCCA, CCA, PLS
 from .gcca import GCCA, KGCCA
 from .iterative import (
     _Iterative,
@@ -14,5 +15,4 @@ from .iterative import (
 )
 from .mcca import MCCA, KCCA
 from .ncca import NCCA
-from .rcca import rCCA, CCA, PLS
 from .tcca import TCCA, KTCCA
