+++ ccagame/pls/experiments/game.py
@@ -54,26 +54,26 @@ class Game(PLSExperiment):
         self._U = jax.pmap(lambda U: U / jnp.linalg.norm(U, axis=1, keepdims=True))(U)
         self._V = jax.pmap(lambda V: V / jnp.linalg.norm(V, axis=1, keepdims=True))(V)
         # This line parallelizes over data sending different data to each device
-        self._grads_and_update_x = functools.partial(self._grads_and_update, view="x")
-        self._grads_and_update_y = functools.partial(self._grads_and_update, view="y")
-        self._grads_and_update_x = jax.pmap(
-            self._grads_and_update_x,
-            in_axes=(0, 0, 0, None, None, 0, 0, 0),
-            axis_name="i",
-        )
-        self._grads_and_update_y = jax.pmap(
-            self._grads_and_update_y,
-            in_axes=(0, 0, 0, None, None, 0, 0, 0),
+        self._update_with_grads = jax.pmap(
+            jax.vmap(
+            self._update_with_grads,
+            in_axes=(0, 0, 0),
+            ),
+            in_axes=(0, 0, 0),
             axis_name="i",
         )
         # This parallelizes gradient calcs and updates for eigenvectors within a given device
-        self._grads_and_utils = jax.vmap(
-            self._grads_and_utils, in_axes=(0, 0, 0, None, None, None, None)
+        self._grads = jax.pmap(
+            jax.vmap(
+            self._grads, 
+            in_axes=(0, 0, 0, None,None,None,None),
+            ),
+            in_axes=(0, 0, 0, None,None,0,0),
+            axis_name="i"
         )
-        self._optimizer_x = optax.sgd(learning_rate=1e-5, momentum=0.9, nesterov=True)
-        self._opt_state_x = jax.pmap(lambda U: self._optimizer_x.init(U))(self._U)
-        self._optimizer_y = optax.sgd(learning_rate=1e-5, momentum=0.9, nesterov=True)
-        self._opt_state_y = jax.pmap(lambda V: self._optimizer_y.init(V))(self._V)
+        self._optimizer = optax.sgd(learning_rate=1e-6, momentum=0.9, nesterov=True)
+        self._opt_state_x = jax.pmap(lambda U: self._optimizer.init(U))(self._U)
+        self._opt_state_y = jax.pmap(lambda V: self._optimizer.init(V))(self._V)
 
     def _update(self, views, global_step):
         X_i, Y_i = views
@@ -81,67 +81,34 @@ class Game(PLSExperiment):
         Y_i = jnp.reshape(Y_i, (self._num_devices, -1, self._dims[1]))
         self._local_U = jnp.reshape(self._U, (self._total_k, self._dims[0]))
         self._local_V = jnp.reshape(self._V, (self._total_k, self._dims[1]))
-        self._U, self._opt_state_x = self._grads_and_update_x(
-            self._U,
-            self._V,
-            self._weights,
-            self._local_U,
-            self._local_V,
-            X_i,
-            Y_i,
-            self._opt_state_x,
-        )
-        self._V, self._opt_state_y = self._grads_and_update_y(
-            self._V,
-            self._U,
-            self._weights,
-            self._local_V,
-            self._local_U,
-            Y_i,
-            X_i,
-            self._opt_state_y,
-        )
+        grads_x = self._grads(self._U, self._V, self._weights, self._local_U,self._local_V, X_i, Y_i)
+        grads_y = self._grads(self._V, self._U, self._weights, self._local_V, self._local_U, Y_i, X_i)
[…]
