### 2021-11-13 2221131
Switching tutorials to sphinx gallery
