+++ cca_zoo/test/test_deepmodels.py
@@ -63,8 +63,8 @@ def test_tutorial():
     test_ids = np.arange(min(N, len(test_dataset)))
     np.random.shuffle(test_ids)
     test_dataset = Subset(test_dataset, test_ids)
-    encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
-    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784)
+    encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, layer_sizes=[800, 800])
+    encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=784, layer_sizes=[800, 800])
     train_loader, val_loader = get_dataloaders(train_dataset, val_dataset)
     test_loader = torch.utils.data.DataLoader(test_dataset, batch_size=len(test_dataset))
     dcca_model = DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
