+++ cca_zoo/deepmodels/__init__.py
@@ -7,4 +7,5 @@ from .dccae import DCCAE
 from .dtcca import DTCCA
 from .dvcca import DVCCA
 from .splitae import SplitAE
+from .trainers import CCALightning
 from .utils import get_dataloaders, process_data
