### 2021-11-13 b64fbcb
Switching tutorials to sphinx gallery
