+++ docs/requirements.txt
@@ -14,3 +14,4 @@ torch>=1.9.0
 torchvision
 Pillow
 sphinx-autodoc-typehints
+sphinx-gallery
