### 2021-11-13 610d717
Switching tutorials to sphinx gallery
