### 2021-11-13 303a1bc
Switching tutorials to sphinx gallery
