### 2021-11-13 dee4766
Switching tutorials to sphinx gallery
