+++ docs/requirements.txt
@@ -13,5 +13,6 @@ joblib
 torch>=1.9.0
 torchvision
 Pillow
+sphinx
 sphinx-autodoc-typehints
 sphinx-gallery
