### 2021-11-13 4ccc829
Switching tutorials to sphinx gallery
