+++ docs/requirements.txt
@@ -1,18 +1,3 @@
-jax
-numpyro
-arviz
-numpy
-scikit-learn
-mvlearn
-scipy>=1.5
-matplotlib
-seaborn
-pandas
-tensorly
-joblib
-torch>=1.9.0
-torchvision
-Pillow
 sphinx
 sphinx-autodoc-typehints
 sphinx-gallery
