### 2021-11-13 fcef45b
Switching tutorials to sphinx gallery
