+++ docs/source/conf.py
@@ -14,31 +14,40 @@ import os
 import sys
 import warnings
 
-sys.path.insert(0, os.path.abspath('../..'))
+sys.path.insert(0, os.path.abspath("../.."))
 warnings.filterwarnings("ignore", category=UserWarning,
-                        message='Matplotlib is currently using agg, which is a'
-                                ' non-GUI backend, so cannot show the figure.')
+                        message="Matplotlib is currently using agg, which is a"
+                                " non-GUI backend, so cannot show the figure.")
 # -- Project information -----------------------------------------------------
 
-project = 'cca-zoo'
-copyright = '2021, James Chapman'
-author = 'James Chapman'
+project = "cca-zoo"
+copyright = "2021, James Chapman"
+author = "James Chapman"
 
 # -- General configuration ---------------------------------------------------
 
 extensions = [
-    'sphinx.ext.autodoc',
+    "sphinx.ext.autodoc",
     "sphinx.ext.autosummary",
-    'sphinx_autodoc_typehints',
-    'sphinx.ext.viewcode'
-    'sphinx.ext.viewcodesphinx_gallery'
-    'sphinx_gallery.gen_gallery',
-    'sphinx.ext.viewcodesphinx'
+    "sphinx_autodoc_typehints",
+    "sphinx.ext.intersphinx",
+    "sphinx.ext.viewcode"
+    "sphinx.ext.viewcodesphinx_gallery"
+    "sphinx_gallery.gen_gallery",
+    "sphinx.ext.viewcodesphinx"
 ]
 
 sphinx_gallery_conf = {
-    'examples_dirs': '../examples',  # path to your example scripts
-    'gallery_dirs': 'auto_examples',  # path to where to save gallery generated output
+    "examples_dirs": "../examples",  # path to your example scripts
+    "gallery_dirs": "auto_examples",  # path to where to save gallery generated output
+}
+
+# -- sphinx.ext.intersphinx
+intersphinx_mapping = {
+    "numpy": ("https://docs.scipy.org/doc/numpy", None),
+    "python": ("https://docs.python.org/3", None),
+    "scipy": ("https://docs.scipy.org/doc/scipy/reference", None),
+    "sklearn": ("http://scikit-learn.org/dev", None),
 }
 
 # -- sphinx.ext.autosummary
@@ -48,10 +57,10 @@ autosummary_generate = True
 autoclass_content = "both"
 autodoc_default_flags = ["members", "show-inheritance"]
 autodoc_member_order = "bysource"  # default is alphabetical
-special_members = '--init__'
+special_members = "--init__"
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
@@ -68,4 +77,4 @@ html_theme = "sphinx_rtd_theme"
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = ["_static"]
