### 2021-11-13 9a77ab6
Switching tutorials to sphinx gallery
