### 2021-11-13 f361ab0
Switching tutorials to sphinx gallery
