### 2021-11-13 c0bcf6f
Switching tutorials to sphinx gallery
