### 2021-11-13 b9a90e3
Switching tutorials to sphinx gallery
