### 2021-11-13 dbdf962
Switching tutorials to sphinx gallery
