### 2021-11-13 cd346e9
Switching tutorials to sphinx gallery
