### 2021-11-13 cffff4a
Switching tutorials to sphinx gallery
