+++ docs/conf.py
@@ -25,7 +25,6 @@ copyright = "2021, James Chapman"
 author = "James Chapman"
 
 # -- General configuration ---------------------------------------------------
-master_doc = 'index'
 
 extensions = [
     "sphinx.ext.autodoc",
@@ -77,3 +76,5 @@ html_theme = "sphinx_rtd_theme"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
+
+master_doc = 'index'
