### 2021-11-13 774b8ad
Switching tutorials to sphinx gallery
