### 2021-11-13 512d243
Switching tutorials to sphinx gallery
