+++ docs/source/conf.py
@@ -36,8 +36,8 @@ extensions = [
 
 sphinx_gallery_conf = {
     "doc_module": "cca-zoo",
-    "examples_dirs": "../examples",  # path to your example scripts
-    "gallery_dirs": "auto_examples",  # path to where to save gallery generated output
+    "examples_dirs": "../../examples",  # path to your example scripts
+    "gallery_dirs": "../auto_examples",  # path to where to save gallery generated output
 }
 
 # -- sphinx.ext.intersphinx
@@ -76,5 +76,3 @@ html_theme = "sphinx_rtd_theme"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
-
-master_doc = 'index'
