### 2021-11-13 03d5f54
Switching tutorials to sphinx gallery
