+++ docs/requirements.txt
@@ -1,3 +1,6 @@
 sphinx==1.8.5
 sphinx-autodoc-typehints
 sphinx-gallery
+pandas
+numpy
+matplotlib
