### 2021-11-14 c666eca
Switching tutorials to sphinx gallery
