### 2021-11-14 d69cb50
Switching tutorials to sphinx gallery
