### 2021-11-14 80f9845
Switching tutorials to sphinx gallery
