### 2021-11-14 302ec31
Switching tutorials to sphinx gallery
