### 2021-11-14 fa74e85
Switching tutorials to sphinx gallery
