### 2021-11-14 9bf2aca
Switching tutorials to sphinx gallery
