### 2021-11-14 21e24cf
Switching tutorials to sphinx gallery
