### 2021-11-14 7700420
Switching tutorials to sphinx gallery
