### 2021-11-14 045e9a3
Switching tutorials to sphinx gallery
