+++ docs/source/conf.py
@@ -14,7 +14,7 @@ import os
 import sys
 import warnings
 
-sys.path.insert(0, os.path.abspath(os.path.join("..", "..", "..", "cca_zoo")))
+sys.path.insert(0, os.path.abspath('../..'))
 warnings.filterwarnings("ignore", category=UserWarning,
                         message="Matplotlib is currently using agg, which is a"
                                 " non-GUI backend, so cannot show the figure.")
