### 2021-11-14 928789b
Switching tutorials to sphinx gallery
