### 2021-11-14 f12c430
Switching tutorials to sphinx gallery
