### 2021-11-14 fe295d1
Switching tutorials to sphinx gallery
