+++ examples/plot_sparse_cca.py
@@ -1,17 +1,10 @@
 """
-"This" is my example-script
+Sparse CCA Methods
 ===========================
 
-This example doesn't do much, it just makes a simple plot
 """
 
 # %%
-
-# A tutorial on using cca-zoo to generate multiview models with sparsity on weights
-
-
-# %%
-
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -21,7 +14,6 @@ from cca_zoo.model_selection import GridSearchCV
 from cca_zoo.models import PMD, SCCA, ElasticCCA, CCA, PLS, SCCA_ADMM, SpanCCA
 
 # %%
-
 np.random.seed(42)
 n = 200
 p = 100
@@ -32,13 +24,11 @@ true_latent_dims = 1
 
 (X, Y), (tx, ty) = generate_covariance_data(n, view_features=[p, q], latent_dims=true_latent_dims,
                                             view_sparsity=[view_1_sparsity, view_2_sparsity], correlation=[0.9])
-# normalize weights for comparability
 tx /= np.sqrt(n)
 ty /= np.sqrt(n)
 
 
 # %%
-
 def plot_true_weights_coloured(ax, weights, true_weights, title=''):
     ind = np.arange(len(true_weights))
     mask = np.squeeze(true_weights == 0)
@@ -58,47 +48,19 @@ def plot_model_weights(wx, wy, tx, ty):
 
 
 # %%
-
-## First try with CCA
-
-# %%
-
-# fit a cca model
 cca = CCA().fit([X, Y])
-
 plot_model_weights(cca.weights[0], cca.weights[1], tx, ty)
 
 # %%
-
-## PLS
-
-# %%
-
-# fit a pls model
 pls = PLS().fit([X, Y])
-
 plot_model_weights(pls.weights[0], pls.weights[1], tx, ty)
 
 # %%
-
-## Penalized Matrix Decomposition (Sparse CCA by Witten)
-# Initially set c=2 for both views arbitrarily
-
-# %%
-
-# fit a pmd model
 pmd = PMD(c=[2, 2]).fit([X, Y])
-
 plot_model_weights(pmd.weights[0], pmd.weights[1], tx, ty)
 
-# %%
-
-## Tracking the objective
-# For these iterative algorithms, you can access the convergence over iterations
 
 # %%
-
-# Convergence
 plt.figure()
 plt.title('Objective Convergence')
 plt.plot(np.array(pmd.track[0]['objective']).T)
@@ -106,38 +68,18 @@ plt.ylabel('Objective')
 plt.xlabel('#iterations')
 
 # %%
-
-### We can also tune the hyperparameter using GridSearchCV
-
-# %%
-
-# Set up a grid. We can't use c<1 or c>sqrt(#features)
 c1 = [1, 3, 7, 9]
 c2 = [1, 3, 7, 9]
 param_grid = {'c': [c1, c2]}
-
-# GridSearchCV can use multiple cores (jobs) and takes folds (number of cv folds) as a parameter. It can also produce a plot.
 pmd = GridSearchCV(PMD(), param_grid=param_grid,
                    cv=3,
                    verbose=True).fit([X, Y])
 
 # %%
-
-# Also the model object now has a pandas dataframe containing the results from each fold
-
-# %%
-
 pd.DataFrame(pmd.cv_results_)
 
 # %%
-
-## Sparse CCA by iterative lasso (Mai)
-
-# %%
-
-# fit a scca model
 scca = SCCA(c=[1e-3, 1e-3]).fit([X, Y])
-
 plot_model_weights(scca.weights[0], scca.weights[1], tx, ty)
 
 # Convergence
@@ -148,15 +90,7 @@ plt.ylabel('Objective')
[…]
