+++ examples/README.rst
@@ -1,4 +1,4 @@
-This is my gallery
-==================
+Tutorials and Examples Gallery
+================================
 
 Below is a gallery of examples
