### 2021-11-14 2918dae
Switching tutorials to sphinx gallery
