### 2021-11-14 88190a2
Switching tutorials to sphinx gallery
