+++ examples/plot_dcca_custom.py
@@ -1,6 +1,6 @@
 """
-Deep CCA
-===========================
+Deep CCA with Bespoke Encoders
+===============================
 
 This example doesn't do much, it just makes a simple plot
 """
