### 2021-11-14 e11260f
Switching tutorials to sphinx gallery
