+++ examples/plot_kernel_ncca.py
@@ -1,0 +1,6 @@
+"""
+Kernel and Nonparametric CCA
+===============================
+
+This example doesn't do much, it just makes a simple plot
+"""
