### 2021-11-14 e449dc9
Switching tutorials to sphinx gallery
