+++ examples/plot_hyperparameter_selection.py
@@ -1,0 +1,6 @@
+"""
+Hyperparameter Selection
+===========================
+
+This will compare MCCA, GCCA, TCCA
+"""
