### 2021-11-14 6fdf50f
Switching tutorials to sphinx gallery
