### 2021-11-14 7d694a8
Switching tutorials to sphinx gallery
