+++ docs/source/index.rst
@@ -17,7 +17,7 @@ Documentation
 
 
 .. toctree::
-   :maxdepth: 1
+   :maxdepth: 2
    :caption: Reference
 
    api/data
