+++ docs/source/api/iterative.rst
@@ -1,0 +1,50 @@
+Iterative Models
+=======================
+
+Normal CCA and PLS by alternating least squares
+--------------------------------------------------
+Potentially quicker for bery large data
+
+.. automodule:: cca_zoo.models.CCA_ALS
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.PLS_ALS
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+
+Sparsity Inducing Models
+--------------------------------------------------
+
+
+.. automodule:: cca_zoo.models.PMD
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.SCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.ElasticCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.SpanCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.ParkhomenkoCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+.. automodule:: cca_zoo.models.SCCA_ADMM
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+Miscellaneous
+--------------------------------------------------
+
+.. automodule:: cca_zoo.models.SWCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
