### 2021-11-14 7fe2c6a
Switching tutorials to sphinx gallery
