+++ cca_zoo/deepmodels/dcca.py
@@ -10,6 +10,11 @@ class DCCA(_DCCA_base):
     """
     A class used to fit a DCCA model.
 
+    Citation
+    --------
+
+    Andrew, Galen, et al. "Deep canonical correlation analysis." International conference on machine learning. PMLR, 2013.
+
     Examples
     --------
     >>> from cca_zoo.deepmodels import DCCA
