+++ cca_zoo/models/iterative.py
@@ -201,6 +201,7 @@ class ElasticCCA(_Iterative):
 
     Citation
     --------
+
     Waaijenborg, Sandra, Philip C. Verselewel de Witt Hamer, and Aeilko H. Zwinderman. "Quantifying the association between gene expressions and DNA-markers by penalized canonical correlation analysis." Statistical applications in genetics and molecular biology 7.1 (2008).
 
     :Example:
@@ -293,6 +294,7 @@ class CCA_ALS(ElasticCCA):
 
     Citation
     --------
+
     Golub, Gene H., and Hongyuan Zha. "The canonical correlations of matrix pairs and their numerical computation." Linear algebra for signal processing. Springer, New York, NY, 1995. 27-49.
 
     :Example:
@@ -357,6 +359,7 @@ class SCCA(ElasticCCA):
 
     Citation
     --------
+
     Mai, Qing, and Xin Zhang. "An iterative penalized least squares approach to sparse canonical correlation analysis." Biometrics 75.3 (2019): 734-744.
 
     :Example:
@@ -423,6 +426,7 @@ class PMD(_Iterative):
 
     Citation
     --------
+
     Witten, Daniela M., Robert Tibshirani, and Trevor Hastie. "A penalized matrix decomposition, with applications to sparse principal components and canonical correlation analysis." Biostatistics 10.3 (2009): 515-534.
 
     :Example:
@@ -495,6 +499,7 @@ class ParkhomenkoCCA(_Iterative):
 
     Citation
     --------
+
     Parkhomenko, Elena, David Tritchler, and Joseph Beyene. "Sparse canonical correlation analysis with application to genomic data integration." Statistical applications in genetics and molecular biology 8.1 (2009).
 
     :Example:
@@ -563,6 +568,7 @@ class SCCA_ADMM(_Iterative):
 
     Citation
     --------
+
     Suo, Xiaotong, et al. "Sparse canonical correlation analysis." arXiv preprint arXiv:1705.10865 (2017).
 
     :Example:
@@ -643,6 +649,7 @@ class SpanCCA(_Iterative):
 
     Citation
     --------
+
     Asteris, Megasthenis, et al. "A simple and provable algorithm for sparse diagonal CCA." International Conference on Machine Learning. PMLR, 2016.
 
     """
@@ -715,6 +722,7 @@ class SWCCA(_Iterative):
 
     Citation
     --------
+
     Wenwen, M. I. N., L. I. U. Juan, and Shihua Zhang. "Sparse Weighted Canonical Correlation Analysis." Chinese Journal of Electronics 27.3 (2018): 459-466.
 
 
