### 2021-11-14 fa44dcc
Switching tutorials to sphinx gallery
