### 2021-11-14 4054793
Switching tutorials to sphinx gallery
