### 2021-11-14 1c41fcb
Switching tutorials to sphinx gallery
