### 2021-11-14 a2d37df
Switching tutorials to sphinx gallery
