### 2021-11-14 e4107e4
Switching tutorials to sphinx gallery
