### 2021-11-14 02f991e
Switching tutorials to sphinx gallery
