### 2021-11-14 2a5fc2e
Switching tutorials to sphinx gallery
