+++ examples/plot_kernel_cca.py
@@ -30,17 +30,19 @@ kernel_reg = GridSearchCV(KCCA(latent_dims=latent_dims), param_grid=param_grid,
                           verbose=True).fit([X, Y]).best_estimator_
 
 # %%
-c1 = [0.9, 0.99]
-c2 = [0.9, 0.99]
-param_grid = {'kernel': ['linear'], 'c': [c1, c2]}
-kernel_reg = GridSearchCV(KCCA(latent_dims=latent_dims), param_grid=param_grid,
-                          cv=cv,
-                          verbose=True).fit([X, Y]).best_estimator_
+degree1 = [2, 3]
+degree2 = [2, 3]
+param_grid = {'kernel': ['poly'], 'degree': [degree1, degree2],
+              'c': [c1, c2]}
+kernel_poly = GridSearchCV(KCCA(latent_dims=latent_dims), param_grid=param_grid,
+                           cv=cv,
+                           verbose=True).fit([X, Y]).best_estimator_
 
-# %%
-c1 = [0.9, 0.99]
-c2 = [0.9, 0.99]
-param_grid = {'kernel': ['linear'], 'c': [c1, c2]}
-kernel_reg = GridSearchCV(KCCA(latent_dims=latent_dims), param_grid=param_grid,
-                          cv=cv,
-                          verbose=True).fit([X, Y]).best_estimator_
+# kernel cca (gaussian)
+gamma1 = [1e+1, 1e+2, 1e+3]
+gamma2 = [1e+1, 1e+2, 1e+3]
+param_grid = {'kernel': ['rbf'], 'gamma': [gamma1, gamma2],
+              'c': [c1, c2]}
+kernel_poly = GridSearchCV(KCCA(latent_dims=latent_dims), param_grid=param_grid,
+                           cv=cv,
+                           verbose=True).fit([X, Y]).best_estimator_
