+++ .github/workflows/.readthedocs.yml
@@ -1,6 +1,7 @@
 # .readthedocs.yml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
+on [workflow_dispatch]
 
 # Required
 version: 2
