### 2021-11-15 96c67f2
Update .readthedocs.yml
