+++ .github/workflows/.readthedocs.yml
@@ -1,7 +1,7 @@
+on: [workflow_dispatch]
 # .readthedocs.yml
 # Read the Docs configuration file
 # See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
-on [workflow_dispatch]
 
 # Required
 version: 2
