### 2021-11-15 4063e84
Switching tutorials to sphinx gallery
