### 2021-11-15 ded8a9c
Delete .readthedocs.yml
