### 2021-11-15 e255bc6
Switching tutorials to sphinx gallery
