### 2021-11-15 ccc11a1
Switching tutorials to sphinx gallery
