+++ requirements/deep.txt
@@ -1,4 +1,4 @@
 torch>=1.9.0
 torchvision
-pytorch-lightning==1.4.9
+pytorch-lightning
 Pillow
