### 2021-11-15 67ca66a
Switching tutorials to sphinx gallery
