### 2021-11-15 2adfa6b
Switching tutorials to sphinx gallery
