### 2021-11-15 b74ffbd
Switching tutorials to sphinx gallery
