### 2021-11-15 323518a
Switching tutorials to sphinx gallery
