### 2021-11-15 5465444
Switching tutorials to sphinx gallery
