### 2021-11-15 8c4a36f
Switching tutorials to sphinx gallery
