### 2021-11-15 e88c804
Switching tutorials to sphinx gallery
