+++ cca_zoo/models/ncca.py
@@ -26,10 +26,17 @@ class NCCA(_CCA_Base):
     >>> model.fit([X1,X2])
     """
 
-    def __init__(self, latent_dims: int = 1, scale=True, centre=True, copy_data=True, accept_sparse=False,
-                 random_state: Union[int, np.random.RandomState] = None, nearest_neighbors=None,
-                 gamma: Iterable[float] = None,
-                 ):
+    def __init__(
+            self,
+            latent_dims: int = 1,
+            scale=True,
+            centre=True,
+            copy_data=True,
+            accept_sparse=False,
+            random_state: Union[int, np.random.RandomState] = None,
+            nearest_neighbors=None,
+            gamma: Iterable[float] = None,
+    ):
         """
         Constructor for NCCA
 
@@ -42,12 +49,16 @@ class NCCA(_CCA_Base):
         :param nearest_neighbors: Number of neaest neighbors (l2 distance) to consider when constructing affinity
         :param gamma: Bandwidth parameter for rbf kernel
         """
-        super().__init__(latent_dims, scale, centre, copy_data, accept_sparse, random_state)
+        super().__init__(
+            latent_dims, scale, centre, copy_data, accept_sparse, random_state
+        )
         self.nearest_neighbors = nearest_neighbors
         self.gamma = gamma
 
     def _check_params(self):
-        self.nearest_neighbors = _process_parameter("nearest_neighbors", self.nearest_neighbors, 1, self.n_views)
+        self.nearest_neighbors = _process_parameter(
+            "nearest_neighbors", self.nearest_neighbors, 1, self.n_views
+        )
         self.gamma = _process_parameter("gamma", self.gamma, None, self.n_views)
         self.kernel = _process_parameter("kernel", None, "rbf", self.n_views)
 
@@ -60,18 +71,25 @@ class NCCA(_CCA_Base):
         self.n = views[0].shape[0]
         self._check_params()
         self.train_views = views
-        self.KNs = [NearestNeighbors(n_neighbors=self.nearest_neighbors[i]).fit(view) for i, view in
-                    enumerate(views)]
-        NNs = [self.KNs[i].kneighbors(view, self.nearest_neighbors[i]) for i, view in enumerate(views)]
+        self.KNs = [
+            NearestNeighbors(n_neighbors=self.nearest_neighbors[i]).fit(view)
+            for i, view in enumerate(views)
+        ]
+        NNs = [
+            self.KNs[i].kneighbors(view, self.nearest_neighbors[i])
+            for i, view in enumerate(views)
+        ]
         kernels = [self._get_kernel(i, view) for i, view in enumerate(self.train_views)]
         self.Ws = [fill_W(kernel, inds) for kernel, (dists, inds) in zip(kernels, NNs)]
-        self.Ws = [self.Ws[0] / self.Ws[0].sum(axis=1, keepdims=True),
-                   self.Ws[1] / self.Ws[1].sum(axis=0, keepdims=True)]
+        self.Ws = [
+            self.Ws[0] / self.Ws[0].sum(axis=1, keepdims=True),
+            self.Ws[1] / self.Ws[1].sum(axis=0, keepdims=True),
+        ]
         S = self.Ws[0] @ self.Ws[1]
         U, S, Vt = np.linalg.svd(S)
-        self.f = U[:, 1:self.latent_dims + 1] * np.sqrt(self.n)
-        self.g = Vt[1:self.latent_dims + 1, :].T * np.sqrt(self.n)
-        self.S = S[1:self.latent_dims + 1]
[…]
