+++ docs/source/conf.py
@@ -14,10 +14,13 @@ import os
 import sys
 import warnings
 
-sys.path.insert(0, os.path.abspath('../..'))
-warnings.filterwarnings("ignore", category=UserWarning,
-                        message="Matplotlib is currently using agg, which is a"
-                                " non-GUI backend, so cannot show the figure.")
+sys.path.insert(0, os.path.abspath("../.."))
+warnings.filterwarnings(
+    "ignore",
+    category=UserWarning,
+    message="Matplotlib is currently using agg, which is a"
+            " non-GUI backend, so cannot show the figure.",
+)
 # -- Project information -----------------------------------------------------
 
 project = "cca-zoo"
@@ -33,7 +36,7 @@ extensions = [
     "sphinx.ext.viewcode",
     "sphinx_gallery.gen_gallery",
     "sphinx.ext.napoleon",
-    "sphinx.ext.githubpages"
+    "sphinx.ext.githubpages",
 ]
 
 sphinx_gallery_conf = {
@@ -79,4 +82,4 @@ html_theme = "sphinx_rtd_theme"
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ["_static"]
 
-master_doc = 'index'
+master_doc = "index"
