### 2021-11-15 #3823 @emdupre
Thanks @ejolly for your review !

It looks like https://github.com/jameschapman19/cca_zoo/issues/74, https://github.com/jameschapman19/cca_zoo/issues/75, https://github.com/jameschapman19/cca_zoo/issues/76, and https://github.com/jameschapman19/cca_zoo/issues/77 were all created in response to reviewer feedback, so I'll monitor those issues to see how the revisions are going.

@hugorichard @robbisg @ejolly if there are any additional issues you noted that you consider to _not_ be covered by those issues, please let us know here !
