### 2021-11-15 a0292ed
Switching tutorials to sphinx gallery
