+++ cca_zoo/test/test_deepmodels.py
@@ -60,13 +60,13 @@ def test_large_p():
 
 def test_DCCA_methods():
     latent_dims = 2
-    epochs = 50
+    epochs = 100
     cca = CCA(latent_dims=latent_dims).fit((X, Y))
     # Soft Decorrelation (stochastic Decorrelation Loss)
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
     encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
-    sdl = DCCA_SDL(latent_dims, X.shape[0], encoders=[encoder_1, encoder_2], lam=1e-2)
-    optimizer = optim.SGD(sdl.parameters(), lr=1e-3)
+    sdl = DCCA_SDL(latent_dims, X.shape[0], encoders=[encoder_1, encoder_2], lam=1e-3)
+    optimizer = optim.SGD(sdl.parameters(), lr=1e-1)
     sdl = CCALightning(sdl, optimizer=optimizer)
     trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
     trainer.fit(sdl, train_loader)
@@ -80,9 +80,9 @@ def test_DCCA_methods():
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
     encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=12)
     dcca_noi = DCCA_NOI(latent_dims, X.shape[0], encoders=[encoder_1, encoder_2], rho=0)
-    optimizer = optim.Adam(dcca_noi.parameters(), lr=1e-3)
+    optimizer = optim.SGD(dcca_noi.parameters(), lr=1e-2)
     dcca_noi = CCALightning(dcca_noi, optimizer=optimizer)
-    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
+    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10, enable_checkpointing=False)
     trainer.fit(dcca_noi, train_loader)
     assert (
             np.testing.assert_array_less(
@@ -100,7 +100,7 @@ def test_DCCA_methods():
     )
     optimizer = optim.SGD(dcca.parameters(), lr=1e-3)
     dcca = CCALightning(dcca, optimizer=optimizer)
-    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
+    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10, enable_checkpointing=False)
     trainer.fit(dcca, train_loader)
     assert (
             np.testing.assert_array_less(
@@ -118,7 +118,7 @@ def test_DCCA_methods():
     )
     optimizer = optim.SGD(dgcca.parameters(), lr=1e-3)
     dgcca = CCALightning(dgcca, optimizer=optimizer)
-    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
+    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10, enable_checkpointing=False)
     trainer.fit(dgcca, train_loader)
     assert (
             np.testing.assert_array_less(
@@ -136,7 +136,7 @@ def test_DCCA_methods():
     )
     optimizer = optim.SGD(dmcca.parameters(), lr=1e-3)
     dmcca = CCALightning(dmcca, optimizer=optimizer)
-    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
+    trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10, enable_checkpointing=False)
     trainer.fit(dmcca, train_loader)
     assert (
             np.testing.assert_array_less(
[…]
