### 2021-11-15 1aec37b
Switching tutorials to sphinx gallery
