### 2021-11-15 7fed1b0
Switching tutorials to sphinx gallery
