+++ .github/workflows/python-package.yml
@@ -13,6 +13,8 @@ jobs:
   build:
 
     runs-on: ubuntu-latest
+    env:
+      MODULE_NAME: cca_zoo
     strategy:
       matrix:
         python-version: [3.7, 3.8, 3.9]
