### 2021-11-15 49e5d87
Switching tutorials to sphinx gallery
