+++ cca_zoo/models/gcca.py
@@ -24,8 +24,9 @@ class GCCA(rCCA):
     >>> rng=np.random.RandomState(0)
     >>> X1 = rng.random(10,5)
     >>> X2 = np.random.rand(10,5)
+    >>> X3 = np.random.rand(10,5)
     >>> model = GCCA()
-    >>> model.fit([X1,X2])
+    >>> model.fit((X1,X2,X3))
     """
 
     def __init__(
@@ -109,8 +110,9 @@ class KGCCA(GCCA):
     >>> rng=np.random.RandomState(0)
     >>> X1 = rng.random(10,5)
     >>> X2 = np.random.rand(10,5)
+    >>> X3 = np.random.rand(10,5)
     >>> model = KGCCA()
-    >>> model.fit([X1,X2])
+    >>> model.fit((X1,X2,X3))
     """
 
     def __init__(
