+++ .github/workflows/builddocs.yml
@@ -1,15 +1,0 @@
-on: workflow_dispatch
-  # Build docs
-  - run:
-      name: make html
-      command: |
-        cd docs;
-        make html;
-  # Save the outputs
-  - store_artifacts:
-      path: docs/_build/html/
-      destination: dev
-  - persist_to_workspace:
-      root: docs/_build
-      paths:
-        - html
