### 2021-11-16 022d6c5
Doctests
