### 2021-11-16 a1c43db
Doctests
