+++ cca_zoo/deepmodels/dcca_barlow_twins.py
@@ -15,9 +15,6 @@ class BarlowTwins(DCCA):
 
     Zbontar, Jure, et al. "Barlow twins: Self-supervised learning via redundancy reduction." arXiv preprint arXiv:2103.03230 (2021).
 
-    Examples
-    --------
-
     """
 
     def __init__(
