### 2021-11-16 93aa41c
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
