### 2021-11-16 67948a2
Update setup.py
added 'all' option as per review
