### 2021-11-16 b643e05
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
