### 2021-11-16 1ab5ed7
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
