### 2021-11-16 f0dc07b
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
