### 2021-11-16 7de358b
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
