+++ examples/plot_dcca.py
@@ -36,12 +36,14 @@ epochs = 10
 encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=392)
 encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=392)
 
+# %%
 # Deep CCA
 dcca = DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
 dcca = CCALightning(dcca)
 trainer = pl.Trainer(max_epochs=epochs, enable_checkpointing=False)
 trainer.fit(dcca, train_loader, val_loader)
 
+# %%
 # Deep CCA by Non-Linear Orthogonal Iterations
 dcca_noi = DCCA_NOI(
     latent_dims=latent_dims, N=len(train_dataset), encoders=[encoder_1, encoder_2]
@@ -50,6 +52,7 @@ dcca_noi = CCALightning(dcca_noi)
 trainer = pl.Trainer(max_epochs=epochs, enable_checkpointing=False)
 trainer.fit(dcca_noi, train_loader, val_loader)
 
+# %%
 # Deep CCA by Stochastic Decorrelation Loss
 dcca_sdl = DCCA_SDL(
     latent_dims=latent_dims, N=len(train_dataset), encoders=[encoder_1, encoder_2]
@@ -58,6 +61,7 @@ dcca_sdl = CCALightning(dcca_sdl)
 trainer = pl.Trainer(max_epochs=epochs, enable_checkpointing=False)
 trainer.fit(dcca_sdl, train_loader, val_loader)
 
+# %%
 # Deep CCA by Barlow Twins
 barlowtwins = BarlowTwins(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
 barlowtwins = CCALightning(barlowtwins)
