### 2021-11-16 3282116
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
