+++ docs/source/index.rst
@@ -17,7 +17,7 @@ Documentation
 
 
 .. toctree::
-   :maxdepth: 2
+   :maxdepth: 3
    :caption: Reference
 
    api/data
