### 2021-11-16 29a2bc3
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
