### 2021-11-16 a26bfb4
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
