### 2021-11-16 19bcce5
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
