### 2021-11-16 6e70bc9
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
