+++ cca_zoo/models/rcca.py
@@ -31,6 +31,7 @@ class rCCA(_CCA_Base):
 
     Citation
     --------
+
     .. Vinod, Hrishikesh D. "Canonical ridge and econometrics of joint production." Journal of econometrics 4.2 (1976): 147-166.
 
     :Example:
@@ -193,7 +194,8 @@ class CCA(rCCA):
         w_2^Tw_2=1
 
     Citation
-    --------
+    ---------
+
     .. Hotelling, Harold. "Relations between two sets of variates." Breakthroughs in statistics. Springer, New York, NY, 1992. 162-190.
 
     :Example:
