+++ cca_zoo/models/rcca.py
@@ -24,9 +24,9 @@ class rCCA(_CCA_Base):
 
         \text{subject to:}
 
-        w_1^Tw_1=1
+        (1-c_1)w_1^TX_1^TX_1w_1+c_1w_1^Tw_1=1
 
-        w_2^Tw_2=1
+        (1-c_2)w_2^TX_2^TX_2w_2+c_2w_2^Tw_2=1
 
 
     Citation
@@ -189,9 +189,9 @@ class CCA(rCCA):
 
         \text{subject to:}
 
-        w_1^Tw_1=1
+        w_1^TX_1^TX_1w_1=1
 
-        w_2^Tw_2=1
+        w_2^TX_2^TX_2w_2=1
 
     Citation
     ---------
