### 2021-11-16 df37754
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
