### 2021-11-16 aa73ff8
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
