+++ cca_zoo/models/gcca.py
@@ -162,6 +162,8 @@ class KGCCA(GCCA):
             kernel_params: Iterable[dict] = None,
     ):
         """
+        Constructor for PLS
+
         :param latent_dims: number of latent dimensions to fit
         :param scale: normalize variance in each column before fitting
         :param centre: demean data by column before fitting (and before transforming out of sample
