+++ cca_zoo/deepmodels/dcca.py
@@ -12,6 +12,7 @@ class DCCA(_DCCA_base):
 
     Citation
     --------
+
     .. Andrew, Galen, et al. "Deep canonical correlation analysis." International conference on machine learning. PMLR, 2013.
 
     """
