+++ cca_zoo/deepmodels/dcca_noi.py
@@ -11,6 +11,7 @@ class DCCA_NOI(DCCA):
 
     Citation
     --------
+
     .. Wang, Weiran, et al. "Stochastic optimization for deep CCA via nonlinear orthogonal iterations." 2015 53rd Annual Allerton Conference on Communication, Control, and Computing (Allerton). IEEE, 2015.
 
     """
