### 2021-11-16 af07f18
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
