### 2021-11-16 8141bec
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
