### 2021-11-16 31cd5b7
Remove Tutorial Notebooks as they are replaced with sphinx-gallery
