+++ cca_zoo/models/iterative.py
@@ -218,8 +218,8 @@ class ElasticCCA(_Iterative):
 
         w_i^TX_i^TX_iw_i=1
 
-    Citation
-    --------
+    :Citation:
+
     .. Waaijenborg, Sandra, Philip C. Verselewel de Witt Hamer, and Aeilko H. Zwinderman. "Quantifying the association between gene expressions and DNA-markers by penalized canonical correlation analysis." Statistical applications in genetics and molecular biology 7.1 (2008).
 
     :Example:
@@ -321,8 +321,8 @@ class CCA_ALS(ElasticCCA):
 
         w_i^TX_i^TX_iw_i=1
 
-    Citation
-    --------
+    :Citation:
+
     .. Golub, Gene H., and Hongyuan Zha. "The canonical correlations of matrix pairs and their numerical computation." Linear algebra for signal processing. Springer, New York, NY, 1995. 27-49.
 
     :Example:
@@ -396,8 +396,8 @@ class SCCA(ElasticCCA):
 
         w_i^TX_i^TX_iw_i=1
 
-    Citation
-    --------
+    :Citation:
+
     .. Mai, Qing, and Xin Zhang. "An iterative penalized least squares approach to sparse canonical correlation analysis." Biometrics 75.3 (2019): 734-744.
 
     :Example:
@@ -465,8 +465,18 @@ class PMD(_Iterative):
     """
     Fits a Sparse CCA (Penalized Matrix Decomposition) model.
 
-    Citation
-    --------
+    .. math::
+
+        w_{opt}=\underset{w}{\mathrm{argmax}}\{ w_1^TX_1^TX_2w_2  \}\\
+
+        \text{subject to:}
+
+        w_i^Tw_i=1
+
+        \|w_i\|<=c_i
+
+    :Citation:
+
     .. Witten, Daniela M., Robert Tibshirani, and Trevor Hastie. "A penalized matrix decomposition, with applications to sparse principal components and canonical correlation analysis." Biostatistics 10.3 (2009): 515-534.
 
     :Example:
@@ -540,8 +550,16 @@ class ParkhomenkoCCA(_Iterative):
     """
     Fits a sparse CCA (penalized CCA) model
 
-    Citation
-    --------
+    .. math::
+
+        w_{opt}=\underset{w}{\mathrm{argmax}}\{ w_1^TX_1^TX_2w_2  \} + c_i\|w_i\|\\
+
+        \text{subject to:}
+
+        w_i^Tw_i=1
+
+    :Citation:
+
     .. Parkhomenko, Elena, David Tritchler, and Joseph Beyene. "Sparse canonical correlation analysis with application to genomic data integration." Statistical applications in genetics and molecular biology 8.1 (2009).
 
     :Example:
@@ -611,8 +629,8 @@ class SCCA_ADMM(_Iterative):
     """
     Fits a sparse CCA model by alternating ADMM
 
-    Citation
-    --------
+    :Citation:
+
     .. Suo, Xiaotong, et al. "Sparse canonical correlation analysis." arXiv preprint arXiv:1705.10865 (2017).
 
     :Example:
@@ -694,8 +712,8 @@ class SpanCCA(_Iterative):
     """
     Fits a Sparse CCA model using SpanCCA.
 
-    Citation
-    --------
+    :Citation:
+
     .. Asteris, Megasthenis, et al. "A simple and provable algorithm for sparse diagonal CCA." International Conference on Machine Learning. PMLR, 2016.
 
 
@@ -780,7 +798,6 @@ class SWCCA(_Iterative):
     :Citation:
 
[…]
