+++ cca_zoo/models/iterative.py
@@ -33,7 +33,6 @@ class _Iterative(_CCA_Base):
             random_state=None,
             deflation="cca",
             max_iter: int = 100,
-            generalized: bool = False,
             initialization: str = "unregularized",
             tol: float = 1e-9,
     ):
@@ -47,7 +46,7 @@ class _Iterative(_CCA_Base):
         :param random_state: Pass for reproducible output across multiple function calls
         :param deflation: the type of deflation.
         :param max_iter: the maximum number of iterations to perform in the inner optimization loop
-        :param generalized: use auxiliary variables (required for >2 views)
+
         :param initialization: intialization for optimisation. 'unregularized' uses CCA or PLS solution,'random' uses random initialization,'uniform' uses uniform initialization of weights and scores
         :param tol: tolerance value used for early stopping
         """
@@ -59,7 +58,6 @@ class _Iterative(_CCA_Base):
             accept_sparse=["csc", "csr"],
         )
         self.max_iter = max_iter
-        self.generalized = generalized
         self.initialization = initialization
         self.tol = tol
         self.deflation = deflation
@@ -126,7 +124,6 @@ class _Iterative(_CCA_Base):
         """
         self.loop = PLSInnerLoop(
             max_iter=self.max_iter,
-            generalized=self.generalized,
             initialization=self.initialization,
             random_state=self.random_state,
         )
@@ -166,7 +163,6 @@ class PLS_ALS(_Iterative):
             copy_data=True,
             random_state=None,
             max_iter: int = 100,
-            generalized: bool = False,
             initialization: str = "unregularized",
             tol: float = 1e-9,
     ):
@@ -179,7 +175,6 @@ class PLS_ALS(_Iterative):
         :param copy_data: If True, X will be copied; else, it may be overwritten
         :param random_state: Pass for reproducible output across multiple function calls
         :param max_iter: the maximum number of iterations to perform in the inner optimization loop
-        :param generalized: use auxiliary variables (required for >2 views)
         :param initialization: intialization for optimisation. 'unregularized' uses CCA or PLS solution,'random' uses random initialization,'uniform' uses uniform initialization of weights and scores
         :param tol: tolerance value used for early stopping
         """
@@ -190,7 +185,6 @@ class PLS_ALS(_Iterative):
             copy_data=copy_data,
             deflation="pls",
             max_iter=max_iter,
-            generalized=generalized,
             initialization=initialization,
             tol=tol,
             random_state=random_state,
@@ -199,7 +193,6 @@ class PLS_ALS(_Iterative):
     def _set_loop_params(self):
         self.loop = PLSInnerLoop(
             max_iter=self.max_iter,
-            generalized=self.generalized,
             initialization=self.initialization,
             tol=self.tol,
             random_state=self.random_state,
@@ -208,7 +201,7 @@ class PLS_ALS(_Iterative):
 
 class ElasticCCA(_Iterative):
     """
-    Fits an elastic CCA by iterative rescaled elastic net regression
+    Fits an elastic CCA by iterating elastic net regression
 
     .. math::
 
@@ -220,7 +213,7 @@ class ElasticCCA(_Iterative):
 
     :Citation:
 
-    .. Waaijenborg, Sandra, Philip C. Verselewel de Witt Hamer, and Aeilko H. Zwinderman. "Quantifying the association between gene expressions and DNA-markers by penalized canonical correlation analysis." Statistical applications in genetics and molecular biology 7.1 (2008).
+    .. Fu, Xiao, et al. "Scalable and flexible multiview MAX-VAR canonical correlation analysis." IEEE Transactions on Signal Processing 65.16 (2017): 4150-4165.
 
     :Example:
 
[…]
