### 2021-11-16 4048c08
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
