+++ cca_zoo/deepmodels/dcca_barlow_twins.py
@@ -13,7 +13,7 @@ class BarlowTwins(DCCA):
     Citation
     --------
 
-    .. Zbontar, Jure, et al. "Barlow twins: Self-supervised learning via redundancy reduction." arXiv preprint arXiv:2103.03230 (2021).
+    Zbontar, Jure, et al. "Barlow twins: Self-supervised learning via redundancy reduction." arXiv preprint arXiv:2103.03230 (2021).
 
     """
 
