+++ cca_zoo/deepmodels/dcca.py
@@ -13,7 +13,7 @@ class DCCA(_DCCA_base):
     Citation
     --------
 
-    .. Andrew, Galen, et al. "Deep canonical correlation analysis." International conference on machine learning. PMLR, 2013.
+    Andrew, Galen, et al. "Deep canonical correlation analysis." International conference on machine learning. PMLR, 2013.
 
     """
 
