+++ cca_zoo/deepmodels/dccae.py
@@ -13,7 +13,7 @@ class DCCAE(_DCCA_base):
     Citation
     --------
 
-    .. Wang, Weiran, et al. "On deep multi-view representation learning." International conference on machine learning. PMLR, 2015.
+    Wang, Weiran, et al. "On deep multi-view representation learning." International conference on machine learning. PMLR, 2015.
 
     """
 
