+++ docs/source/index.rst
@@ -17,7 +17,7 @@ Documentation
 
 
 .. toctree::
-   :maxdepth: 3
+   :maxdepth: 4
    :caption: Reference
 
    api/data
