### 2021-11-16 cbbc4a3
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
