+++ cca_zoo/deepmodels/dcca_sdl.py
@@ -8,8 +8,7 @@ class DCCA_SDL(DCCA_NOI):
     """
     A class used to fit a Deep CCA by Stochastic Decorrelation model.
 
-    Citation
-    --------
+    :Citation:
 
     Chang, Xiaobin, Tao Xiang, and Timothy M. Hospedales. "Scalable and effective deep CCA via soft decorrelation." Proceedings of the IEEE Conference on Computer Vision and Pattern Recognition. 2018.
 
