+++ cca_zoo/probabilisticmodels/vcca.py
@@ -15,9 +15,9 @@ class VariationalCCA(_CCA_Base):
     """
     A class used to fit a variational bayesian CCA. Not quite the same due to using VI methods rather than EM
 
-    Citation
-    --------
-    .. Wang, Chong. "Variational Bayesian approach to canonical correlation analysis." IEEE Transactions on Neural Networks 18.3 (2007): 905-910.
+    :Citation:
+
+    Wang, Chong. "Variational Bayesian approach to canonical correlation analysis." IEEE Transactions on Neural Networks 18.3 (2007): 905-910.
 
     """
 
