### 2021-11-16 c24e82f
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
