+++ cca_zoo/models/mcca.py
@@ -180,7 +180,7 @@ class KCCA(MCCA):
         self.c = _process_parameter("c", self.c, 0, self.n_views)
 
     def _get_kernel(self, view, X, Y=None):
-        if callable(self.kernel):
+        if callable(self.kernel[view]):
             params = self.kernel_params[view] or {}
         else:
             params = {
