+++ examples/plot_kernel_cca.py
@@ -24,6 +24,21 @@ cv = 3
     n, view_features=[p, q], latent_dims=latent_dims, correlation=[0.9]
 )
 
+
+# %%
+# Custom Kernel
+def my_kernel(X, Y, param=0):
+    """
+    We create a custom kernel:
+
+    """
+
+    return np.random.normal(0, param)
+
+
+kernel_custom = KCCA(latent_dims=latent_dims, kernel=[my_kernel, my_kernel],
+                     kernel_params=[{'param': 1}, {'param': 1}]).fit((X, Y))
+
 # %%
 # Linear
 c1 = [0.9, 0.99]
@@ -52,8 +67,8 @@ kernel_poly = (
 
 # %%
 # kernel cca (gaussian/rbf)
-gamma1 = [1e1, 1e2, 1e3]
-gamma2 = [1e1, 1e2, 1e3]
+gamma1 = [1e-1, 1e-2, 1e-3]
+gamma2 = [1e-1, 1e-2, 1e-3]
 param_grid = {"kernel": ["rbf"], "gamma": [gamma1, gamma2], "c": [c1, c2]}
 kernel_poly = (
     GridSearchCV(
@@ -62,3 +77,18 @@ kernel_poly = (
         .fit([X, Y])
         .best_estimator_
 )
+
+
+# %%
+# Custom Kernel
+def my_kernel(X, Y, param=0):
+    """
+    We create a custom kernel:
+
+    """
+    M = np.random.rand(X.shape[1], X.shape[1]) + param
+    return X @ M @ M.T @ Y.T
+
+
+kernel_custom = KCCA(latent_dims=latent_dims, kernel=[my_kernel, my_kernel],
+                     kernel_params=[{'param': 1}, {'param': 1}]).fit((X, Y))
