+++ cca_zoo/models/tcca.py
@@ -243,7 +243,7 @@ class KTCCA(TCCA):
         self.c = _process_parameter("c", self.c, 0, self.n_views)
 
     def _get_kernel(self, view, X, Y=None):
-        if callable(self.kernel):
+        if callable(self.kernel[view]):
             params = self.kernel_params[view] or {}
         else:
             params = {
