### 2021-11-16 f5c1a1b
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
