### 2021-11-16 810ba76
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
