### 2021-11-16 e2c60a8
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
