### 2021-11-17 7826a4b
Removing the slow constrained option for Elastic CCA. Now uses maxvar by default and sumcor for lasso where valid
