### 2021-11-17 d142fba
Changed iterative error to warning. This stops gridsearch breaking if just one parameter is in the wrong range
