+++ cca_zoo/models/innerloop.py
@@ -6,9 +6,6 @@ import numpy as np
 from sklearn.exceptions import ConvergenceWarning
 from sklearn.linear_model import (
     ElasticNet,
-    Lasso,
-    LinearRegression,
-    Ridge,
     SGDRegressor,
 )
 from sklearn.utils._testing import ignore_warnings
@@ -91,9 +88,12 @@ class _InnerLoop:
         for _ in range(self.max_iter):
             self._inner_iteration()
             if np.isnan(self.scores).sum() > 0:
+                warnings.warn(
+                    f"Some scores are nan. Usually regularisation is too high."
+                )
                 break
             self.track["objective"].append(self._objective())
-            if _ > 0 and self._early_stop():
+            if _ > 1 and self._early_stop():
                 self.track["converged"] = True
                 break
             self.old_scores = self.scores.copy()
@@ -142,6 +142,7 @@ class PLSInnerLoop(_InnerLoop):
     def _inner_iteration(self):
         # Update each view using loop update function
         for i, view in enumerate(self.views):
+            # if no nans
             if np.isnan(self.scores).sum() == 0:
                 self._update_view(i)
 
@@ -310,80 +311,28 @@ class ElasticInnerLoop(PLSInnerLoop):
         self.regressors = []
         for alpha, l1_ratio, positive in zip(self.c, self.l1_ratio, self.positive):
             if self.stochastic:
-                if l1_ratio == 0:
-                    self.regressors.append(
-                        SGDRegressor(
-                            penalty="l2",
-                            alpha=alpha / len(self.views),
-                            fit_intercept=False,
-                            tol=self.tol,
-                            warm_start=True,
-                            random_state=self.random_state,
-                        )
+                self.regressors.append(
+                    SGDRegressor(
+                        penalty="elasticnet",
+                        alpha=alpha / len(self.views),
+                        l1_ratio=l1_ratio,
+                        fit_intercept=False,
+                        tol=self.tol,
+                        warm_start=True,
+                        random_state=self.random_state,
                     )
-                elif l1_ratio == 1:
-                    self.regressors.append(
-                        SGDRegressor(
-                            penalty="l1",
-                            alpha=alpha / len(self.views),
-                            fit_intercept=False,
-                            tol=self.tol,
-                            warm_start=True,
-                            random_state=self.random_state,
-                        )
-                    )
-                else:
-                    self.regressors.append(
-                        SGDRegressor(
-                            penalty="elasticnet",
-                            alpha=alpha / len(self.views),
-                            l1_ratio=l1_ratio,
-                            fit_intercept=False,
-                            tol=self.tol,
-                            warm_start=True,
-                            random_state=self.random_state,
-                        )
-                    )
+                )
             else:
-                if alpha == 0:
-                    self.regressors.append(LinearRegression(fit_intercept=False))
-                elif l1_ratio == 0:
-                    if positive:
-                        self.regressors.append(
-                            ElasticNet(
-                                alpha=alpha / len(self.views),
-                                l1_ratio=0,
-                                fit_intercept=False,
-                                warm_start=True,
-                                positive=positive,
-                                random_state=self.random_state,
-                            )
-                        )
-                    else:
-                        self.regressors.append(
-                            Ridge(alpha=alpha / len(self.views), fit_intercept=False)
-                        )
-                elif l1_ratio == 1:
-                    self.regressors.append(
-                        Lasso(
-                            alpha=alpha / len(self.views),
[…]
