### 2021-11-17 980dcd4
Changed iterative error to warning. This stops gridsearch breaking if just one parameter is in the wrong range
