+++ cca_zoo/test/test_models.py
@@ -189,18 +189,20 @@ def test_regularized_methods():
 
 def test_non_negative_methods():
     latent_dims = 2
-    nnscca = SCCA(latent_dims=latent_dims, tol=1e-9, positive=True, c=[1e-5, 1e-5], random_state=0).fit(
-        (X, Y)
-    )
+    nnscca = SCCA(
+        latent_dims=latent_dims, tol=1e-9, positive=True, c=[1e-1, 1e-1], random_state=0
+    ).fit((X, Y))
     nnelastic = ElasticCCA(
         latent_dims=latent_dims,
         tol=1e-9,
         positive=True,
         l1_ratio=[0.5, 0.5],
         c=[1e-4, 1e-5],
-        random_state=0
+        random_state=0,
     ).fit([X, Y])
-    nnals = CCA_ALS(latent_dims=latent_dims, tol=1e-9, positive=True, random_state=0).fit([X, Y])
+    nnals = CCA_ALS(
+        latent_dims=latent_dims, tol=1e-9, positive=True, random_state=0
+    ).fit([X, Y])
     assert np.testing.assert_array_less(-1e-9, nnelastic.weights[0]) is None
     assert np.testing.assert_array_less(-1e-9, nnelastic.weights[1]) is None
     assert np.testing.assert_array_less(-1e-9, nnals.weights[0]) is None
@@ -222,11 +224,15 @@ def test_sparse_methods():
     c1 = [1e-1]
     c2 = [1e-1]
     param_grid = {"c": [c1, c2]}
-    parkhomenko_cv = GridSearchCV(ParkhomenkoCCA(random_state=rng), param_grid=param_grid).fit([X, Y])
+    parkhomenko_cv = GridSearchCV(
+        ParkhomenkoCCA(random_state=rng), param_grid=param_grid
+    ).fit([X, Y])
     c1 = [2e-2]
     c2 = [1e-2]
     param_grid = {"c": [c1, c2]}
-    admm_cv = GridSearchCV(SCCA_ADMM(random_state=rng), param_grid=param_grid).fit([X, Y])
+    admm_cv = GridSearchCV(SCCA_ADMM(random_state=rng), param_grid=param_grid).fit(
+        [X, Y]
+    )
     c1 = loguniform(1e-1, 2e-1)
     c2 = loguniform(1e-1, 2e-1)
     param_grid = {"c": [c1, c2], "l1_ratio": [[0.9], [0.9]]}
