### 2021-11-17 eb309ee
Changed iterative error to warning. This stops gridsearch breaking if just one parameter is in the wrong range
