### 2021-11-17 75101ed
Small doc changes updating examples
