+++ examples/plot_ridge_reg.py
@@ -1,6 +1,0 @@
-"""
-Ridge Regularised CCA: From CCA to PLS
-===========================
-
-This script will show how CCA and PLS form opposite ends of a ridge regularisation spectrum
-"""
