+++ examples/plot_dcca_custom.py
@@ -32,8 +32,8 @@ encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=392)
 
 # Deep CCA
 dcca = DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
-dcca = CCALightning(dcca)
 optimizer = optim.Adam(dcca.parameters(), lr=1e-3)
 scheduler = optim.lr_scheduler.CosineAnnealingLR(optimizer, 1)
+dcca = CCALightning(dcca, optimizer=optimizer, scheduler=scheduler)
 trainer = pl.Trainer(max_epochs=epochs, enable_checkpointing=False)
 trainer.fit(dcca, train_loader, val_loader)
