### 2021-11-17 3de1634
Small doc changes updating examples
