### 2021-11-17 dfe46ee
Small doc changes updating examples
