### 2021-11-17 b4f1d28
Small doc changes updating examples
