### 2021-11-17 d316f94
Small doc changes updating examples
