### 2021-11-17 6db3cb4
Small doc changes updating examples
