+++ cca_zoo/test/test_deepmodels.py
@@ -25,8 +25,8 @@ rng = check_random_state(0)
 X = rng.rand(200, 10)
 Y = rng.rand(200, 12)
 Z = rng.rand(200, 14)
-X_conv = rng.rand(100, 1, 16, 16)
-Y_conv = rng.rand(100, 1, 16, 16)
+X_conv = rng.rand(200, 1, 16, 16)
+Y_conv = rng.rand(200, 1, 16, 16)
 dataset = data.CCA_Dataset([X, Y, Z])
 train_dataset, val_dataset = process_data(dataset, val_split=0.2)
 train_dataset_numpy, val_dataset_numpy = process_data((X, Y, Z), val_split=0.2)
@@ -53,10 +53,10 @@ def test_DCCA_methods():
     )
     trainer.fit(dcca_noi, train_loader)
     assert (
-        np.testing.assert_array_less(
-            cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
-        )
-        is None
+            np.testing.assert_array_less(
+                cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
+            )
+            is None
     )
     # Soft Decorrelation (stochastic Decorrelation Loss)
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
@@ -67,10 +67,10 @@ def test_DCCA_methods():
     trainer = pl.Trainer(max_epochs=epochs, log_every_n_steps=10)
     trainer.fit(sdl, train_loader)
     assert (
-        np.testing.assert_array_less(
-            cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
-        )
-        is None
+            np.testing.assert_array_less(
+                cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
+            )
+            is None
     )
     # DCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
@@ -87,10 +87,10 @@ def test_DCCA_methods():
     )
     trainer.fit(dcca, train_loader)
     assert (
-        np.testing.assert_array_less(
-            cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
-        )
-        is None
+            np.testing.assert_array_less(
+                cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
+            )
+            is None
     )
     # DGCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
@@ -107,10 +107,10 @@ def test_DCCA_methods():
     )
     trainer.fit(dgcca, train_loader)
     assert (
-        np.testing.assert_array_less(
-            cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
-        )
-        is None
+            np.testing.assert_array_less(
+                cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
+            )
+            is None
     )
     # DMCCA
     encoder_1 = architectures.Encoder(latent_dims=latent_dims, feature_size=10)
@@ -127,10 +127,10 @@ def test_DCCA_methods():
     )
     trainer.fit(dmcca, train_loader)
     assert (
-        np.testing.assert_array_less(
-            cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
-        )
-        is None
+            np.testing.assert_array_less(
+                cca.score((X, Y)).sum(), trainer.model.score(train_loader).sum()
+            )
+            is None
     )
     # Barlow Twins
[…]
