+++ examples/plot_dcca_custom.py
@@ -34,6 +34,6 @@ encoder_2 = architectures.Encoder(latent_dims=latent_dims, feature_size=392)
 dcca = DCCA(latent_dims=latent_dims, encoders=[encoder_1, encoder_2])
 optimizer = optim.Adam(dcca.parameters(), lr=1e-3)
 scheduler = optim.lr_scheduler.CosineAnnealingLR(optimizer, 1)
-dcca = CCALightning(dcca, optimizer=optimizer, scheduler=scheduler)
+dcca = CCALightning(dcca, optimizer=optimizer, lr_scheduler=scheduler)
 trainer = pl.Trainer(max_epochs=epochs, enable_checkpointing=False)
 trainer.fit(dcca, train_loader, val_loader)
