+++ cca_zoo/deepmodels/splitae.py
@@ -30,9 +30,17 @@ class SplitAE(_DCCA_base):
 
     def forward(self, *args):
         z = self.encoder(args[0])
-        return [z]
+        return z
 
-    def decode(self, z):
+    def recon(self, *args):
+        """
+        :param args:
+        :return:
+        """
+        z = self(*args)
+        return self._decode(z)
+
+    def _decode(self, z):
         """
         This method is used to decode from the latent space to the best prediction of the original views
 
@@ -45,7 +53,7 @@ class SplitAE(_DCCA_base):
 
     def loss(self, *args):
         z = self(*args)
-        recon = self.decode(*z)
+        recon = self._decode(z)
         recon_loss = self.recon_loss(args, recon)
         return recon_loss
 

[…]
