+++ cca_zoo/data/toy.py
@@ -4,10 +4,9 @@ implementation of their work soon) Check out their paper at https://arxiv.org/ab
 import numpy as np
 import torch
 import torch.utils.data
-from PIL import Image
+import torchvision
 from torch.utils.data import Dataset
 from torchvision import datasets, transforms
-from torchvision.transforms.functional import InterpolationMode
 
 
 class Split_MNIST_Dataset(Dataset):
@@ -16,7 +15,7 @@ class Split_MNIST_Dataset(Dataset):
     """
 
     def __init__(
-        self, mnist_type: str = "MNIST", train: bool = True, flatten: bool = True
+            self, mnist_type: str = "MNIST", train: bool = True, flatten: bool = True
     ):
         """
 
@@ -72,7 +71,7 @@ class Noisy_MNIST_Dataset(Dataset):
     """
 
     def __init__(
-        self, mnist_type: str = "MNIST", train: bool = True, flatten: bool = True
+            self, mnist_type: str = "MNIST", train: bool = True, flatten: bool = True
     ):
         """
 
@@ -81,25 +80,32 @@ class Noisy_MNIST_Dataset(Dataset):
         :param flatten: whether to flatten the data into array or use 2d images
         """
         if mnist_type == "MNIST":
-            self.dataset = datasets.MNIST("../../data", train=train, download=True)
+            self.dataset = datasets.MNIST("../../data", train=train, download=True,
+                                          transform=torchvision.transforms.Compose([
+                                              torchvision.transforms.ToTensor()]))
         elif mnist_type == "FashionMNIST":
             self.dataset = datasets.FashionMNIST(
-                "../../data", train=train, download=True
-            )
+                "../../data", train=train, download=True, transform=torchvision.transforms.Compose([
+                    torchvision.transforms.ToTensor()]))
         elif mnist_type == "KMNIST":
-            self.dataset = datasets.KMNIST("../../data", train=train, download=True)
+            self.dataset = datasets.KMNIST("../../data", train=train, download=True,
+                                           transform=torchvision.transforms.Compose([
+                                               torchvision.transforms.ToTensor(),
+                                               transforms.Normalize((0.1307,), (0.3081,))]))
 
-        self.data = self.dataset.data
         self.base_transform = transforms.ToTensor()
         self.a_transform = transforms.Compose(
             [
-                transforms.ToTensor(),  # first, convert image to PyTorch tensor
-                transforms.ToPILImage(),
+                torchvision.transforms.RandomRotation((-45, 45))
+            ]
+        )
+        self.a_transform = transforms.Compose(
+            [
+                torchvision.transforms.RandomRotation((-45, 45))
             ]
         )
         self.b_transform = transforms.Compose(
             [
-                transforms.ToTensor(),
                 transforms.Lambda(_add_mnist_noise),
                 transforms.Lambda(self.__threshold_func__),
             ]
@@ -108,8 +114,7 @@ class Noisy_MNIST_Dataset(Dataset):
         self.filtered_classes = []
         self.filtered_nums = []
         for i in range(10):
-            self.filtered_classes.append(self.data[self.targets == i])
-            self.filtered_nums.append(self.filtered_classes[i].shape[0])
+            self.filtered_nums.append(np.where(self.targets == i)[0])
         self.flatten = flatten
 
     def __threshold_func__(self, x):
@@ -117,28 +122,18 @@ class Noisy_MNIST_Dataset(Dataset):
         return x
 
     def __len__(self):
-        return len(self.data)
+        return len(self.dataset)
 
     def __getitem__(self, idx):
-        x_a = self.a_transform(self.data[idx].numpy() / 255)
[…]
