+++ examples/plot_many_views.py
@@ -23,10 +23,9 @@ cv = 3
     n, view_features=[p, q, r], latent_dims=latent_dims, correlation=[0.9]
 )
 
-"""
-Eigendecomposition-Based Methods
----------------------------------
-"""
+# %%
+# Eigendecomposition-Based Methods
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 # %%
 mcca = MCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
@@ -34,10 +33,9 @@ mcca = MCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 # %%
 gcca = GCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 
-"""
-We can also use kernel versions of these methods
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-"""
+# %%
+# We can also use kernel versions of these methods
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 # %%
 kcca = KCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
@@ -45,10 +43,10 @@ kcca = KCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 # %%
 kgcca = KGCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 
-"""
-Higher order correlation methods
----------------------------------
-"""
+# %%
+# Higher order correlation methods
+# ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
 
 # %%
 # Tensor CCA finds higher order correlations so scores are not comparable (but TCCA is equivalent for 2 views)
@@ -57,10 +55,10 @@ tcca = TCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 # %%
 ktcca = KTCCA(latent_dims=latent_dims).fit((X, Y, X)).score((X, Y, Z))
 
-"""
-Iterative Methods
----------------------
+# %%
+# Iterative Methods
+# ^^^^^^^^^^^^^^^^^^^^^^
+#
+# Most of the iterative methods can also use multiple views e.g.
 
-Most of the iterative methods can also use multiple views e.g.
-"""
 pmd = PMD(latent_dims=latent_dims, c=1).fit((X, Y, X)).score((X, Y, Z))
