### 2021-11-18 07e2640
Touching up examples
