### 2021-11-18 jameschapman19/cca_zoo#75 @jameschapman19
This has been done in the latest documentation on the main branch. I'm waiting till I've completed all of these before updating the version.

Really nice change thanks to the reviewers for the suggestion. Didn't realise sphinx-gallery could be done in notebook style!
