### 2021-11-18 jameschapman19/cca_zoo#77 @jameschapman19
As I mentioned in #75 I have followed all reviewers suggestions and put tutorials in sphinx-gallery so they are tested (and look great!)

Thanks again for the suggestion and something new I have learnt.
