### 2021-11-18 jameschapman19/cca_zoo#74 @jameschapman19
Made a start on this in the latest docs. Still a bit of work to do on an intro to methods which would be a nice gateway to the package (and also a selling point).
