### 2021-11-18 2de67f6
Update .scrutinizer.yml
