+++ .scrutinizer.yml
@@ -5,6 +5,11 @@ checks:
 
 build:
   nodes:
+    python:
+        version: "3.7.1"
+
+        # Whether to setup a virtualenv in the repository root directory.
+        virtualenv: true
     analysis:
       tests:
         override:
