### 2021-11-18 a93fe8a
Update .scrutinizer.yml
