### 2021-11-18 45f20d9
Delete .scrutinizer.yml
