+++ .scrutinizer.yml
@@ -1,14 +1,0 @@
-checks:
-    python:
-        version: "3.7.1"
-        # Whether to setup a virtualenv in the repository root directory.
-        virtualenv: true
-        code_rating: true
-        duplicate_code: true
-
-build:
-  nodes:
-    analysis:
-      tests:
-        override:
-          - py-scrutinizer-run
