+++ ccagame/pls/__init__.py
@@ -1,1 +1,4 @@
-
+from .game import Game
+from .msg import MSG
+from .oja import Oja
+from .stochasticpower import StochasticPower
