+++ ccagame/pls/experiments/__init__.py
@@ -1,2 +1,0 @@
-from .plsexperiment import PLSExperiment
-from .game import Game
