+++ ccagame/pca/experiments/__init__.py
@@ -1,3 +1,0 @@
-from .pcaexperiment import PCAExperiment
-from .game import Game
-from .oja import Oja
