### 2021-11-18 7a129fd
setting up mnist and xrmb experiments
