+++ ccagame/pca/__init__.py
@@ -1,3 +1,4 @@
+from .pcaexperiment import PCAExperiment
 from .game import Game
 from .gha import GHA
 from .oja import Oja

[…]
