+++ ccagame/pls/oja.py
@@ -25,7 +25,6 @@ class Oja(PLSExperiment):
             n_components=n_components,
             data=data,
             batch_size=batch_size,
-            **kwargs
         )
         """Constructs the experiment.
         Args:
