+++ experiments/pls/mnist/train.py
@@ -17,7 +17,7 @@ import jax.numpy as jnp
 DEVICES = 4
 N_COMPONENTS = 4
 BATCH_SIZE = None
-LEARNING_RATE = 1e-3
+LEARNING_RATE = 1e-7
 MODEL = "game"
 # This is used to turn name of model on command line into model class
 MODEL_DICT = {
