### 2021-11-19 ba11ea4
Starting to set up some logging stuff
