+++ ccagame/utils.py
@@ -1,9 +1,45 @@
 import numbers
-
+import logging
 import numpy as np
 from jax._src import prng
 from jax._src.random import PRNGKey
-from scipy.io import loadmat
+import os
+
+log = logging.getLogger(__name__)
+
+def _get_next_version():
+        root_dir = os.getcwd()
+
+        try:
+            listdir_info = os.listdir(root_dir)
+        except OSError:
+            log.warning("Missing logger folder: %s", root_dir)
+            return 0
+
+        existing_versions = []
+        for listing in listdir_info:
+            if os.path.isdir(listing) and listing.startswith("version_"):
+                dir_ver = listing.split("_")[1].replace("/", "")
+                existing_versions.append(int(dir_ver))
+        if len(existing_versions) == 0:
+            return 0
+        return max(existing_versions) + 1
+
+
+def log_dir(version=None) -> str:
+    """The directory for this run's tensorboard checkpoint.
+    By default, it is named ``'version_${self.version}'`` but it can be overridden by passing a string value for the
+    constructor's version parameter instead of ``None`` or an int.
+    """
+    if version is None:
+        version=_get_next_version()
+    # create a pseudo standard path ala test-tube
+    version = version if isinstance(version, str) else f"version_{version}"
+    log_dir = os.path.join(os.getcwd(), version)
+    log_dir = os.path.expandvars(log_dir)
+    log_dir = os.path.expanduser(log_dir)
+    os.mkdir(log_dir)
+    return log_dir
 
 
 def get_num_batches(X, Y=None, batch_size=None):
