+++ ccagame/pls/game.py
@@ -18,7 +18,7 @@ class Game(PLSExperiment):
         learning_rate=1e-3,
         momentum=0.9,
         nesterov=True,
-        batch_size=None,
+        batch_size=0,
         **kwargs
     ):
         super(Game, self).__init__(
