+++ ccagame/pca/oja.py
@@ -41,4 +41,3 @@ class Oja(PCAExperiment):
     def _update(self, inputs, global_step):
         self._V = self._V @ inputs.T @ inputs
         self._V = (jnp.linalg.qr(self._V.T)[0]).T
-        return self._V

[…]
