+++ experiments/pls/mnist/train.py
@@ -11,10 +11,11 @@ import jax.numpy as jnp
 import os
 from experiments import parse_args, get_config
 
+from jax.config import config
+config.update('jax_disable_jit', True)
+
 # Right so basically this should run from command line/bash script
 # mnist.py --cores 4 --n_components 4 --batch_size 16 --lr 0.001 --model game
-DEVICES = 1
-MODEL = "msg"
 MODEL_DICT = {
     "game": pls.Game,
     "msg": pls.MSG,
@@ -50,4 +51,5 @@ if __name__ == "__main__":
     flags.mark_flag_as_required("config")
     #magic function which does what pytorch-lightning does which is to make a new numbered version in the directory for each run
     os.chdir(log_dir())
-    app.run(functools.partial(platform.main, MODEL_DICT['oja']))
+    #TODO THIS IS CURRENTLY A HACK WHI
+    app.run(functools.partial(platform.main, MODEL_DICT['game']))
