+++ experiments/experiment_config.py
@@ -4,7 +4,7 @@ import argparse
 # These are the defaults for the above arguments
 DEVICES = 1
 N_COMPONENTS = 4
-LEARNING_RATE = 1e-5
+LEARNING_RATE = 1e-7
 MODEL = "game"
 BATCH_SIZE = 0
 # This is used to turn name of model on command line into model class
