### 2021-11-19 80fe910
problem with vmap gradient bit
