+++ ccagame/pls/__init__.py
@@ -3,3 +3,4 @@ from .game import Game
 from .msg import MSG
 from .oja import Oja
 from .stochasticpower import StochasticPower
+from .incremental import Incremental
