+++ ccagame/pls/incremental.py
@@ -34,7 +34,9 @@ class Incremental(PLSExperiment):
         """
         """Initialization function for a Jaxline experiment."""
         self._U = jax.random.normal(self.local_rng, (self.n_components, dims[0]))
+        self._U=(1/jnp.linalg.norm(self._U,axis=1)*self._U.T).T
         self._V = jax.random.normal(self.local_rng, (self.n_components, dims[1]))
+        self._V=(1/jnp.linalg.norm(self._V,axis=1)*self._V.T).T
         self._S = jax.random.normal(self.local_rng, (self.n_components,))
 
     def _update(self, views, global_step):
