+++ ccagame/pls/oja.py
@@ -3,7 +3,8 @@ from os import environ
 import jax
 import jax.numpy as jnp
 from . import PLSExperiment
-
+from functools import partial
+from jax import jit
 
 class Oja(PLSExperiment):
     def __init__(
@@ -36,6 +37,7 @@ class Oja(PLSExperiment):
         self._U = jax.random.normal(self.local_rng, (self.n_components, dims[0]))
         self._V = jax.random.normal(self.local_rng, (self.n_components, dims[1]))
 
+    @partial(jit, static_argnums=(0))
     def _update(self, views, global_step):
         X_i, Y_i = views
         C = X_i.T @ Y_i
