+++ ccagame/pls/msg.py
@@ -3,7 +3,8 @@ import jax
 import jax.numpy as jnp
 import optax
 from . import PLSExperiment
-
+from functools import partial
+from jax import jit
 
 
 
@@ -44,6 +45,7 @@ class MSG(PLSExperiment):
         self._optimizer =  optax.sgd(learning_rate=learning_rate, momentum=momentum, nesterov=nesterov)
         self._opt_state = self._optimizer.init(self._M)
 
+    @partial(jit, static_argnums=(0))
     def _update(self, views, global_step):
         X_i, Y_i = views
         self._M=self._U.T@self._V
