+++ ccagame/pca/oja.py
@@ -3,7 +3,8 @@ from os import environ
 import jax.numpy as jnp
 import jax
 from . import PCAExperiment
-
+from functools import partial
+from jax import jit
 
 class Oja(PCAExperiment):
     def __init__(
@@ -38,6 +39,7 @@ class Oja(PCAExperiment):
         """
         self._V = jax.random.normal(self.local_rng, (self.n_components, dims))
 
+    @partial(jit, static_argnums=(0))
     def _update(self, inputs, global_step):
         self._V = self._V @ inputs.T @ inputs
         self._V = (jnp.linalg.qr(self._V.T)[0]).T
