+++ experiments/experiment_config.py
@@ -6,7 +6,7 @@ DEVICES = 1
 N_COMPONENTS = 16
 LEARNING_RATE = 1e-2
 MODEL = "game"
-BATCH_SIZE = 64
+BATCH_SIZE = 1
 # This is used to turn name of model on command line into model class
 TRAINING_STEPS = 1000
 
