+++ ccagame/cca/game.py
@@ -1,0 +1,136 @@
+from functools import partial
+from os import environ
+import jax
+import jax.numpy as jnp
+import numpy as np
+import optax
+from jax import jit
+from . import CCAExperiment
+
+class Game(CCAExperiment):
+    def __init__(
+        self,
+        mode,
+        init_rng=None,
+        num_devices=1,
+        n_components=1,
+        dims=None,
+        data=None,
+        learning_rate=1e-6,
+        momentum=0.9,
+        nesterov=True,
+        batch_size=0,
+        **kwargs
+    ):
+        super(Game, self).__init__(
+            mode,
+            init_rng=init_rng,
+            num_devices=num_devices,
+            dims=dims,
+            n_components=n_components,
+            data=data,
+            batch_size=batch_size,
+            **kwargs
+        )
+        """Constructs the experiment.
+        Args:
+          mode: A string, equivalent to FLAGS.jaxline_mode when running normally.
+          init_rng: A `PRNGKey` to use for experiment initialization.
+        """
+        """Initialization function for a Jaxline experiment."""
+        k_per_device=int(n_components/num_devices)
+        weights = np.eye(self.n_components) * 2 - np.ones((self.n_components, self.n_components))
+        weights[np.triu_indices(self.n_components, 1)] = 0
+        self._weights = jnp.reshape(weights, [num_devices, k_per_device, self.n_components])
+        # generates a key for each device
+        keys = jax.random.split(self.local_rng, num_devices)
+        # generates weights for each component on each device
+        U = jax.pmap(lambda key: jax.random.normal(key, (k_per_device, dims[0])))(keys)
+        V = jax.pmap(lambda key: jax.random.normal(key, (k_per_device, dims[1])))(keys)
+        # normalizes the weights for each component
+        self._U = jax.pmap(lambda U: U / jnp.linalg.norm(U, axis=1, keepdims=True))(U)
+        self._V = jax.pmap(lambda V: V / jnp.linalg.norm(V, axis=1, keepdims=True))(V)
+        # This line parallelizes over data sending different data to each device
+        self._update_with_grads = jax.pmap(
+            jax.vmap(
+            self._update_with_grads,
+            in_axes=(0, 0, 0),
+            ),
+            in_axes=(0, 0, 0),
+            axis_name="i",
+        )
+        # This parallelizes gradient calcs and updates for eigenvectors within a given device
+        self._grads = jax.pmap(
+            jax.vmap(
+            self._grads, 
+            in_axes=(0, 0, 0, None,None,None,None),
+            ),
+            in_axes=(0, 0, 0, None,None,0,0),
+            axis_name="i"
+        )
+        self._optimizer =  optax.sgd(learning_rate=learning_rate, momentum=momentum, nesterov=nesterov)
+        self._opt_state_x = jax.pmap(lambda U: self._optimizer.init(U))(self._U)
+        self._opt_state_y = jax.pmap(lambda V: self._optimizer.init(V))(self._V)
+
+    #@partial(jit, static_argnums=(0))
+    def _update(self, views, global_step):
+        X_i, Y_i = views
+        X_i = jnp.reshape(X_i, (self.num_devices, -1, self.dims[0]))
+        Y_i = jnp.reshape(Y_i, (self.num_devices, -1, self.dims[1]))
+        self._local_U = jnp.reshape(self._U, (self.n_components, self.dims[0]))
+        self._local_V = jnp.reshape(self._V, (self.n_components, self.dims[1]))
[…]
