+++ ccagame/cca/__init__.py
@@ -1,0 +1,2 @@
+from .ccaexperiment import CCAExperiment
+from .game import Game
