+++ ccagame/__init__.py
@@ -1,2 +1,3 @@
 from . import pca
 from . import pls
+from . import cca
