+++ experiments/cca/xrmb/train.py
@@ -1,24 +1,23 @@
-from ccagame import pls
+from ccagame import cca
+from experiments.experiment_config import LEARNING_RATE
 import functools
 from os import environ
 from absl import app, flags
 from ccagame.utils import data_stream, log_dir
-from datasets.xrmb import xrmb
+from datasets.mnist import mnist
 from jaxline_fork import platform
 import jax.numpy as jnp
 import os
 from experiments import parse_args, get_config
+from cca_zoo.models import rCCA
+import numpy as np
+from datasets.xrmb import xrmb
 
 # Right so basically this should run from command line/bash script
 # mnist.py --cores 4 --n_components 4 --batch_size 16 --lr 0.001 --model game
 MODEL_DICT = {
-    "game": pls.Game,
-    "msg": pls.MSG,
-    "oja": pls.Oja,
-    "power": pls.StochasticPower,
-    "incremental":pls.Incremental
+    "game": cca.Game,
 }
-
 
 # TO RUN AN EXPERIMENT YOU HAVE TO TINKER HERE A BIT.
 if __name__ == "__main__":
@@ -26,25 +25,27 @@ if __name__ == "__main__":
     environ["XLA_FLAGS"] = f"--xla_force_host_platform_device_count={args.devices}"
     FLAGS = flags.FLAGS
     X, Y, X_te, Y_te = xrmb()
-    input_data_iterator = data_stream(
-        X, Y=Y, batch_size=args.batch_size
+    input_data_iterator = data_stream(X, Y=Y, batch_size=args.batch_size)
+    cca = rCCA(latent_dims=args.n_components, scale=False, c=0.1).fit(
+        (np.array(X), np.array(Y))
     )
-    correct_U, correct_S, correct_V = jnp.linalg.svd(X.T @ Y)
-    correct_U = correct_U[:, :args.n_components]
-    correct_V = correct_V[:args.n_components, :].T
+    print(f"TV : {cca.score((X,Y))}")
+    correct_U = cca.weights[0]
+    correct_V = cca.weights[1]
     FLAGS.config = get_config(
         input_data_iterator,
         dims=[X.shape[1], Y.shape[1]],
         num_devices=args.devices,
         n_components=args.n_components,
         training_steps=args.training_steps,
-        correct_eigenvectors=[correct_U,correct_V],
+        correct_eigenvectors=[correct_U, correct_V],
         learning_rate=args.learning_rate,
-        batch_size=args.batch_size,
         model=args.model,
-        holdout=[X_te,Y_te]
+        batch_size=args.batch_size,
+        holdout=[X_te, Y_te],
     )
     flags.mark_flag_as_required("config")
-    #magic function which does what pytorch-lightning does which is to make a new numbered version in the directory for each run
+    # magic function which does what pytorch-lightning does which is to make a new numbered version in the directory for each run
     os.chdir(log_dir())
-    app.run(functools.partial(platform.main, MODEL_DICT['incremental']))
+    # TODO THIS IS CURRENTLY A HACK WHI
+    app.run(functools.partial(platform.main, MODEL_DICT["game"]))
