+++ experiments/experiment_config.py
@@ -3,10 +3,10 @@ import argparse
 
 # These are the defaults for the above arguments
 DEVICES = 1
-N_COMPONENTS = 16
-LEARNING_RATE = 1e-2
+N_COMPONENTS = 1
+LEARNING_RATE = 1e-1
 MODEL = "game"
-BATCH_SIZE = 0
+BATCH_SIZE = 10
 # This is used to turn name of model on command line into model class
 TRAINING_STEPS = 1000
 
