### 2021-11-28 7907165
Deflation options missing from some methods
