+++ cca_zoo/models/innerloop.py
@@ -20,11 +20,11 @@ from ..utils.check_values import (
 
 class _InnerLoop:
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol: float = 1e-5,
-        initialization: str = "unregularized",
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol: float = 1e-9,
+            initialization: str = "unregularized",
+            random_state=None,
     ):
         """
         :param max_iter: maximum number of iterations to perform if tol is not reached
@@ -45,7 +45,7 @@ class _InnerLoop:
     def _initialize(self):
         if self.initialization == "random":
             self.scores = np.array(
-                [self.random_state.randn(view.shape[0], 1) for view in self.views]
+                [self.random_state.normal(0, 1, size=(view.shape[0], 1)) for view in self.views]
             )
         elif self.initialization == "uniform":
             self.scores = np.array([np.ones((view.shape[0], 1)) for view in self.views])
@@ -56,15 +56,17 @@ class _InnerLoop:
                     random_state=self.random_state,
                     tol=self.tol,
                 )
-                ._fit(*self.views)
-                .scores
+                    ._fit(*self.views)
+                    .scores
             )
+        elif callable(self.initialization):
+            self.scores = next(self.initialization())
         else:
             raise ValueError("initialize must be random, uniform or unregularized")
         self.scores = (
-            self.scores
-            * np.sqrt(self.n - 1)
-            / np.linalg.norm(self.scores, axis=1)[:, np.newaxis]
+                self.scores
+                * np.sqrt(self.n - 1)
+                / np.linalg.norm(self.scores, axis=1)[:, np.newaxis]
         )
         self.weights = [
             self.random_state.randn(view.shape[1], 1) for view in self.views
@@ -119,11 +121,11 @@ class _InnerLoop:
 
 class PLSInnerLoop(_InnerLoop):
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol=1e-5,
-        initialization: str = "unregularized",
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol=1e-9,
+            initialization: str = "unregularized",
+            random_state=None,
     ):
         super().__init__(
             max_iter=max_iter,
@@ -156,7 +158,7 @@ class PLSInnerLoop(_InnerLoop):
         targets = np.ma.array(self.scores, mask=False)
         targets.mask[view_index] = True
         self.weights[view_index] = (
-            self.views[view_index].T @ targets.sum(axis=0).filled()
+                self.views[view_index].T @ targets.sum(axis=0).filled()
         )
         self.weights[view_index] /= np.linalg.norm(self.weights[view_index])
         self.scores[view_index] = self.views[view_index] @ self.weights[view_index]
@@ -164,8 +166,8 @@ class PLSInnerLoop(_InnerLoop):
     def _early_stop(self) -> bool:
         # Some kind of early stopping
         if all(
-            _cosine_similarity(self.scores[n], self.old_scores[n]) > (1 - self.tol)
-            for n, view in enumerate(self.scores)
+                _cosine_similarity(self.scores[n], self.old_scores[n]) > (1 - self.tol)
+                for n, view in enumerate(self.scores)
         ):
             return True
         else:
@@ -174,13 +176,13 @@ class PLSInnerLoop(_InnerLoop):
 
 class PMDInnerLoop(PLSInnerLoop):
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol=1e-5,
-        initialization: str = "unregularized",
-        c=None,
-        positive=None,
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol=1e-9,
[…]
