### 2021-11-29 4c6361f
fix problem with post_transform for deep models
