+++ ccagame/cca/game.py
@@ -50,8 +50,12 @@ class Game(CCAExperiment):
         # generates a key for each device
         keys = jax.random.split(self.local_rng, num_devices)
         # generates weights for each component on each device
-        self._U = jax.pmap(lambda key: jax.random.normal(key, (k_per_device, dims[0]))/10000)(keys)
-        self._V = jax.pmap(lambda key: jax.random.normal(key, (k_per_device, dims[1]))/10000)(keys)
+        self._U = jax.pmap(
+            lambda key: jax.random.normal(key, (k_per_device, dims[0])) / 100
+        )(keys)
+        self._V = jax.pmap(
+            lambda key: jax.random.normal(key, (k_per_device, dims[1])) / 100
+        )(keys)
         # This line parallelizes over data sending different data to each device
         self._update_with_grads = jax.pmap(
             jax.vmap(
@@ -65,16 +69,17 @@ class Game(CCAExperiment):
         self._grads = jax.pmap(
             jax.vmap(
                 self._grads,
-                in_axes=(0, 0, 0, None, None, None),
+                in_axes=(0, 0, 0, None, None),
             ),
-            in_axes=(0, 0, 0, None, 0, 0),
+            in_axes=(0, 0, 0, None, 0),
             axis_name="i",
         )
         self._optimizer = optax.sgd(
-            learning_rate=learning_rate, momentum=momentum, nesterov=nesterov
+            learning_rate=learning_rate, momentum=0, nesterov=False
         )
         self._opt_state_x = jax.pmap(lambda U: self._optimizer.init(U))(self._U)
         self._opt_state_y = jax.pmap(lambda V: self._optimizer.init(V))(self._V)
+        self.learning_rate = learning_rate
 
     # @partial(jit, static_argnums=(0))
     def _update(self, views, global_step):
@@ -83,12 +88,16 @@ class Game(CCAExperiment):
         Y_i = jnp.reshape(Y_i, (self.num_devices, -1, self.dims[1]))
         self._local_U = jnp.reshape(self._U, (self.n_components, self.dims[0]))
         self._local_V = jnp.reshape(self._V, (self.n_components, self.dims[1]))
-        T = X_i @ self._local_U.T + Y_i @ self._local_V.T
-        T = T / jnp.linalg.norm(T, axis=0)
-        local_T = T
+        #TODO batchnorm views separately like barlowtwins
+        Zx= X_i @ self._local_U.T
+        Zx=Zx / jnp.linalg.norm(Zx, axis=1)
+        Zy= Y_i @ self._local_V.T
+        Zy=Zy / jnp.linalg.norm(Zy, axis=1)
+        T = Zx+Zy
         T = jnp.reshape(T, (self.num_devices, -1, T.shape[1]))
-        grads_x = self._grads(self._U, T, self._weights, self._local_U, local_T, X_i)
-        grads_y = self._grads(self._V, T, self._weights, self._local_V, local_T, Y_i)
+        local_T = jnp.reshape(T, (self.n_components, -1))
+        grads_x = self._grads(self._U, T, self._weights, local_T, X_i)
+        grads_y = self._grads(self._V, T, self._weights, local_T, Y_i)
         self._U, self._opt_state_x = self._update_with_grads(
             self._U, grads_x, self._opt_state_x
         )
@@ -98,17 +107,17 @@ class Game(CCAExperiment):
 
     @staticmethod
     @jit
-    def _grads(ui, ti, weights, U, T, X):
+    def _grads(ui, ti, weights, T, X):
         weights_ij = (jnp.sign(weights + 0.5) - 1.0) / 2.0  # maps -1 to -1 else to 0
-        reward = X.T @ ti
[…]
