+++ experiments/cca/rand/train.py
@@ -9,6 +9,7 @@ from experiments import parse_args, get_config
 from cca_zoo.models import rCCA
 import numpy as np
 import jax.numpy as jnp
+
 # Right so basically this should run from command line/bash script
 # mnist.py --cores 4 --n_components 4 --batch_size 16 --lr 0.001 --model game
 MODEL_DICT = {
@@ -20,19 +21,17 @@ if __name__ == "__main__":
     args = parse_args()
     environ["XLA_FLAGS"] = f"--xla_force_host_platform_device_count={args.devices}"
     FLAGS = flags.FLAGS
-    X = np.random.normal(size=(100, 11))
+    X = np.random.normal(size=(128, 11))
     X -= X.mean(axis=0)
-    X/=X.std(axis=0)
-    Y = np.random.normal(size=(100, 10))
+    X /= X.std(axis=0)
+    Y = np.random.normal(size=(128, 10))
     Y -= Y.mean(axis=0)
-    Y/=Y.std(axis=0)
+    Y /= Y.std(axis=0)
     input_data_iterator = data_stream(X, Y=Y, batch_size=args.batch_size)
     cca = rCCA(latent_dims=args.n_components, scale=False, c=0).fit(
         (np.array(X), np.array(Y))
     )
-    print(
-        f"TV : {cca.score((X,Y))}"
-    )  # jnp.corrcoef((X@correct_U).T,(Y@correct_V).T)[args.n_components:,:args.n_components]
+    print(f"TV : {cca.score((X,Y))}, Sum : {cca.score((X,Y)).sum()}")
     correct_U = cca.weights[0]
     correct_V = cca.weights[1]
     FLAGS.config = get_config(
