+++ experiments/experiment_config.py
@@ -3,11 +3,11 @@ import argparse
 
 # These are the defaults for the above arguments
 DEVICES = 1
-N_COMPONENTS = 5
+N_COMPONENTS = 15
 LEARNING_RATE = 1e-3
 BATCH_SIZE = 64
 # This is used to turn name of model on command line into model class
-TRAINING_STEPS = 5000
+TRAINING_STEPS = 10000
 MODEL='game'
 
 def parse_args():
