+++ ccagame/pls/msg.py
@@ -15,7 +15,6 @@ class MSG(PLSExperiment):
         init_rng=None,
         num_devices=1,
         n_components=1,
-        dims=None,
         data=None,
         learning_rate=1e-3,
         momentum=0.9,
@@ -27,7 +26,6 @@ class MSG(PLSExperiment):
             mode,
             init_rng=init_rng,
             num_devices=num_devices,
-            dims=dims,
             n_components=n_components,
             data=data,
             batch_size=batch_size,
