+++ ccagame/__init__.py
@@ -1,3 +1,4 @@
+from . import utils
 from . import pca
 from . import pls
 from . import cca
