+++ ccagame/pls/online.py
@@ -22,7 +22,6 @@ class Online(PLSExperiment):
         init_rng=None,
         num_devices=1,
         n_components=1,
-        dims=None,
         data=None,
         learning_rate=1e-6,
         momentum=0.9,
@@ -34,7 +33,6 @@ class Online(PLSExperiment):
             mode,
             init_rng=init_rng,
             num_devices=num_devices,
-            dims=dims,
             n_components=n_components,
             data=data,
             batch_size=batch_size,
