### 2021-12-01 1da3cb3
Initialization change for iterative methods
