+++ cca_zoo/models/innerloop.py
@@ -23,7 +23,7 @@ class _InnerLoop:
             self,
             max_iter: int = 100,
             tol: float = 1e-9,
-            initialization: str = "unregularized",
+            initial_scores: np.ndarray = None,
             random_state=None,
     ):
         """
@@ -31,7 +31,7 @@ class _InnerLoop:
         :param tol: tolerance value used for stopping criteria
         :param initialization: initialise the optimisation with either the 'unregularized' (CCA/PLS) solution, or a 'random' initialisation
         """
-        self.initialization = initialization
+        self.scores = initial_scores
         self.max_iter = max_iter
         self.tol = tol
         self.random_state = check_random_state(random_state)
@@ -42,47 +42,13 @@ class _InnerLoop:
         """
         pass
 
-    def _initialize(self):
-        if self.initialization == "random":
-            self.scores = np.array(
-                [
-                    self.random_state.normal(0, 1, size=(view.shape[0], 1))
-                    for view in self.views
-                ]
-            )
-        elif self.initialization == "uniform":
-            self.scores = np.array([np.ones((view.shape[0], 1)) for view in self.views])
-        elif self.initialization == "unregularized":
-            self.scores = (
-                PLSInnerLoop(
-                    initialization="random",
-                    random_state=self.random_state,
-                    tol=self.tol,
-                )
-                    ._fit(*self.views)
-                    .scores
-            )
-        elif callable(self.initialization):
-            self.scores = next(self.initialization())
-        else:
-            raise ValueError("initialize must be random, uniform or unregularized")
-        self.scores = (
-                self.scores
-                * np.sqrt(self.n - 1)
-                / np.linalg.norm(self.scores, axis=1)[:, np.newaxis]
-        )
-        self.weights = [
-            self.random_state.randn(view.shape[1], 1) for view in self.views
-        ]
-
     def _fit(self, *views: np.ndarray):
         self.views = views
         self.n = views[0].shape[0]
 
         # Check that the parameters that have been passed are valid for these views given #views and #features
         self._check_params()
-        self._initialize()
-
+        self.weights = [self.random_state.randn(view.shape[1]) for view in self.views]
         self.track = {}
         self.track["converged"] = False
         # Iterate until convergence
@@ -127,13 +93,13 @@ class PLSInnerLoop(_InnerLoop):
             self,
             max_iter: int = 100,
             tol=1e-9,
-            initialization: str = "unregularized",
+            initial_scores: np.ndarray = None,
             random_state=None,
     ):
         super().__init__(
             max_iter=max_iter,
             tol=tol,
-            initialization=initialization,
+            initial_scores=initial_scores,
             random_state=random_state,
         )
 
@@ -164,7 +130,9 @@ class PLSInnerLoop(_InnerLoop):
                 self.views[view_index].T @ targets.sum(axis=0).filled()
         )
         self.weights[view_index] /= np.linalg.norm(self.weights[view_index])
-        self.scores[view_index] = self.views[view_index] @ self.weights[view_index]
+        self.scores[view_index] = self.views[view_index] @ np.squeeze(
+            np.array(self.weights[view_index])
+        )
 
     def _early_stop(self) -> bool:
         # Some kind of early stopping
@@ -182,7 +150,7 @@ class PMDInnerLoop(PLSInnerLoop):
             self,
             max_iter: int = 100,
             tol=1e-9,
-            initialization: str = "unregularized",
+            initial_scores: np.ndarray = None,
             c=None,
[…]
