### 2021-12-01 79b993a
Initialization change for iterative methods
