### 2021-12-02 8fe4a32
update to run with ukbb data and add config_ukbb.py
