### 2021-12-02 138173f
Initialization change for iterative methods
