+++ cca_zoo/utils/__init__.py
@@ -1,2 +1,3 @@
-from .check_values import *
-from .plotting import plot_latent_train_test, plot_latent_label
+from .check_values import _check_views, _check_batch_size, _check_Parikh2014, _process_parameter, \
+    _check_parameter_number, _check_converged_weights
+from .plotting import plot_latent_train_test, plot_latent_label, cv_plot
