+++ cca_zoo/deepmodels/dtcca.py
@@ -20,7 +20,7 @@ class DTCCA(DCCA):
     """
 
     def __init__(
-        self, latent_dims: int, encoders=None, r: float = 0, eps: float = 1e-3
+            self, latent_dims: int, encoders=None, r: float = 0, eps: float = 1e-3
     ):
         """
 
