### 2021-12-02 6b74684
Initialization change for iterative methods
