+++ cca_zoo/models/__init__.py
@@ -1,3 +1,4 @@
+from .rcca import rCCA, CCA, PLS
 from .gcca import GCCA, KGCCA
 from .iterative import (
     _Iterative,
@@ -13,5 +14,4 @@ from .iterative import (
 )
 from .mcca import MCCA, KCCA
 from .ncca import NCCA
-from .rcca import rCCA, CCA, PLS
 from .tcca import TCCA, KTCCA
