+++ cca_zoo/test/test_models.py
@@ -62,56 +62,56 @@ def test_unregularized_methods():
     assert np.testing.assert_array_almost_equal(corr_cca, corr_kcca, decimal=1) is None
     assert np.testing.assert_array_almost_equal(corr_cca, corr_tcca, decimal=1) is None
     assert (
-            np.testing.assert_array_almost_equal(corr_kgcca, corr_gcca, decimal=1) is None
+        np.testing.assert_array_almost_equal(corr_kgcca, corr_gcca, decimal=1) is None
     )
     # Check standardized models have standard outputs
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(iter.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(iter.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(cca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(cca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(mcca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(mcca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(kcca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(kcca.transform((X, Y))[0], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(iter.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(iter.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(cca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(cca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(mcca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(mcca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
     assert (
-            np.testing.assert_allclose(
-                np.linalg.norm(kcca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
-            )
-            is None
+        np.testing.assert_allclose(
+            np.linalg.norm(kcca.transform((X, Y))[1], axis=0) ** 2, n, rtol=0.2
+        )
+        is None
     )
 
 
@@ -179,7 +179,7 @@ def test_regularized_methods():
     # Check the correlations from each unregularized method are the same
[…]
