+++ cca_zoo/test/test_models.py
@@ -215,8 +215,8 @@ def test_non_negative_methods():
 
 
 def test_sparse_methods():
-    c1 = [0.1, 0.3]
-    c2 = [0.1, 0.3]
+    c1 = [0.1, 0.2]
+    c2 = [0.1, 0.2]
     param_grid = {"c": [c1, c2]}
     pmd_cv = GridSearchCV(PMD(random_state=rng), param_grid=param_grid).fit([X, Y])
     cv_plot(pmd_cv.cv_results_)
@@ -304,10 +304,15 @@ def test_l0():
 
 
 def test_partialcca():
-    #Tests that partial CCA scores are not correlated with confounds.
+    # Tests that partial CCA scores are not correlated with confounds.
     pcca = PartialCCA(latent_dims=3)
     pcca.fit((X, Y), confounds=Z)
-    assert np.allclose(np.corrcoef(pcca.transform((X, Y),confounds=Z)[0], Z, rowvar=False)[:3,:3]-np.eye(3), 0, atol=0.001)
+    assert np.allclose(
+        np.corrcoef(pcca.transform((X, Y), confounds=Z)[0], Z, rowvar=False)[:3, :3]
+        - np.eye(3),
+        0,
+        atol=0.001,
+    )
 
 
 def test_pls():
