+++ cca_zoo/data/simulated.py
@@ -10,16 +10,16 @@ from ..utils import _process_parameter
 
 
 def generate_covariance_data(
-    n: int,
-    view_features: List[int],
-    latent_dims: int = 1,
-    view_sparsity: List[Union[int, float]] = None,
-    correlation: Union[List[float], float] = 1,
-    structure: Union[str, List[str]] = None,
-    sigma: Union[List[float], float] = None,
-    decay: float = 0.5,
-    positive=None,
-    random_state: Union[int, np.random.RandomState] = None,
+        n: int,
+        view_features: List[int],
+        latent_dims: int = 1,
+        view_sparsity: List[Union[int, float]] = None,
+        correlation: Union[List[float], float] = 1,
+        structure: Union[str, List[str]] = None,
+        sigma: Union[List[float], float] = None,
+        decay: float = 0.5,
+        positive=None,
+        random_state: Union[int, np.random.RandomState] = None,
 ):
     """
     Function to generate CCA dataset with defined population correlations
@@ -49,7 +49,8 @@ def generate_covariance_data(
     positive = _process_parameter("positive", positive, False, len(view_features))
     sigma = _process_parameter("sigma", sigma, 0.5, len(view_features))
     completed = False
-    while not completed:
+    attempts = 0
+    while not completed and attempts < 10:
         try:
             mean = np.zeros(sum(view_features))
             if not isinstance(correlation, list):
@@ -58,7 +59,7 @@ def generate_covariance_data(
             covs = []
             true_features = []
             for view_p, sparsity, view_structure, view_positive, view_sigma in zip(
-                view_features, view_sparsity, structure, positive, sigma
+                    view_features, view_sparsity, structure, positive, sigma
             ):
                 # Covariance Bit
                 if view_structure == "identity":
@@ -110,12 +111,12 @@ def generate_covariance_data(
                     # Cross Bit
                     cross += covs[i] @ A @ covs[j]
                 cov[
-                    splits[i] : splits[i] + view_features[i],
-                    splits[j] : splits[j] + view_features[j],
+                splits[i]: splits[i] + view_features[i],
+                splits[j]: splits[j] + view_features[j],
                 ] = cross
                 cov[
-                    splits[j] : splits[j] + view_features[j],
-                    splits[i] : splits[i] + view_features[i],
+                splits[j]: splits[j] + view_features[j],
+                splits[i]: splits[i] + view_features[i],
                 ] = cross.T
 
             X = np.zeros((n, sum(view_features)))
@@ -126,16 +127,19 @@ def generate_covariance_data(
             completed = True
         except:
             completed = False
+            attempts += 1
+    if not completed:
+        raise ValueError('Could not generate positive definite covariance matrix for supplied parameters.')
     return views, true_features
 
 
 def generate_simple_data(
-    n: int,
-    view_features: List[int],
-    view_sparsity: List[Union[int, float]] = None,
-    eps: float = 0,
-    transform=True,
-    random_state=None,
+        n: int,
+        view_features: List[int],
+        view_sparsity: List[Union[int, float]] = None,
+        eps: float = 0,
+        transform=False,
+        random_state=None,
 ):
     """
     Simple latent variable model to generate data with one latent factor
@@ -161,11 +165,11 @@ def generate_simple_data(
         "view_sparsity", view_sparsity, 0, len(view_features)
     )
     for p, sparsity in zip(view_features, view_sparsity):
-        weights = random_state.randn(p, 1)
+        weights = random_state.normal(size=(p, 1))
[…]
