+++ cca_zoo/test/test_models.py
@@ -330,17 +330,17 @@ def test_VCCA():
     numpyro = pytest.importorskip("numpyro")
     from cca_zoo.probabilisticmodels import VariationalCCA
     from cca_zoo.data import generate_simple_data
-
+    np.random.seed(0)
     # Tests tensor CCA methods
     (X, Y), (_) = generate_covariance_data(20, [5, 5])
     latent_dims = 1
     cca = CCA(latent_dims=latent_dims).fit([X, Y])
     vcca = VariationalCCA(latent_dims=latent_dims, num_warmup=1000, num_samples=1000).fit([X, Y])
-    # Test that vanilla CCA and VCCA produce roughly similar latent space
+    # Test that vanilla CCA and VCCA produce roughly similar latent space ie they are correlated
     assert (
-        np.corrcoef(
+        np.abs(np.corrcoef(
             cca.transform([X, Y])[1].T,
             vcca.posterior_samples["z"].mean(axis=0)[:, 0],
-        )[0, 1]
+        )[0, 1])
         > 0.9
     )
