+++ docs/source/api/models.rst
@@ -131,6 +131,18 @@ Sparse CCA by ADMM
 Miscellaneous
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Nonparametric CCA
+"""""""""""""""""""""""""""""""""""""""""""
+.. autoclass:: cca_zoo.models.NCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
+Partial CCA
+"""""""""""""""""""""""""""""""""""""""""""
+.. autoclass:: cca_zoo.models.PartialCCA
+    :inherited-members:
+    :exclude-members: get_params, set_params
+
 Sparse Weighted CCA
 """""""""""""""""""""""""""""""""""""""""""
 .. autoclass:: cca_zoo.models.SWCCA
