### 2021-12-03 jameschapman19/cca_zoo#76 @jameschapman19
Satisfied with this too. Tests still take a little while but I think they are now minimal and important (in the sense of checking crucial things ran) whereas before they were neither minimal nor all important
