### 2021-12-03 d764b83
Speed up unregularized als by not pushing through elastic solver
