+++ cca_zoo/models/_innerloop.py
@@ -6,7 +6,7 @@ import numpy as np
 from sklearn.exceptions import ConvergenceWarning
 from sklearn.linear_model import (
     ElasticNet,
-    SGDRegressor,
+    SGDRegressor, LinearRegression,
 )
 from sklearn.utils._testing import ignore_warnings
 from sklearn.utils.validation import check_random_state
@@ -20,10 +20,10 @@ from cca_zoo.utils.check_values import (
 
 class _InnerLoop:
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol: float = 1e-9,
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol: float = 1e-9,
+            random_state=None,
     ):
         """
         :param max_iter: maximum number of iterations to perform if tol is not reached
@@ -86,10 +86,10 @@ class _InnerLoop:
 
 class _PLSInnerLoop(_InnerLoop):
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol=1e-9,
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol=1e-9,
+            random_state=None,
     ):
         super().__init__(
             max_iter=max_iter,
@@ -121,7 +121,7 @@ class _PLSInnerLoop(_InnerLoop):
         targets = np.ma.array(self.scores, mask=False)
         targets.mask[view_index] = True
         self.weights[view_index] = (
-            self.views[view_index].T @ targets.sum(axis=0).filled()
+                self.views[view_index].T @ targets.sum(axis=0).filled()
         )
         self.weights[view_index] /= np.linalg.norm(self.weights[view_index])
         self.scores[view_index] = self.views[view_index] @ np.squeeze(
@@ -131,8 +131,8 @@ class _PLSInnerLoop(_InnerLoop):
     def _early_stop(self) -> bool:
         # Some kind of early stopping
         if all(
-            _cosine_similarity(self.scores[n], self.old_scores[n]) > (1 - self.tol)
-            for n, view in enumerate(self.scores)
+                _cosine_similarity(self.scores[n], self.old_scores[n]) > (1 - self.tol)
+                for n, view in enumerate(self.scores)
         ):
             return True
         else:
@@ -141,12 +141,12 @@ class _PLSInnerLoop(_InnerLoop):
 
 class _PMDInnerLoop(_PLSInnerLoop):
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol=1e-9,
-        c=None,
-        positive=None,
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol=1e-9,
+            c=None,
+            positive=None,
+            random_state=None,
     ):
         super().__init__(
             max_iter=max_iter,
@@ -182,7 +182,7 @@ class _PMDInnerLoop(_PLSInnerLoop):
         targets = np.ma.array(self.scores, mask=False)
         targets.mask[view_index] = True
         self.weights[view_index] = (
-            self.views[view_index].T @ targets.sum(axis=0).filled()
+                self.views[view_index].T @ targets.sum(axis=0).filled()
         )
         self.weights[view_index] = _delta_search(
             self.weights[view_index],
@@ -196,11 +196,11 @@ class _PMDInnerLoop(_PLSInnerLoop):
 
 class _ParkhomenkoInnerLoop(_PLSInnerLoop):
     def __init__(
-        self,
-        max_iter: int = 100,
-        tol=1e-9,
-        c=None,
-        random_state=None,
+            self,
+            max_iter: int = 100,
+            tol=1e-9,
+            c=None,
+            random_state=None,
     ):
         super().__init__(
[…]
