### 2021-12-03 c9ff5e6
Format read me
make the landing page fitting markdown syntax
