+++ README.md
@@ -5,22 +5,35 @@
 [![version](https://img.shields.io/pypi/v/cca-zoo)](https://pypi.org/project/cca-zoo/)
 [![downloads](https://img.shields.io/pypi/dm/cca-zoo)](https://pypi.org/project/cca-zoo/)
 
-# Installation
-Note: for standard install use: 
-pip install cca-zoo
+# CCA zoo
 
+`cca-zoo` is a collection of linear, kernel, and deep methods for canonical correlation analysis of multiview data. 
+Where possible I have followed the `scikit-learn`/`mvlearn` APIs and models therefore have `fit`/`transform`/`fit_transform` methods as standard.
+
+## Installation
+
+Dependency of some implemented algorithms are heavy, such as `pytorch` and `numpyro`. 
+We provide several options to accomandate the user's needs.
+For full details of algoritms inclused, please refere to section [Implemented Methods](#implemented-methods)
+
+Standard installation: 
+
+```
+pip install cca-zoo
+```
 For deep learning elements use:
+```
 pip install cca-zoo[deep]
+```
 
 For probabilistic elements use:
+```
 pip install cca-zoo[probabilistic]
-
-This means that there is no need to install the large pytorch package or numpyro to run cca-zoo unless you wish to use deep learning
-
-# Documentation
+```
+## Documentation
 Available at https://cca-zoo.readthedocs.io/en/latest/
   
-# Citation:
+## Citation:
 If this repository was helpful to you please do give a star.
 
 In case this work is used as part of research I attach a DOI bibtex entry:
@@ -38,58 +51,47 @@ In case this work is used as part of research I attach a DOI bibtex entry:
 }
 ```
 
-# Implemented Methods
-## Standard Install
-### CCA (Canonical Correlation Analysis)
-Solutions based on either alternating least squares or as the solution to genrralized eigenvalue problem
-### PLS (Partial Least Squares)
+## Implemented Methods
 
-### rCCA (Ridge Regularized Canonical Correlation Analysis)
-https://www.sciencedirect.com/science/article/abs/pii/0304407676900105?via%3Dihub
-### GCCA (Generalized CCA)  :
-https://academic.oup.com/biomet/article-abstract/58/3/433/233349?redirectedFrom=fulltext
-### MCCA (Multiset CCA) : 
-### K(M)CCA (kernel Multiset CCA) : 
-### TCCA (Tensor CCA) :
-https://arxiv.org/pdf/1502.02330.pdf
-### KTCCA (kernel Tensor CCA) :
-https://arxiv.org/pdf/1502.02330.pdf
-### SCCA (Sparse CCA) :
-https://onlinelibrary.wiley.com/doi/abs/10.1111/biom.13043
-### SPLS (Sparse PLS/Penalized Matrix Decomposition)  :
-https://web.stanford.edu/~hastie/Papers/PMD_Witten.pdf
-### ElasticCCA (Penalized CCA) :
-https://pubmed.ncbi.nlm.nih.gov/19689958/
-### SWCCA (Sparse Weighted CCA) : 
-https://arxiv.org/abs/1710.04792v1#:~:text=However%2C%20classical%20and%20sparse%20CCA%20models%20consider%20the,where%20weights%20are%20used%20for%20regularizing%20different%20samples.
-### SpanCCA
-http://akyrillidis.github.io/pubs/Conferences/cca.pdf
+### Standard Install
+- CCA (Canonical Correlation Analysis): Solutions based on either alternating least squares or as the solution to genrralized eigenvalue problem
+- PLS (Partial Least Squares)
[…]
