### 2021-12-03 jameschapman19/cca_zoo#76 @emdupre
This is great, thank you @jameschapman19 ! Would you say that with closing this you've addressed all of the reviewer feedback ? I see that #122 is open and currently failing tests, so I'm not sure if you were planning to work on that first.

Please just ping into the JOSS issue whenever you're ready, and I'll make sure that this gets to the reviewers !
