### 2021-12-03 8e7c1e8
Adding a small amount of regularisation to CCA by alternating least squares.
Fix iterative tests by adding random seeds.
