+++ requirements/basic.txt
@@ -1,5 +1,5 @@
 numpy
-scikit-learn>=0.23
+scikit-learn>=1.0
 mvlearn
 scipy>=1.7
 matplotlib

[…]
