### 2021-12-03 aa9e12b
Removing main from partial cca
