### 2021-12-03 #3823 @jameschapman19
@emdupre closed that final issue. Thanks for all of the suggestions I hope that they have all been addressed in some way.
