### 2021-12-04 9f96a5a
adding a template bash script for wanbd running
