### 2021-12-06 7f9cb7e
Remove lightning module loss method. Never used and not required
