### 2021-12-06 #3823 @emdupre
Thank you, @jameschapman19 !

@hugorichard and @robbisg I want to check with you explicitly since you have a few unchecked items on your reviewer checklist : do the provided revisions (as discussed in the cross-linked issues) address your previous concerns ? Of course, @ejolly, please also let us know if there's anything that you don't consider resolved !
