+++ cluster/wandb_agent.sh
@@ -7,7 +7,7 @@
 #$ -wd /home/jchapman/projects/barlowtwins
 hostname
 date
-source /home/jchapmanbarbano/jaxsourcefile # source
+source /home/jchapman/jaxsourcefile # source
 source /home/jchapman/venvs/ccagame/bin/activate
 wandb login --relogin f6437b9357053a0a34efc60cbfaa03087e8984f9
 wandb agent blah
