+++ experiments/pls/sweep.yaml
@@ -1,4 +1,4 @@
-program: train.py
+program: experiments/pls/train.py
 method: grid
 parameters:
   model:
