### 2021-12-07 439674b
tweaks to get wandb outputs nicely
