+++ experiments/pls/sweep.yaml
@@ -2,12 +2,12 @@ program: experiments/pls/train.py
 method: grid
 parameters:
   model:
-    values: ['game','msg','oja', 'power']
+    values: ['game','oja', 'power']
   config.data:
     values: ['mnist']
   config.training_steps:
-    values: [5000]
+    values: [2000]
   config.learning_rate:
-    values: [1e-3,1e-4,1e-5,1e-6]
+    values: [1e-1,1e-2,1e-3, 1e-4, 1e-5]
   config.batch_size:
-    values: [1024,512,256]
+    values: [256, 512, 1024]

[…]
