### 2021-12-07 1d44b53
Change sqrt function for deep models to a more stable version (and associated epsilons)
Fix API problem with GridSearchCV!
