+++ experiments/cca/config.py
@@ -3,18 +3,18 @@ from ml_collections import config_dict
 
 
 def get_config() -> config_dict.ConfigDict:
-    #get the basic jax config
+    # get the basic jax config
     config = get_base_config()
 
-    #these are given by wandb
-    config.learning_rate=1e-3
-    config.num_devices=1
-    config.n_components=16
-    config.batch_size=0
-    config.training_steps = 100
-    config.data='mnist'
+    # these are given by wandb
+    config.learning_rate = 1e-5
+    config.num_devices = 1
+    config.n_components = 4
+    config.batch_size = 0
+    config.training_steps = 1000
+    config.data = "mnist"
 
-    #defaults
+    # defaults
     config.checkpoint_dir = "jaxlog"
     config.train_checkpoint_all_hosts = True
     config.log_tensors_interval = 1
