+++ experiments/pls/config.py
@@ -7,7 +7,7 @@ def get_config() -> config_dict.ConfigDict:
     config = get_base_config()
 
     #these are given by wandb
-    config.learning_rate=1e-3
+    config.learning_rate=1e-2
     config.num_devices=1
     config.n_components=16
     config.batch_size=0
