+++ ccagame/cca/ccaexperiment.py
@@ -1,15 +1,14 @@
 from abc import abstractmethod
 from typing import Optional
-from jax._src.numpy.lax_numpy import zeros_like
 import jax.numpy as jnp
 from ccagame.baseexperiment import BaseExperiment
 from jaxline import utils
 from jax import jit
 from functools import partial
-
-from ccagame.utils import data_stream
 from datasets.mnist import mnist_iterator
+from datasets.ukbiobank import ukbb_iterator
 from datasets.xrmb import xrmb_iterator
+
 
 class CCAExperiment(BaseExperiment):
     def __init__(
@@ -23,14 +22,36 @@ class CCAExperiment(BaseExperiment):
         batch_size=0,
         correct_eigenvectors=None,
         holdout=None,
+        path=None,
         **kwargs,
     ):
+        if data == "mnist":
+            (
+                self.data,
+                self.holdout,
+                self.correct_eigenvectors,
+                self.dims,
+            ) = mnist_iterator(batch_size=batch_size, n_components=n_components, cca=True)
+        elif data == "xrmb":
+            (
+                self.data,
+                self.holdout,
+                self.correct_eigenvectors,
+                self.dims,
+            ) = xrmb_iterator(batch_size=batch_size, n_components=n_components, cca=True)
+        elif data == "ukbb":
+            self.data, self.holdout, self.dims = ukbb_iterator(
+                path, batch_size=batch_size
+            )
+            self.correct_eigenvectors = None
+        else:
+            raise ValueError("Data {data} not implemented yet")
         super(CCAExperiment, self).__init__(
             mode=mode,
             init_rng=init_rng,
             num_devices=num_devices,
             n_components=n_components,
-            data=data,
+            data=self.data,
             batch_size=batch_size,
         )
         """Constructs the experiment.
@@ -39,29 +60,29 @@ class CCAExperiment(BaseExperiment):
           init_rng: A `PRNGKey` to use for experiment initialization.
         """
         """Initialization function for a Jaxline experiment."""
-        self.dims = dims
-        self.correct_eigenvectors = correct_eigenvectors
-        if data=='mnist':
-            self.data,self.holdout=mnist_iterator(batch_size=batch_size)
-        elif data=='xrmb':
-            self.data,self.holdout=xrmb_iterator(batch_size=batch_size)
-        else:
-            raise ValueError('Data {data} not implemented yet')
 
     @abstractmethod
     def _update(self, views, global_step):
         raise NotImplementedError
 
-    #@partial(jit, static_argnums=(0))
+    # @partial(jit, static_argnums=(0))
     def _get_scalars(self):
         U = jnp.reshape(self._U, (self.n_components, self.dims[0]))
         V = jnp.reshape(self._V, (self.n_components, self.dims[1]))
         return {
-            "TC": self._TC(U,V),
-            "correct x":self._correct_eigenvector_streak(U, self.correct_eigenvectors[0],self.holdout[0]),
-            "correct y":self._correct_eigenvector_streak(V, self.correct_eigenvectors[1],self.holdout[1]),
-            "subspace x":self._normalized_subspace_distance(U, self.correct_eigenvectors[0]),
-            "subspace y":self._normalized_subspace_distance(V, self.correct_eigenvectors[1]),
+            "TC": self._TC(U, V),
+            "correct x": self._correct_eigenvector_streak(
+                U, self.correct_eigenvectors[0], self.holdout[0]
+            ),
[…]

[…]
