### 2021-12-13 #3823 @emdupre
Thank you @ejolly @hugorichard @robbisg for your thoughtful reviews and @jameschapman19 for your impressive work on `cca-zoo` !

At this point, @jameschapman19, could you please:

- [ ] Make a tagged release of your software, and list the version tag of the archived version here. I see that you released version `v1.10.6` 5 days ago and have made no changes to `main` since; please let me know if you'd like to use that as the  JOSS-associated release.
- [ ] Check the archival deposit (e.g., in Zenodo) has the correct metadata (e.g., the JOSS paper and software archive have the same title and author listing)
- [ ] List the DOI of the archived version here.

I can then move forward with accepting the submission :rocket:
