### 2021-12-14 70a598b
Excited to have fixed DTCCA.
Simplified CCA_Dataset (removed labels since it wasn't used - this is a case where user can customise for themselves).

Fixed get_dataloaders to allow a bit more flexibility.
