+++ cca_zoo/__init__.py
@@ -1,6 +1,7 @@
 import cca_zoo.data
 import cca_zoo.models
 import cca_zoo.model_selection
+
 try:
     import cca_zoo.deepmodels
 except ModuleNotFoundError:

[…]
