### 2021-12-14 #3823 @jameschapman19
Thanks all.

- I've made a tiny fix so please go with v1.10.7
- It is archived in Zenodo.
- Hao-Ting and I are listed as authors on both the JOSS paper and software archive.
- The DOI is 10.5281/zenodo.5781712 and can be viewed at https://zenodo.org/record/5781712
