### 2021-12-15 #3823 @emdupre
Thank you, @jameschapman19 !

Could you also please update the Zenodo archive title so it matches the title of the JOSS paper ? Specifically, "cca-zoo : a python package for implementing models from the canonical correlation analysis family".
