### 2021-12-15 #3823 @kyleniemeyer
Hi @jameschapman19, I'm just doing some final checks before accepting.

Can you capitalize "python" in the paper's title? There are some uses throughout the paper where it should be capitalized as well.

Also, the Bach and Jordan reference needs a few more details. This is a technical report, published by University of California Berkeley Department of Statistics, Tech. Rep. 688. You can also add the URL https://statistics.berkeley.edu/sites/default/files/tech-reports/688.pdf.

Lastly, please confirm that there are no DOIs available for the Wenwen et al. and Wong et al. references.
