### 2021-12-15 #3823 @emdupre
Thank you for catching those points, @kyleniemeyer !

> Can you capitalize "python" in the paper's title? There are some uses throughout the paper where it should be capitalized as well.

@jameschapman19, when you're making those changes please update it in the zenodo archive as well, just so those two documents match !
