+++ docs/joss/paper.bib
@@ -24,7 +24,8 @@
   title={A probabilistic interpretation of canonical correlation analysis},
   author={Bach, Francis R and Jordan, Michael I},
   year={2005},
-  publisher={Technical Report 688, Department of Statistics, University of California~…}
+  url={https://statistics.berkeley.edu/sites/default/files/tech-reports/688.pdf}
+  publisher={Technical Report 688, Department of Statistics, University of California Berkeley}
 }
 
 @article{hardoon2004canonical,
@@ -114,7 +115,8 @@
   number={3},
   pages={459--466},
   year={2018},
-  publisher={IET}
+  publisher={IET},
+  doi={doi={10.1049/cje.2017.08.004}}
 }
 
 @inproceedings{andrew2013deep,
@@ -214,17 +216,6 @@
   pages={1148--1157},
   year={2016},
   organization={PMLR}
-}
-
-@article{wenwen2018sparse,
-  title={Sparse Weighted Canonical Correlation Analysis},
-  author={Wenwen, MIN and Juan, LIU and Zhang, Shihua},
-  journal={Chinese Journal of Electronics},
-  volume={27},
-  number={3},
-  pages={459--466},
-  year={2018},
-  publisher={IET}
 }
 
 @inproceedings{wang2015deep,
