### 2021-12-16 a9b2c66
Changed name of variational cca to probabilistic cca (solved by VI).
Changes to references in paper.

Change to title of paper
