+++ docs/joss/paper.bib
@@ -24,7 +24,7 @@
   title={A probabilistic interpretation of canonical correlation analysis},
   author={Bach, Francis R and Jordan, Michael I},
   year={2005},
-  url={https://statistics.berkeley.edu/sites/default/files/tech-reports/688.pdf}
+  url={https://statistics.berkeley.edu/sites/default/files/tech-reports/688.pdf},
   publisher={Technical Report 688, Department of Statistics, University of California Berkeley}
 }
 
