### 2021-12-17 #3823 @jameschapman19
Thanks @kyleniemeyer - I was able to track down a doi for Wenwen et al. too

@emdupre I made a related change to the references in the docstrings themselves so best to also change to 1.10.8 (but don't worry if it's too much trouble).
