### 2021-12-17 #3823 @emdupre
> @emdupre I made a related change to the references in the docstrings themselves so best to also change to 1.10.8 (but don't worry if it's too much trouble).

Thank you, @jameschapman19 ! To confirm, did you also create a corresponding archive for the new release ? Can you list the DOI here if so ?
