### 2021-12-17 #3823 @jameschapman19
New DOI: 10.5281/zenodo.5786616

Just realised there is also the DOI which resolves to the most recent: 10.5281/zenodo.4382739

I don't know what is more appropriate for JOSS but I'm happy with either.
