### 2021-12-17 #3823 @kyleniemeyer
We prefer the DOI that points to the specific version
