### 2021-12-17 #3823 @kyleniemeyer
@jameschapman19 sorry, the Wenwen et al. reference still needs one final change—the DOI field should not have the extra `doi=` in it, just do `doi={10.1049/cje.2017.08.004}`
