+++ docs/joss/paper.bib
@@ -116,7 +116,7 @@
   pages={459--466},
   year={2018},
   publisher={IET},
-  doi={doi={10.1049/cje.2017.08.004}}
+  doi={10.1049/cje.2017.08.004}
 }
 
 @inproceedings{andrew2013deep,
