### 2021-12-17 7d3d904
Fixed doc examples
