+++ cca_zoo/models/tcca.py
@@ -43,13 +43,13 @@ class TCCA(_CCA_Base):
     """
 
     def __init__(
-            self,
-            latent_dims: int = 1,
-            scale=True,
-            centre=True,
-            copy_data=True,
-            random_state=None,
-            c: Union[Iterable[float], float] = None,
+        self,
+        latent_dims: int = 1,
+        scale=True,
+        centre=True,
+        copy_data=True,
+        random_state=None,
+        c: Union[Iterable[float], float] = None,
     ):
         """
         Constructor for TCCA
@@ -127,7 +127,10 @@ class TCCA(_CCA_Base):
         ]
         multiplied_views = np.stack(transformed_views, axis=0).prod(axis=0).sum(axis=0)
         norms = np.stack(
-            [np.linalg.norm(transformed_view, axis=0) for transformed_view in transformed_views],
+            [
+                np.linalg.norm(transformed_view, axis=0)
+                for transformed_view in transformed_views
+            ],
             axis=0,
         ).prod(axis=0)
         corrs = multiplied_views / norms
@@ -190,19 +193,19 @@ class KTCCA(TCCA):
     """
 
     def __init__(
-            self,
-            latent_dims: int = 1,
-            scale: bool = True,
-            centre=True,
-            copy_data=True,
-            random_state=None,
-            eps=1e-3,
-            c: Union[Iterable[float], float] = None,
-            kernel: Iterable[Union[float, callable]] = None,
-            gamma: Iterable[float] = None,
-            degree: Iterable[float] = None,
-            coef0: Iterable[float] = None,
-            kernel_params: Iterable[dict] = None,
+        self,
+        latent_dims: int = 1,
+        scale: bool = True,
+        centre=True,
+        copy_data=True,
+        random_state=None,
+        eps=1e-3,
+        c: Union[Iterable[float], float] = None,
+        kernel: Iterable[Union[float, callable]] = None,
+        gamma: Iterable[float] = None,
+        degree: Iterable[float] = None,
+        coef0: Iterable[float] = None,
+        kernel_params: Iterable[dict] = None,
     ):
         """
         Constructor for TCCA
