+++ cca_zoo/data/utils.py
@@ -20,4 +20,4 @@ class CCA_Dataset(Dataset):
 
     def __getitem__(self, idx):
         views = [view[idx].astype(np.float32) for view in self.views]
-        return tuple(views)
+        return {"views": views}
