### 2021-12-18 3eb8df1
Fixed doc examples
