+++ docs/source/api/probabilisticmodels.rst
@@ -4,6 +4,6 @@ Probabilistic Models
 Variational CCA
 ----------------------------------------
 
-.. automodule:: cca_zoo.probabilisticmodels.vcca
+.. automodule:: cca_zoo.probabilisticmodels.probabilisticcca
     :inherited-members:
     :exclude-members: get_params, set_params
