### 2021-12-18 52cd4e9
Fixed doc examples
